The basic sircovid model The "basic" sircovid model. This is a dust model.

The basic sircovid model

The "basic" sircovid model. This is a dust model.

Examples

# Set up the basic model for England with default parameters and
# an initial seeding of early February:
p <- basic_parameters(sircovid_date("2020-02-07"), "england")
mod <- basic$new(p, 0, 10)

# Set the initial state and index as we would us for a run
# (without setting an initial state there is no seeding)
initial <- basic_initial(mod$info(), 10, p)
mod$update_state(state = initial$state, step = initial$step)
mod$set_index(basic_index(mod$info())$run)

# Run the model up to the end of march
step_end <- sircovid_date("2020-03-31") / p$dt

# The filtered state is returned at the end of the run
mod$run(step_end)
#>            [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> icu         208   86  260  118  241  192   66  144   96    92
#> deaths      678  241  760  327  620  573  212  404  279   267
#> deaths_inc  120   39  127   74  100  104   38   65   45    51

# More state can be retrieved using the "state" method
mod$state(1:6)
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,]       91.0       91.0       91.0       91.0       91.0       91.0
#> [2,] 56550138.0 56550138.0 56550138.0 56550138.0 56550138.0 56550138.0
#> [3,]      208.0       86.0      260.0      118.0      241.0      192.0
#> [4,]      678.0      241.0      760.0      327.0      620.0      573.0
#> [5,]      120.0       39.0      127.0       74.0      100.0      104.0
#> [6,]        0.1        0.1        0.1        0.1        0.1        0.1
#>            [,7]       [,8]       [,9]      [,10]
#> [1,]       91.0       91.0       91.0       91.0
#> [2,] 56550138.0 56550138.0 56550138.0 56550138.0
#> [3,]       66.0      144.0       96.0       92.0
#> [4,]      212.0      404.0      279.0      267.0
#> [5,]       38.0       65.0       45.0       51.0
#> [6,]        0.1        0.1        0.1        0.1