Parameters for the basic model

Parameters for the "basic" model.

basic_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  exp_noise = 1e+06
)

Arguments

start_date
Thee start date, as a sircovid_date() (i.e., the number of days into 2020)

region
The region to run the model for. This will bee used to get population data, which is currently fixed within the package and is limited to "uk", the four constituent nations ("england", "wales", "scotland", "northern_ireland") and the 7 NHS regions (e.g., "midlands"). These names are case insensitive.

beta_date
A vector of dates for changes in beta (the contact rate parameter), or NULL if a single value is used for all times (see sircovid_parameters_piecewise_linear(), where this is passed as date).

beta_value
A vector of values for beta (the contact rate parameter). If not given, and if beta_date is NULL then a value of 0.08 will be used through the whole simulation, otherwise if beta_date is NULL this must be a scalar. If beta_date is given then beta_date and beta_value must have the same length (see sircovid_parameters_piecewise_linear(), where this is passed as value).

severity
Severity data, via Bob Verity's markovid package. This needs to be NULL (use the default bundled data version in the package), a data.frame object (for raw severity data) or a list (for data that has already been processed by sircovid for use). New severity data comes from Bob Verity via the markovid package, and needs to be carefully calibrated with the progression parameters.

exp_noise
Rate of exponential noise used in the compare function - typically set to a large value so that noise is small but non-zero. If set to Inf then there is no noise in the observation process (not realistic but useful for testing).

Value

A list of inputs to the model, many of which are fixed and represent data. These correspond largely to user() calls within the odin code, though some are also used in processing just before the model is run.

Examples

basic_parameters(sircovid_date("2020-02-01"), "uk")

#> $hosp_transmission
#> [1] 0
#> 
#> $ICU_transmission
#> [1] 0
#> 
#> $G_D_transmission
#> [1] 0
#> 
#> $dt
#> [1] 0.25
#> 
#> $steps_per_day
#> [1] 4
#> 
#> $initial_step
#> [1] 128
#> 
#> $n_age_groups
#> [1] 17
#> 
#> $beta_step
#> [1] 0.08
#> 
#> $population
#>  [1] 3857263 4149852 3953866 3656968 4153080 4514249 4497132 4395667 4019539
#> [10] 4402122 4661015 4405908 3755185 3368199 3318867 2325296 3362599
#> 
#> $m
#>                 [0,4)        [5,9)      [10,14)      [15,19)      [20,24)
#> [0,4)    4.966707e-07 1.980202e-07 1.146078e-07 7.547886e-08 1.138137e-07
#> [5,9)    1.980202e-07 1.599396e-06 2.972491e-07 1.352721e-07 1.101069e-07
#> [10,14)  1.146078e-07 2.972491e-07 1.733225e-06 3.334453e-07 6.026086e-08
#> [15,19)  7.547886e-08 1.352721e-07 3.334453e-07 1.836025e-06 3.278825e-07
#> [20,24)  1.138137e-07 1.101069e-07 6.026086e-08 3.278825e-07 6.244090e-07
#> [25,29)  1.819153e-07 1.626085e-07 8.744544e-08 1.819264e-07 3.222430e-07
#> [30,34)  1.940880e-07 2.332740e-07 1.297928e-07 1.120304e-07 1.882853e-07
#> [35,39)  2.235471e-07 2.784775e-07 2.396054e-07 1.996738e-07 1.624326e-07
#> [40,44)  9.944949e-08 2.332186e-07 2.679243e-07 2.341623e-07 1.838660e-07
#> [45,49)  6.175863e-08 8.096090e-08 1.519569e-07 2.467186e-07 2.148012e-07
#> [50,54)  8.214138e-08 7.752704e-08 8.335625e-08 1.392241e-07 1.530723e-07
#> [55,59)  6.104769e-08 5.013907e-08 8.513107e-08 7.833919e-08 1.178128e-07
#> [60,64)  6.057159e-08 9.418311e-08 5.929099e-08 5.863294e-08 8.866898e-08
#> [65,69)  4.767006e-08 7.402091e-08 6.898452e-08 9.552015e-08 7.226646e-08
#> [70,74)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [75,79)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [80,100) 2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#>               [25,29)      [30,34)      [35,39)      [40,44)      [45,49)
#> [0,4)    1.819153e-07 1.940880e-07 2.235471e-07 9.944949e-08 6.175863e-08
#> [5,9)    1.626085e-07 2.332740e-07 2.784775e-07 2.332186e-07 8.096090e-08
#> [10,14)  8.744544e-08 1.297928e-07 2.396054e-07 2.679243e-07 1.519569e-07
#> [15,19)  1.819264e-07 1.120304e-07 1.996738e-07 2.341623e-07 2.467186e-07
#> [20,24)  3.222430e-07 1.882853e-07 1.624326e-07 1.838660e-07 2.148012e-07
#> [25,29)  4.054957e-07 2.389792e-07 1.870791e-07 1.918110e-07 2.005309e-07
#> [30,34)  2.389792e-07 3.706066e-07 2.728896e-07 2.145623e-07 1.654374e-07
#> [35,39)  1.870791e-07 2.728896e-07 3.412451e-07 3.032663e-07 2.149544e-07
#> [40,44)  1.918110e-07 2.145623e-07 3.032663e-07 3.370632e-07 3.085384e-07
#> [45,49)  2.005309e-07 1.654374e-07 2.149544e-07 3.085384e-07 4.254147e-07
#> [50,54)  1.882440e-07 1.671228e-07 1.606864e-07 1.873283e-07 1.702589e-07
#> [55,59)  1.560297e-07 1.454919e-07 1.379231e-07 1.451614e-07 1.482183e-07
#> [60,64)  1.189756e-07 1.374885e-07 1.668294e-07 1.791329e-07 1.414642e-07
#> [65,69)  1.147535e-07 8.979643e-08 1.540697e-07 1.348771e-07 8.620090e-08
#> [70,74)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [75,79)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [80,100) 4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#>               [50,54)      [55,59)      [60,64)      [65,69)      [70,74)
#> [0,4)    8.214138e-08 6.104769e-08 6.057159e-08 4.767006e-08 2.100618e-08
#> [5,9)    7.752704e-08 5.013907e-08 9.418311e-08 7.402091e-08 2.893729e-08
#> [10,14)  8.335625e-08 8.513107e-08 5.929099e-08 6.898452e-08 5.503570e-08
#> [15,19)  1.392241e-07 7.833919e-08 5.863294e-08 9.552015e-08 9.539437e-08
#> [20,24)  1.530723e-07 1.178128e-07 8.866898e-08 7.226646e-08 5.468746e-08
#> [25,29)  1.882440e-07 1.560297e-07 1.189756e-07 1.147535e-07 4.760237e-08
#> [30,34)  1.671228e-07 1.454919e-07 1.374885e-07 8.979643e-08 3.516410e-08
#> [35,39)  1.606864e-07 1.379231e-07 1.668294e-07 1.540697e-07 4.653274e-08
#> [40,44)  1.873283e-07 1.451614e-07 1.791329e-07 1.348771e-07 1.307023e-07
#> [45,49)  1.702589e-07 1.482183e-07 1.414642e-07 8.620090e-08 1.059467e-07
#> [50,54)  1.592838e-07 2.060829e-07 1.178586e-07 9.108480e-08 9.232320e-08
#> [55,59)  2.060829e-07 2.647960e-07 1.998500e-07 1.379069e-07 8.128181e-08
#> [60,64)  1.178586e-07 1.998500e-07 1.734975e-07 1.674043e-07 1.041277e-07
#> [65,69)  9.108480e-08 1.379069e-07 1.674043e-07 2.089258e-07 1.332611e-07
#> [70,74)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [75,79)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [80,100) 9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#>               [75,79)     [80,100)
#> [0,4)    2.100618e-08 2.100618e-08
#> [5,9)    2.893729e-08 2.893729e-08
#> [10,14)  5.503570e-08 5.503570e-08
#> [15,19)  9.539437e-08 9.539437e-08
#> [20,24)  5.468746e-08 5.468746e-08
#> [25,29)  4.760237e-08 4.760237e-08
#> [30,34)  3.516410e-08 3.516410e-08
#> [35,39)  4.653274e-08 4.653274e-08
#> [40,44)  1.307023e-07 1.307023e-07
#> [45,49)  1.059467e-07 1.059467e-07
#> [50,54)  9.232320e-08 9.232320e-08
#> [55,59)  8.128181e-08 8.128181e-08
#> [60,64)  1.041277e-07 1.041277e-07
#> [65,69)  1.332611e-07 1.332611e-07
#> [70,74)  1.493131e-07 1.493131e-07
#> [75,79)  1.493131e-07 1.493131e-07
#> [80,100) 1.493131e-07 1.493131e-07
#> 
#> $p_star
#>  [1] 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2
#> 
#> $p_C
#>  [1] 0.1809 0.1809 0.1809 0.1809 0.2241 0.2241 0.2241 0.2241 0.3054 0.3054
#> [11] 0.3054 0.3054 0.3546 0.3546 0.3546 0.3546 0.6456
#> 
#> $p_G_D
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05
#> [16] 0.05 0.05
#> 
#> $p_H_D
#>  [1] 0.1258933 0.1226134 0.1356729 0.1526679 0.1743031 0.1941879 0.2093617
#>  [8] 0.2244326 0.2370135 0.2578281 0.2908746 0.3207640 0.3625638 0.3909655
#> [15] 0.4211515 0.4475459 0.4820000
#> 
#> $p_ICU_D
#>  [1] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#>  [8] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> [15] 0.5234896 0.5234896 0.5234896
#> 
#> $p_W_D
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05
#> [16] 0.05 0.05
#> 
#> $p_ICU
#>  [1] 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387
#>  [7] 0.11845387 0.12551064 0.14091042 0.17595537 0.24437611 0.35243489
#> [13] 0.52779221 0.72404472 0.18197685 0.05439050 0.01673267
#> 
#> $p_sero_pos_1
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394
#> 
#> $p_sero_pos_2
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394
#> 
#> $p_H
#>  [1] 0.000632689 0.000539148 0.000995640 0.002035807 0.004545371 0.008748134
#>  [7] 0.013802147 0.019850725 0.024607803 0.032822485 0.049084367 0.061571007
#> [13] 0.086379194 0.099452536 0.124208801 0.150169192 0.153030303
#> 
#> $p_recov_ICU
#>  [1] 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104
#>  [8] 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104
#> [15] 0.4765104 0.4765104 0.4765104
#> 
#> $p_recov_hosp
#>  [1] 0.7705654 0.7734568 0.7619442 0.7469624 0.7278899 0.7103605 0.6969841
#>  [8] 0.6782255 0.6554737 0.6115828 0.5358321 0.4398496 0.3010024 0.1680663
#> [15] 0.4735115 0.5224059 0.5093325
#> 
#> $p_death_hosp
#>  [1] 0.1109807 0.1080894 0.1196019 0.1345838 0.1536562 0.1711856 0.1845620
#>  [8] 0.1962639 0.2036158 0.2124618 0.2197918 0.2077156 0.1712054 0.1078890
#> [15] 0.3445117 0.4232036 0.4739349
#> 
#> $p_recov_sympt
#>  [1] 0.9993673 0.9994609 0.9990044 0.9979642 0.9954546 0.9912519 0.9861979
#>  [8] 0.9801493 0.9753922 0.9671775 0.9509156 0.9384290 0.9136208 0.9005475
#> [15] 0.8757912 0.8498308 0.8469697
#> 
#> $k_E
#> [1] 2
#> 
#> $k_A
#> [1] 1
#> 
#> $k_C
#> [1] 1
#> 
#> $k_hosp
#> [1] 2
#> 
#> $k_ICU
#> [1] 2
#> 
#> $k_rec
#> [1] 2
#> 
#> $gamma_E
#> [1] 0.4357298
#> 
#> $gamma_A
#> [1] 0.4784689
#> 
#> $gamma_C
#> [1] 0.25
#> 
#> $gamma_hosp
#> [1] 2
#> 
#> $gamma_ICU
#> [1] 0.4
#> 
#> $gamma_rec
#> [1] 0.4
#> 
#> $phi_ICU
#> [1] 0.95
#> 
#> $kappa_ICU
#> [1] 2
#> 
#> $phi_death
#> [1] 0.9087341
#> 
#> $kappa_death
#> [1] 2
#> 
#> $exp_noise
#> [1] 1e+06
#>