Forecast the carehomes model

Forecast from the carehomes model; this provides a wrapper around mcstate::pmcmc_sample and mcstate::pmcmc_predict that samples the trajectories then creates samples, setting the sircovid dates and adding trajectories of incidence.

carehomes_forecast(
  samples,
  n_sample,
  burnin,
  forecast_days,
  incidence_states,
  prepend_trajectories = TRUE
)

Arguments

samples
Results of running mcstate::pmcmc()

n_sample
Number of samples to take. If you provide a value of 0 then sampling is skipped (i.e., the entire set of parameters in samples will be forecasted). Otherwise this is passed to mcstate::pmcmc_sample

burnin
Optional integer number of iterations to discard as "burn-in". If given then samples 1:burnin will be excluded from your results. Remember that the first sample represents the starting point of the chain. It is an error if this is not a positive integer or is greater than or equal to the number of samples (i.e., there must be at least one sample remaining after discarding burnin).

forecast_days
The number of days to create a forecast for

incidence_states
A character vector of states for which incidnce should be computed (from cumulative compartments, such as deaths). These will end up in the final trajectories object with the sufix _inc (e.g., deaths becomes deaths_inc).

prepend_trajectories
Prepend trajectories from the particle filter to the predictions created here.