Initial conditions for the carehomes model

Create initial conditions for the carehomes model. This matches the interface required for mcstate

carehomes_initial(info, n_particles, pars)

Arguments

info
The result of running the $info() method on an initialised basic model

n_particles
The number of particles required. Currently only uniform initial seeding is implemented so this has no effect

pars
A parameter list created by basic_parameters(); from this list we will use the population and initial_step elements.

Value

A numeric vector of initial conditions

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_initial(mod$info(), 10, p)

#> $state
#>    [1]        0        0        0        0        0 56286962 56286962        0
#>    [9]        0        0        0        0        0        0        0        0
#>   [17]       10        0        0        0        0        0        0        0
#>   [25]        0        0        0        0        0        0        0        0
#>   [33]        0        0        0        0        0        0        0        0
#>   [41]        0        0        0        0        0        0        0        0
#>   [49]        0        0        0        0        0        0        0        0
#>   [57]        0        0        0        0        0        0        0        0
#>   [65]        0        0        0        0        0        0        0        0
#>   [73]        0        0        0  3299637  3538206  3354246  3090232  3487863
#>   [81]  3758950  3765495  3691183  3371838  3673353  3865002  3628192  3069376
#>   [89]  2779756  2762342  1889735  2582208   339674   339674        0        0
#>   [97]        0        0        0        0        0        0        0        0
#>  [105]        0        0        0        0        0        0        0        0
#>  [113]        0        0        0        0        0        0        0        0
#>  [121]        0        0        0        0        0        0        0        0
#>  [129]        0        0        0        0        0        0        0        0
#>  [137]        0        0        0        0        0        0        0        0
#>  [145]        0        0        0        0        0        0        0        0
#>  [153]        0        0        0        0        0        0        0        0
#>  [161]        0        0        0        0        0        0        0        0
#>  [169]        0        0        0        0        0        0        0        0
#>  [177]        0        0        0        0        0        0        0        0
#>  [185]        0        0        0        0        0  3299637  3538206  3354246
#>  [193]  3090222  3487863  3758950  3765495  3691183  3371838  3673353  3865002
#>  [201]  3628192  3069376  2779756  2762342  1889735  2582208   339674   339674
#>  [209]        0        0        0        0        0        0        0        0
#>  [217]        0        0        0        0        0        0        0        0
#>  [225]        0        0        0        0        0        0        0        0
#>  [233]        0        0        0        0        0        0        0        0
#>  [241]        0        0        0        0        0        0        0        0
#>  [249]        0        0        0        0        0        0        0        0
#>  [257]        0        0        0        0        0        0        0        0
#>  [265]        0        0        0        0        0        0        0        0
#>  [273]        0        0        0        0        0        0        0        0
#>  [281]        0        0        0        0        0        0        0        0
#>  [289]        0        0        0        0        0        0        0        0
#>  [297]        0        0        0        0        0        0        0        0
#>  [305]        0        0       10        0        0        0        0        0
#>  [313]        0        0        0        0        0        0        0        0
#>  [321]        0        0        0        0        0        0        0        0
#>  [329]        0        0        0        0        0        0        0        0
#>  [337]        0        0        0        0        0        0        0        0
#>  [345]        0        0        0        0        0        0        0        0
#>  [353]        0        0        0        0        0        0        0        0
#>  [361]        0        0        0        0        0        0        0        0
#>  [369]        0        0        0        0        0        0        0        0
#>  [377]        0        0        0        0        0        0        0        0
#>  [385]        0        0        0        0        0        0        0        0
#>  [393]        0        0        0        0        0        0        0        0
#>  [401]        0        0        0        0        0        0        0        0
#>  [409]        0        0        0        0        0        0        0        0
#>  [417]        0        0        0        0        0        0        0        0
#>  [425]        0        0        0        0        0        0        0        0
#>  [433]        0        0        0        0        0        0        0        0
#>  [441]        0        0        0        0        0        0        0        0
#>  [449]        0        0        0        0        0        0        0        0
#>  [457]        0        0        0        0        0        0        0        0
#>  [465]        0        0        0        0        0        0        0        0
#>  [473]        0        0        0        0        0        0        0        0
#>  [481]        0        0        0        0        0        0        0        0
#>  [489]        0        0        0        0        0        0        0        0
#>  [497]        0        0        0        0        0        0        0        0
#>  [505]        0        0        0        0        0        0        0        0
#>  [513]        0        0        0        0        0        0        0        0
#>  [521]        0        0        0        0        0        0        0        0
#>  [529]        0        0        0        0        0        0        0        0
#>  [537]        0        0        0        0        0        0        0        0
#>  [545]        0        0        0        0        0        0        0        0
#>  [553]        0        0        0        0        0        0        0        0
#>  [561]        0        0        0        0        0        0        0        0
#>  [569]        0        0        0        0        0        0        0        0
#>  [577]        0        0        0        0        0        0        0        0
#>  [585]        0        0        0        0        0        0        0        0
#>  [593]        0        0        0        0        0        0        0        0
#>  [601]        0        0        0        0        0        0        0        0
#>  [609]        0        0        0        0        0        0        0        0
#>  [617]        0        0        0        0        0        0        0        0
#>  [625]        0        0        0        0        0        0        0        0
#>  [633]        0        0        0        0        0        0        0        0
#>  [641]        0        0        0        0        0        0        0        0
#>  [649]        0        0        0        0        0        0        0        0
#>  [657]        0        0        0        0        0        0        0        0
#>  [665]        0        0        0        0        0        0        0        0
#>  [673]        0        0        0        0        0        0        0        0
#>  [681]        0        0        0        0        0        0        0        0
#>  [689]        0        0        0        0        0        0        0        0
#>  [697]        0        0        0        0        0        0        0        0
#>  [705]        0        0        0        0        0        0        0        0
#>  [713]        0        0        0        0        0        0        0        0
#>  [721]        0        0        0        0        0        0        0        0
#>  [729]        0        0        0        0        0        0        0        0
#>  [737]        0        0        0        0        0        0        0        0
#>  [745]        0        0        0        0        0        0        0        0
#>  [753]        0        0        0        0        0        0        0        0
#>  [761]        0        0        0        0        0        0        0        0
#>  [769]        0        0        0        0        0        0        0        0
#>  [777]        0        0        0        0        0        0        0        0
#>  [785]        0        0        0        0        0        0        0        0
#>  [793]        0        0        0        0        0        0        0        0
#>  [801]        0        0        0        0        0        0        0        0
#>  [809]        0        0        0        0        0        0        0        0
#>  [817]        0        0        0        0        0        0        0        0
#>  [825]        0        0        0        0        0        0        0        0
#>  [833]        0        0        0        0        0        0        0        0
#>  [841]        0        0        0        0        0        0        0        0
#>  [849]        0        0        0        0        0        0        0        0
#>  [857]        0        0        0        0        0        0        0        0
#>  [865]        0        0        0        0        0        0        0        0
#>  [873]        0        0        0        0        0        0        0        0
#>  [881]        0        0        0        0        0        0        0        0
#>  [889]        0        0        0        0        0        0        0        0
#>  [897]        0        0        0        0        0        0        0        0
#>  [905]        0        0        0        0        0        0        0        0
#>  [913]        0        0        0        0        0        0        0        0
#>  [921]        0        0        0        0        0        0        0        0
#>  [929]        0        0        0        0        0        0        0        0
#>  [937]        0        0        0        0        0        0        0        0
#>  [945]        0        0        0        0        0        0        0        0
#>  [953]        0        0        0        0        0        0        0        0
#>  [961]        0        0        0        0        0        0        0        0
#>  [969]        0        0        0        0        0        0        0        0
#>  [977]        0        0        0        0        0        0        0        0
#>  [985]        0        0        0        0        0        0       10        0
#>  [993]        0        0        0        0        0        0        0        0
#> [1001]        0        0        0        0        0        0        0        0
#> [1009]        0        0        0        0        0        0        0        0
#> [1017]        0        0        0        0        0        0        0        0
#> [1025]        0        0        0        0        0        0        0        0
#> [1033]        0        0        0        0        0        0        0        0
#> [1041]        0        0        0        0        0        0        0        0
#> [1049]        0        0        0        0        0        0        0        0
#> [1057]        0        0        0        0        0        0        0        0
#> [1065]        0        0        0        0        0        0        0        0
#> [1073]        0        0        0        0        0        0        0        0
#> [1081]        0        0        0        0        0        0        0        0
#> [1089]        0        0        0        0        0        0        0        0
#> [1097]        0        0        0        0        0        0        0        0
#> [1105]       10        0        0        0        0        0        0        0
#> [1113]        0        0        0        0        0        0        0        0
#> [1121]        0        0        0        0        0        0        0        0
#> [1129]        0        0        0        0        0        0        0        0
#> [1137]        0        0        0
#> 
#> $step
#> [1] 152
#>