Carehomes severity parameters

Carehomes severity parameters

carehomes_parameters_severity(
  dt,
  severity = NULL,
  p_C = NULL,
  p_H = NULL,
  p_H_CHR = NULL,
  p_ICU = NULL,
  p_H_D = NULL,
  p_ICU_D = NULL,
  p_W_D = NULL,
  p_G_D = NULL,
  p_G_D_CHR = NULL,
  p_star = NULL
)

Arguments

dt
The step size

severity
Severity data, used to determine default severity parameter age-scalings and to provide default severity parameter values. Can be NULL (use the default bundled data version in the package), or a data.frame object (for raw severity data).

p_C
Time-varying parameters for p_C (the probability of an infected individual becoming symptomatic). See Details.

p_H
Time-varying parameters for p_H (the probability of a symptomatic individual requiring hospitalisation). See Details.

p_H_CHR
Time-varying parameters for p_H (the probability of a symptomatic individual requiring hospitalisation) for care home residents. If NULL then the value for the oldest age group is used. See Details.

p_ICU
Time-varying parameters for p_ICU (the probability of a hospitalised individual going to ICU). See Details.

p_H_D
Time-varying parameters for p_H_D (the probability of death in general beds).See Details.

p_ICU_D
Time-varying parameters for p_ICU_D (the probability of death in ICU). See Details.

p_W_D
Time-varying parameters for p_W_D (the probability of death in stepdown). See Details.

p_G_D
Time-varying parameters for p_G_D (the probability of individuals requiring hospitalisation dying in the community or a care home). See Details.

p_G_D_CHR
Time-varying parameters for p_G_D (the probability of individuals requiring hospitalisation dying in a care home) for care home residents. If NULL then the value for the oldest age group is used. See Details.

p_star
Time-varying parameters for p_star (the probability of patients being confirmed as covid on admission to hospital). See Details.

Value

A list of severity parameters

Time-varying parameters

Every time varying parameter has the same format, which can be NULL (in which case the value from severity is used) or a list with date and value for changes in the parameter. If value is scalar then date can be NULL or missing. If value is a vector then date must be a vector of sircovid dates of the same length as value.