Combine Rt estimates from EpiEstim

Combine Rt estimates from EpiEstim across multiple runs.

Usage,

combine_rt_epiestim(rt, samples, q = NULL, rank = TRUE)

Arguments

rt

A list of Rt calculations from lancelot_rt_trajectories_epiestim() (though any Rt calculation that confirms to this will work)

samples

A list of samples from lancelot_forecast()

q

A vector of quantiles to return values for

rank

A boolean deciding whether to rank trajectories by increasing incidence or not before combining Rt estimates

Value

A list of Rt output in the same structure as the first element of rt. Rt estimates will be aggregated across regions (or whatever else you are aggregating on) based on the parameters in samples.