Create parameters

Create parameters for use with the model

generate_parameters(
  sircovid_model = basic_model(),
  transmission_model = "POLYMOD",
  country = "United Kingdom",
  severity_data_file = NULL,
  infection_seeding = list(values = c(10), bins = c("15 to 19")),
  beta = 0.1,
  beta_times = "2020-02-02",
  trans_profile = c(1),
  trans_increase = c(1),
  hosp_transmission = 0.1,
  ICU_transmission = 0.05,
  dt = 0.25,
  use_polymod_pop = FALSE
)

Arguments

sircovid_model
Model to use, basic_model() or hospital_model()

transmission_model
Model type. Only 'POLYMOD' currently supported

country
Country name

severity_data_file
Location of file with severity data

infection_seeding
List of vector values of how many cases to seed in a correspond vector of age bins in bins.

beta
Beta, for each time step in beta_times

beta_times
Dates beta changes, in format yyyy-mm-dd

trans_profile
Proportion in each infectivity group

trans_increase
Relative infectivity of each group

hosp_transmission
Transmission rate of hospital cases

ICU_transmission
Tranmissions rate of ICU cases

dt
Time-step to run the model in days

use_polymod_pop
Set to ignore survey_pop_in and use the population from polymod when estimating the transmission matrix

Value

List of parameters for use with sircovid