Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(
  info,
  rt = TRUE,
  cum_admit = TRUE,
  diagnoses_admitted = TRUE,
  cum_infections_disag = TRUE,
  cum_n_vaccinated = TRUE
)

Arguments

info

The result of running the $info() method on an initialised basic model

rt

Logical, whether to output trajectories required for calculating Rt (default = TRUE)

cum_admit

Logical, whether to output cumulative admissions by age (default = TRUE)

diagnoses_admitted

Logical, whether to output cumulative combined confirmed admissions and inpatient diagnoses by age and vaccine class (default = TRUE)

cum_infections_disag

Logical, whether to output cumulative infections by age and vaccine class (default = TRUE)

cum_n_vaccinated

Logical, whether to output cumulative number vaccinated by age and vaccine class (default = TRUE)

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
#> $run
#>                               time                                icu 
#>                                  1                                 12 
#>                            general               deaths_carehomes_inc 
#>                                 13                                 35 
#>                    deaths_comm_inc               deaths_comm_0_49_inc 
#>                                 25                                 26 
#>              deaths_comm_50_54_inc              deaths_comm_55_59_inc 
#>                                 27                                 28 
#>              deaths_comm_60_64_inc              deaths_comm_65_69_inc 
#>                                 29                                 30 
#>              deaths_comm_70_74_inc              deaths_comm_75_79_inc 
#>                                 31                                 32 
#>            deaths_comm_80_plus_inc                    deaths_hosp_inc 
#>                                 33                                 38 
#>               deaths_hosp_0_49_inc              deaths_hosp_50_54_inc 
#>                                 39                                 40 
#>              deaths_hosp_55_59_inc              deaths_hosp_60_64_inc 
#>                                 41                                 42 
#>              deaths_hosp_65_69_inc              deaths_hosp_70_74_inc 
#>                                 43                                 44 
#>              deaths_hosp_75_79_inc            deaths_hosp_80_plus_inc 
#>                                 45                                 46 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 47                                 48 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 59                                 60 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 61                                 63 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 64                                 65 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 66                                 67 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 68                                 62 
#>                          react_pos 
#>                                 69 
#> 
#> $state
#>                                icu                            general 
#>                                 12                                 13 
#>               deaths_carehomes_inc                    deaths_comm_inc 
#>                                 35                                 25 
#>               deaths_comm_0_49_inc              deaths_comm_50_54_inc 
#>                                 26                                 27 
#>              deaths_comm_55_59_inc              deaths_comm_60_64_inc 
#>                                 28                                 29 
#>              deaths_comm_65_69_inc              deaths_comm_70_74_inc 
#>                                 30                                 31 
#>              deaths_comm_75_79_inc            deaths_comm_80_plus_inc 
#>                                 32                                 33 
#>                    deaths_hosp_inc               deaths_hosp_0_49_inc 
#>                                 38                                 39 
#>              deaths_hosp_50_54_inc              deaths_hosp_55_59_inc 
#>                                 40                                 41 
#>              deaths_hosp_60_64_inc              deaths_hosp_65_69_inc 
#>                                 42                                 43 
#>              deaths_hosp_70_74_inc              deaths_hosp_75_79_inc 
#>                                 44                                 45 
#>            deaths_hosp_80_plus_inc                       admitted_inc 
#>                                 46                                  2 
#>                      diagnoses_inc                         sero_pos_1 
#>                                  3                                 47 
#>                         sero_pos_2                    sympt_cases_inc 
#>                                 48                                 59 
#>        sympt_cases_non_variant_inc             sympt_cases_over25_inc 
#>                                 60                                 61 
#>            sympt_cases_under15_inc              sympt_cases_15_24_inc 
#>                                 63                                 64 
#>              sympt_cases_25_49_inc              sympt_cases_50_64_inc 
#>                                 65                                 66 
#>              sympt_cases_65_79_inc            sympt_cases_80_plus_inc 
#>                                 67                                 68 
#> sympt_cases_non_variant_over25_inc                          react_pos 
#>                                 62                                 69 
#>                               hosp                           admitted 
#>                                 14                                  6 
#>                          diagnoses                             deaths 
#>                                  7                                 36 
#>                         deaths_inc                         infections 
#>                                 37                                  4 
#>                     infections_inc                                S_0 
#>                                  5                                433 
#>                                S_5                               S_10 
#>                                434                                435 
#>                               S_15                               S_20 
#>                                436                                437 
#>                               S_25                               S_30 
#>                                438                                439 
#>                               S_35                               S_40 
#>                                440                                441 
#>                               S_45                               S_50 
#>                                442                                443 
#>                               S_55                               S_60 
#>                                444                                445 
#>                               S_65                               S_70 
#>                                446                                447 
#>                               S_75                               S_80 
#>                                448                                449 
#>                              S_CHW                              S_CHR 
#>                                450                                451 
#>                                R_0                                R_5 
#>                                585                                586 
#>                               R_10                               R_15 
#>                                587                                588 
#>                               R_20                               R_25 
#>                                589                                590 
#>                               R_30                               R_35 
#>                                591                                592 
#>                               R_40                               R_45 
#>                                593                                594 
#>                               R_50                               R_55 
#>                                595                                596 
#>                               R_60                               R_65 
#>                                597                                598 
#>                               R_70                               R_75 
#>                                599                                600 
#>                               R_80                              R_CHW 
#>                                601                                602 
#>                              R_CHR                        prob_strain 
#>                                603                                261 
#>                        cum_admit_0                        cum_admit_5 
#>                                223                                224 
#>                       cum_admit_10                       cum_admit_15 
#>                                225                                226 
#>                       cum_admit_20                       cum_admit_25 
#>                                227                                228 
#>                       cum_admit_30                       cum_admit_35 
#>                                229                                230 
#>                       cum_admit_40                       cum_admit_45 
#>                                231                                232 
#>                       cum_admit_50                       cum_admit_55 
#>                                233                                234 
#>                       cum_admit_60                       cum_admit_65 
#>                                235                                236 
#>                       cum_admit_70                       cum_admit_75 
#>                                237                                238 
#>                       cum_admit_80                      cum_admit_CHW 
#>                                239                                240 
#>                      cum_admit_CHR               diagnoses_admitted_0 
#>                                241                                376 
#>               diagnoses_admitted_5              diagnoses_admitted_10 
#>                                377                                378 
#>              diagnoses_admitted_15              diagnoses_admitted_20 
#>                                379                                380 
#>              diagnoses_admitted_25              diagnoses_admitted_30 
#>                                381                                382 
#>              diagnoses_admitted_35              diagnoses_admitted_40 
#>                                383                                384 
#>              diagnoses_admitted_45              diagnoses_admitted_50 
#>                                385                                386 
#>              diagnoses_admitted_55              diagnoses_admitted_60 
#>                                387                                388 
#>              diagnoses_admitted_65              diagnoses_admitted_70 
#>                                389                                390 
#>              diagnoses_admitted_75              diagnoses_admitted_80 
#>                                391                                392 
#>             diagnoses_admitted_CHW             diagnoses_admitted_CHR 
#>                                393                                394 
#>             cum_infections_disag_0             cum_infections_disag_5 
#>                                395                                396 
#>            cum_infections_disag_10            cum_infections_disag_15 
#>                                397                                398 
#>            cum_infections_disag_20            cum_infections_disag_25 
#>                                399                                400 
#>            cum_infections_disag_30            cum_infections_disag_35 
#>                                401                                402 
#>            cum_infections_disag_40            cum_infections_disag_45 
#>                                403                                404 
#>            cum_infections_disag_50            cum_infections_disag_55 
#>                                405                                406 
#>            cum_infections_disag_60            cum_infections_disag_65 
#>                                407                                408 
#>            cum_infections_disag_70            cum_infections_disag_75 
#>                                409                                410 
#>            cum_infections_disag_80           cum_infections_disag_CHW 
#>                                411                                412 
#>           cum_infections_disag_CHR                 cum_n_vaccinated_0 
#>                                413                                357 
#>                 cum_n_vaccinated_5                cum_n_vaccinated_10 
#>                                358                                359 
#>                cum_n_vaccinated_15                cum_n_vaccinated_20 
#>                                360                                361 
#>                cum_n_vaccinated_25                cum_n_vaccinated_30 
#>                                362                                363 
#>                cum_n_vaccinated_35                cum_n_vaccinated_40 
#>                                364                                365 
#>                cum_n_vaccinated_45                cum_n_vaccinated_50 
#>                                366                                367 
#>                cum_n_vaccinated_55                cum_n_vaccinated_60 
#>                                368                                369 
#>                cum_n_vaccinated_65                cum_n_vaccinated_70 
#>                                370                                371 
#>                cum_n_vaccinated_75                cum_n_vaccinated_80 
#>                                372                                373 
#>               cum_n_vaccinated_CHW               cum_n_vaccinated_CHR 
#>                                374                                375 
#>