Grid search of beta and start date

Run a grid search of the particle filter over beta and start date. This is parallelised, first run plan(multiprocess) to set this up.

scan_beta_date(
  min_beta,
  max_beta,
  beta_step,
  first_start_date,
  last_start_date,
  day_step,
  data,
  sircovid_model = basic_model(),
  model_params,
  pars_obs,
  n_particles = 100,
  scale_prior = NULL,
  shape_prior = NULL,
  tolerance = 1e-10
)

Arguments

min_beta
Minimum value of beta in the search

max_beta
Maximum value of beta in the search

beta_step
Step to increment beta between min and max

first_start_date
Earliest start date as 'yyyy-mm-dd'

last_start_date
Latest start date as 'yyyy-mm-dd'

day_step
Step to increment date in days

data
Hosptial data to fit to. See example.csv and particle_filter_data()

sircovid_model
An odin model generator and comparison function.

model_params
Model parameters, from a call to generate_parameters()

pars_obs
list of parameters to use for the comparison function.

n_particles
Number of particles. Positive Integer. Default = 100

scale_prior
Set to use a gamma prior on beta, and report the posterior rather than the likelihood. Sets scale of gamma prior.

shape_prior
Set to use a gamma prior on beta, and report the posterior rather than the likelihood. Sets shape of gamma prior.

tolerance
The smallest difference from 0 that is acceptable in the probability matrix.

Value

List of beta and start date grid values, and normalised probabilities at each point