Prepare data for mcstate

Convert data into mcstate format; this is a thin wrapper around mcstate::particle_filter_data() which adds a dummy step in front of the first data point so that we can use the previous state and the current states to convert cumulative measures into net daily changes.

sircovid_data(data, start_date, dt, expected = NULL)

Arguments

data

A data.frame object suitable for mcstate::particle_filter_data()

start_date

The start date, as a sircovid_date(), R "Date" object or a string in ISO 8601 format (YYYY-MM-DD)

dt

The time step (fraction of a day that each step represents) as used to create the model object

expected

Optional list of expected column names and types that are expected (use NA_real_). These will be filled in with the value where missing.

Value

A data.frame suitable for use with mcstate functions such as mcstate::particle_filter() and mcstate::pmcmc()

Examples

# A data sert that has data from the first of February to the first of
# March (one column of data called 'x')
from <- as.Date("2020-02-01")
to <- as.Date("2020-03-01")
d <- data.frame(date = seq(from, to, by = 1),
                x = runif(to - from + 1),
                stringsAsFactors = FALSE)

# Get this ready for sircovid/mcstate assuming the seeding starts on
# the 15th of January and we take 4 steps per day.
sircovid_data(d, start_date = "2020-01-15", 1 / 4)
#>    date_start date_end step_start step_end         x
#> 1          15       32         60      128 0.0108929
#> 2          32       33        128      132 0.7432243
#> 3          33       34        132      136 0.1132905
#> 4          34       35        136      140 0.1775330
#> 5          35       36        140      144 0.2203747
#> 6          36       37        144      148 0.7298695
#> 7          37       38        148      152 0.7547447
#> 8          38       39        152      156 0.7475137
#> 9          39       40        156      160 0.2732705
#> 10         40       41        160      164 0.6947687
#> 11         41       42        164      168 0.2798471
#> 12         42       43        168      172 0.1439156
#> 13         43       44        172      176 0.9334528
#> 14         44       45        176      180 0.4417651
#> 15         45       46        180      184 0.5402842
#> 16         46       47        184      188 0.7761450
#> 17         47       48        188      192 0.8777766
#> 18         48       49        192      196 0.3241285
#> 19         49       50        196      200 0.7426555
#> 20         50       51        200      204 0.1638625
#> 21         51       52        204      208 0.8354857
#> 22         52       53        208      212 0.7996357
#> 23         53       54        212      216 0.1850547
#> 24         54       55        216      220 0.9908292
#> 25         55       56        220      224 0.3936286
#> 26         56       57        224      228 0.3923258
#> 27         57       58        228      232 0.8904658
#> 28         58       59        232      236 0.3644824
#> 29         59       60        236      240 0.1750675
#> 30         60       61        240      244 0.7895891