Remap state to include vaccination

Remap state that was run with one stratum to support vaccination. This is useful where fitting was done pre-vaccination and the model state needs to be expanded to support vaccination for simulations.

vaccination_remap_state(state_orig, info_orig, info_vacc)

Arguments

state_orig
An original state matrix (rows representing state, columns representing particles or samples). This must be the complete model state.

info_orig
The results of info() from the model object without vaccination

info_vacc
The results of info() from the model object with vaccination

Value

A 3d array with more rows than state.