Compute vaccination order

Compute vaccination priority following JCVI ordering.

vaccine_priority_proportion(
  uptake,
  prop_hcw = NULL,
  prop_very_vulnerable = NULL,
  prop_underlying_condition = NULL
)

vaccine_priority_population(
  region,
  uptake,
  prop_hcw = NULL,
  prop_very_vulnerable = NULL,
  prop_underlying_condition = NULL
)

Arguments

uptake
A vector of length 19 with fractional uptake per group. If a single number is given it is shared across all groups (note that this includes under-18s)

prop_hcw
Assumed fraction of healthcare workers in each group (length 19) - if NULL we use a default that is a guess with hopefully the right general shape.

prop_very_vulnerable
Assumed fraction "very vulnerable" in each group (length 19) - if NULL we use a default that is a guess with hopefully the right general shape.

prop_underlying_condition
Assumed fraction "underlying condition" in each group (length 19) - if NULL we use a default that is a guess with hopefully the right general shape.

region
Region to use to get total population numbers

Value

A matrix with n_groups rows (19) and columns representing priority groups, and element (i, j) is the proportion in group i who should be vaccinated as part of priority group j, accounting for uptake so that the sum over the rows corresponds to the total fractional uptake in that group. For vaccine_priority_population, the total number of individuals replaces the proportion (based on the demography used by sircovid).

Details

https://tinyurl.com/8uwtatvm

Assmuing independance between job (e.g. HCW) and clinical condition

But assuming one is either counted as "clinically extremely vulnerable" or "with underlying health conditions" but not both

The JCVI priority groups, in decending order are:

residents in a care home for older adults and their carers

all those 80 years of age and over and frontline health and social care workers

all those 75 years of age and over

all those 70 years of age and over and clinically extremely vulnerable individuals

all those 65 years of age and over

all individuals aged 16 years to 64 years with underlying health conditions which put them at higher risk of serious disease and mortality

all those 60 years of age and over

all those 55 years of age and over

all those 50 years of age and over

all those 40-49 years of age and over

all those 30-39 years of age and over

all those 18-29 years of age and over