Vaccination schedule using data and future

Helper function to create a vaccination schedule that covers data from the past and projects doses into the future based on the last week of vaccination (based on JCVI order using vaccine_schedule_future). This function is subject to change.

vaccine_schedule_data_future(
  data,
  region,
  uptake,
  end_date,
  mean_days_between_doses
)

Arguments

data
A data.frame with columns date, age_band_min, dose1 and dose2.

region
Region to use to get total population numbers

uptake
A vector of length 19 with fractional uptake per group. If a single number is given it is shared across all groups (note that this includes under-18s)

end_date
The final day in the future to create a schedule for. After this date the model will assume 0 vaccine doses given so an overestimate is probably better than an underestimate.

mean_days_between_doses
Assumed mean days between doses one and two

Value

A vaccine_schedule object