High-level vaccine scenario creation

Create a vaccination scenario

vaccine_schedule_scenario(
  schedule_past,
  doses_future,
  end_date,
  mean_days_between_doses,
  priority_population,
  lag_groups = NULL,
  lag_days = NULL,
  boosters_future = NULL,
  boosters_prepend_zero = TRUE,
  booster_proportion = rep(1L, 19)
)

Arguments

schedule_past

A vaccine_schedule object corresponding to previously carried out vaccination.

doses_future

A named vector of vaccine doses to give in the future. Names must be in ISO date format.

end_date

The final day in the future to create a schedule for. After this date the model will assume 0 vaccine doses given so an overestimate is probably better than an underestimate.

mean_days_between_doses

Assumed mean days between doses one and two

priority_population

Output from vaccine_priority_population, giving the number of people to vaccinate in each age (row) and priority group (column)

lag_groups

Row indices, corresponding to age groups in which a lag should be added to the start time of the dose schedule returned by vaccine_schedule, if NULL then no lag is added. Ignored if lag_groups is NULL.

lag_days

If lag_groups is not NULL then specifies the number of days to add the start of the dose schedule for the given groups. Ignored if lag_groups is NULL.

boosters_future

Optional named vector of booster doses to give in the future. Names must be in ISO date format.

boosters_prepend_zero

If TRUE (default) and boosters_future is not NULL then sets booster doses to zero before the first date in boosters_future. This is in contrast to when it is FALSE and the previous value in schedule_past is replicated until the first date in boosters_future. Note that this should rarely be FALSE as this will likely lead to duplicating daily doses that are already replicated in doses_future.

booster_proportion

Proportion of the groups in priority_population to boost, default is all groups; ignored if booster_daily_doses_value is NULL.

Value

A vaccine_schedule object