Create future betas

Create parameters that include beta values correspond to different Rt scenarios in the future.

add_future_betas(sample, rt, future, rt_type = "Rt_general")

future_Rt(value, relative_to = NULL)

Arguments

sample

A mcstate_pmcmc object

rt

The results of lancelot_Rt_trajectories with matrix elements Rt_general and date which will be used to calculate relative Rt values.

future

A named list of future_Rt values. Each element name corresponds to the date that the change takes effect (in ISO-8601 "YYYY-MM-DD" format) and must be strictly increasing with at least two days separating changes. Note that future values must be of length 1 or match the number of parameters in rt

rt_type

A string giving the entry of rt to modify, defaults to Rt_general

value

A value or vector of values (one for each parameter set) to add in the future. If relative_to is NULL, then this is an absolute Rt value, otherwise it is a relative value.

relative_to

Optionally an ISO 8601 (YYYY-MM-DD) format date string or Date object indicating the date that the value should be taken relative to.

Details

This function is called add_future_betas because updates beta values, however it does this by applying Rt changes in the future via future_Rt.

There are two basic ways that new beta values in the future might be added. First, as relative to some previous day in the past. For example to use a value of beta that corresponds to 1.5x the Rt recorded on the 28th of October, we might write

sircovid::future_Rt(1.5, "2020-10-28")

Second, as an absolute Rt value. For example to use a value of beta corresponding to an Rt of 0.9 we might write

sircovid::future_Rt(0.9)

These new levels need to be associated with dates at which they occur, and will result in a piece-wise constant pattern with 1 day's transition between each level.

Author

Richard Fitzjohn