The basic sircovid model The "basic" sircovid model. This is a dust model.

The basic sircovid model

The "basic" sircovid model. This is a dust model.

Examples

# Set up the basic model for England with default parameters and
# an initial seeding of early February:
p <- basic_parameters(sircovid_date("2020-02-07"), "england")
mod <- basic$new(p, 0, 10)

# Set the initial state and index as we would us for a run
# (without setting an initial state there is no seeding)
initial <- basic_initial(mod$info(), 10, p)
mod$update_state(state = initial)
mod$set_index(basic_index(mod$info())$run)

# Run the model up to the end of march
step_end <- sircovid_date("2020-03-31") / p$dt

# The filtered state is returned at the end of the run
mod$run(step_end)
#>            [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> icu         137  205  214  210  310  232  155  230  171   144
#> deaths      427  618  651  572  883  688  431  682  507   415
#> deaths_inc   78   95  103   93  143  108   71  131  100    55

# More state can be retrieved using the "state" method
mod$state(1:6)
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,]       91.0       91.0       91.0       91.0       91.0       91.0
#> [2,] 56550138.0 56550138.0 56550138.0 56550138.0 56550138.0 56550138.0
#> [3,]      137.0      205.0      214.0      210.0      310.0      232.0
#> [4,]      427.0      618.0      651.0      572.0      883.0      688.0
#> [5,]       78.0       95.0      103.0       93.0      143.0      108.0
#> [6,]        0.1        0.1        0.1        0.1        0.1        0.1
#>            [,7]       [,8]       [,9]      [,10]
#> [1,]       91.0       91.0       91.0       91.0
#> [2,] 56550138.0 56550138.0 56550138.0 56550138.0
#> [3,]      155.0      230.0      171.0      144.0
#> [4,]      431.0      682.0      507.0      415.0
#> [5,]       71.0      131.0      100.0       55.0
#> [6,]        0.1        0.1        0.1        0.1