The basic sircovid model The "basic" sircovid model. This is a dust model.

The basic sircovid model

The "basic" sircovid model. This is a dust model.

Examples

# Set up the basic model for England with default parameters and
# an initial seeding of early February:
p <- basic_parameters(sircovid_date("2020-02-07"), "england")
mod <- basic$new(p, 0, 10)

# Set the initial state and index as we would us for a run
# (without setting an initial state there is no seeding)
initial <- basic_initial(mod$info(), 10, p)
mod$set_state(initial$state, initial$step)

#> NULL

mod$set_index(basic_index(mod$info())$run)

# Run the model up to the end of march
step_end <- sircovid_date("2020-03-31") / p$dt

# The filtered state is returned at the end of the run
mod$run(step_end)

#>            [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> icu          33   26   24   25    8   12   12   10   15    14
#> deaths      102   79   74  102   34   38   49   23   58    41
#> deaths_inc   12   12    8   10    2    9    9    2    9     7

# More state can be retrieved using the "state" method
mod$state(1:6)

#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]       91.00       91.00       91.00       91.00       91.00       91.00
#> [2,] 56286961.00 56286961.00 56286961.00 56286961.00 56286961.00 56286961.00
#> [3,]       33.00       26.00       24.00       25.00        8.00       12.00
#> [4,]      102.00       79.00       74.00      102.00       34.00       38.00
#> [5,]       12.00       12.00        8.00       10.00        2.00        9.00
#> [6,]        0.08        0.08        0.08        0.08        0.08        0.08
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]       91.00       91.00       91.00       91.00
#> [2,] 56286961.00 56286961.00 56286961.00 56286961.00
#> [3,]       12.00       10.00       15.00       14.00
#> [4,]       49.00       23.00       58.00       41.00
#> [5,]        9.00        2.00        9.00        7.00
#> [6,]        0.08        0.08        0.08        0.08