Parameters for the basic model

Parameters for the "basic" model.

basic_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  beta_type = "piecewise-linear",
  severity = NULL,
  exp_noise = 1e+06
)

Arguments

start_date
Thee start date, as a sircovid_date() (i.e., the number of days into 2020)

region
The region to run the model for. This will bee used to get population data, which is currently fixed within the package and is limited to "uk", the four constituent nations ("england", "wales", "scotland", "northern_ireland") and the 7 NHS regions (e.g., "midlands"). These names are case insensitive.

beta_date
A vector of dates (each as a sircovid_date()) for changes in beta (the contact rate parameter), or NULL if a single value is used for all times (see sircovid_parameters_piecewise_linear() or sircovid_parameters_piecewise_constant(), where this is passed as date). If beta_type = "piecewise-constant", then the first date must be 0.

beta_value
A vector of values for beta (the contact rate parameter). If not given, and if beta_date is NULL then a value of 0.08 will be used through the whole simulation, otherwise if beta_date is NULL this must be a scalar. If beta_date is given then beta_date and beta_value must have the same length (see sircovid_parameters_piecewise_linear() or sircovid_parameters_piecewise_constant(), where this is passed as value).

beta_type
The type of form used for beta (the contact rate parameter), which currently can be "piecewise-linear" or "piecewise-constant"

severity
Severity data, via Bob Verity's markovid package. This needs to be NULL (use the default bundled data version in the package), a data.frame object (for raw severity data) or a list (for data that has already been processed by sircovid for use). New severity data comes from Bob Verity via the markovid package, and needs to be carefully calibrated with the progression parameters.

exp_noise
Rate of exponential noise used in the compare function - typically set to a large value so that noise is small but non-zero. If set to Inf then there is no noise in the observation process (not realistic but useful for testing).

Value

A list of inputs to the model, many of which are fixed and represent data. These correspond largely to user() calls within the odin code, though some are also used in processing just before the model is run.

Examples

basic_parameters(sircovid_date("2020-02-01"), "uk")

#> $hosp_transmission
#> [1] 0
#> 
#> $ICU_transmission
#> [1] 0
#> 
#> $G_D_transmission
#> [1] 0
#> 
#> $dt
#> [1] 0.25
#> 
#> $steps_per_day
#> [1] 4
#> 
#> $initial_step
#> [1] 128
#> 
#> $n_age_groups
#> [1] 17
#> 
#> $beta_step
#> [1] 0.08
#> 
#> $population
#>  [1] 3782330 4147413 4045114 3683680 4133158 4476630 4521975 4404100 4091543
#> [10] 4303967 4616017 4510851 3855818 3355381 3363906 2403759 3385592
#> 
#> $m
#>                 [0,4)        [5,9)      [10,14)      [15,19)      [20,24)
#> [0,4)    5.065104e-07 1.997490e-07 1.142840e-07 7.576677e-08 1.152515e-07
#> [5,9)    1.997490e-07 1.600336e-06 2.935787e-07 1.349839e-07 1.103265e-07
#> [10,14)  1.142840e-07 2.935787e-07 1.694128e-06 3.280779e-07 5.961244e-08
#> [15,19)  7.576677e-08 1.349839e-07 3.280779e-07 1.822711e-06 3.268613e-07
#> [20,24)  1.152515e-07 1.103265e-07 5.961244e-08 3.268613e-07 6.274186e-07
#> [25,29)  1.845218e-07 1.632727e-07 8.698579e-08 1.818952e-07 3.243075e-07
#> [30,34)  1.952085e-07 2.326886e-07 1.280368e-07 1.113148e-07 1.881822e-07
#> [35,39)  2.254526e-07 2.783622e-07 2.371701e-07 1.988849e-07 1.626312e-07
#> [40,44)  9.887255e-08 2.310936e-07 2.625822e-07 2.312508e-07 1.827198e-07
#> [45,49)  6.308239e-08 8.236923e-08 1.514590e-07 2.475631e-07 2.178102e-07
#> [50,54)  8.341645e-08 7.794834e-08 8.286894e-08 1.389614e-07 1.540431e-07
#> [55,59)  6.076487e-08 4.961125e-08 8.319027e-08 7.716537e-08 1.171382e-07
#> [60,64)  6.016292e-08 9.333274e-08 5.786570e-08 5.761025e-08 8.832882e-08
#> [65,69)  4.838958e-08 7.411618e-08 6.788907e-08 9.515620e-08 7.258711e-08
#> [70,74)  2.115025e-08 2.881885e-08 5.386913e-08 9.465172e-08 5.467673e-08
#> [75,79)  2.115025e-08 2.881885e-08 5.386913e-08 9.465172e-08 5.467673e-08
#> [80,100) 2.115025e-08 2.881885e-08 5.386913e-08 9.465172e-08 5.467673e-08
#>               [25,29)      [30,34)      [35,39)      [40,44)      [45,49)
#> [0,4)    1.845218e-07 1.952085e-07 2.254526e-07 9.887255e-08 6.308239e-08
#> [5,9)    1.632727e-07 2.326886e-07 2.783622e-07 2.310936e-07 8.236923e-08
#> [10,14)  8.698579e-08 1.280368e-07 2.371701e-07 2.625822e-07 1.514590e-07
#> [15,19)  1.818952e-07 1.113148e-07 1.988849e-07 2.312508e-07 2.475631e-07
#> [20,24)  3.243075e-07 1.881822e-07 1.626312e-07 1.827198e-07 2.178102e-07
#> [25,29)  4.089032e-07 2.391542e-07 1.875372e-07 1.905722e-07 2.037410e-07
#> [30,34)  2.391542e-07 3.685705e-07 2.720147e-07 2.121539e-07 1.663500e-07
#> [35,39)  1.875372e-07 2.720147e-07 3.405917e-07 3.002888e-07 2.167891e-07
#> [40,44)  1.905722e-07 2.121539e-07 3.002888e-07 3.311315e-07 3.091782e-07
#> [45,49)  2.037410e-07 1.663500e-07 2.167891e-07 3.091782e-07 4.351166e-07
#> [50,54)  1.899482e-07 1.671856e-07 1.611473e-07 1.861135e-07 1.728174e-07
#> [55,59)  1.554527e-07 1.433145e-07 1.363783e-07 1.422542e-07 1.489368e-07
#> [60,64)  1.184181e-07 1.356357e-07 1.652679e-07 1.754511e-07 1.411386e-07
#> [65,69)  1.155679e-07 8.948746e-08 1.540300e-07 1.332215e-07 8.780807e-08
#> [70,74)  4.777273e-08 3.488339e-08 4.611959e-08 1.284464e-07 1.068351e-07
#> [75,79)  4.777273e-08 3.488339e-08 4.611959e-08 1.284464e-07 1.068351e-07
#> [80,100) 4.777273e-08 3.488339e-08 4.611959e-08 1.284464e-07 1.068351e-07
#>               [50,54)      [55,59)      [60,64)      [65,69)      [70,74)
#> [0,4)    8.341645e-08 6.076487e-08 6.016292e-08 4.838958e-08 2.115025e-08
#> [5,9)    7.794834e-08 4.961125e-08 9.333274e-08 7.411618e-08 2.881885e-08
#> [10,14)  8.286894e-08 8.319027e-08 5.786570e-08 6.788907e-08 5.386913e-08
#> [15,19)  1.389614e-07 7.716537e-08 5.761025e-08 9.515620e-08 9.465172e-08
#> [20,24)  1.540431e-07 1.171382e-07 8.832882e-08 7.258711e-08 5.467673e-08
#> [25,29)  1.899482e-07 1.554527e-07 1.184181e-07 1.155679e-07 4.777273e-08
#> [30,34)  1.671856e-07 1.433145e-07 1.356357e-07 8.948746e-08 3.488339e-08
#> [35,39)  1.611473e-07 1.363783e-07 1.652679e-07 1.540300e-07 4.611959e-08
#> [40,44)  1.861135e-07 1.422542e-07 1.754511e-07 1.332215e-07 1.284464e-07
#> [45,49)  1.728174e-07 1.489368e-07 1.411386e-07 8.780807e-08 1.068351e-07
#> [50,54)  1.608365e-07 2.050266e-07 1.170549e-07 9.178605e-08 9.246504e-08
#> [55,59)  2.050266e-07 2.586356e-07 1.949070e-07 1.358154e-07 7.969682e-08
#> [60,64)  1.170549e-07 1.949070e-07 1.689694e-07 1.646502e-07 1.016733e-07
#> [65,69)  9.178605e-08 1.358154e-07 1.646502e-07 2.097239e-07 1.329549e-07
#> [70,74)  9.246504e-08 7.969682e-08 1.016733e-07 1.329549e-07 1.469234e-07
#> [75,79)  9.246504e-08 7.969682e-08 1.016733e-07 1.329549e-07 1.469234e-07
#> [80,100) 9.246504e-08 7.969682e-08 1.016733e-07 1.329549e-07 1.469234e-07
#>               [75,79)     [80,100)
#> [0,4)    2.115025e-08 2.115025e-08
#> [5,9)    2.881885e-08 2.881885e-08
#> [10,14)  5.386913e-08 5.386913e-08
#> [15,19)  9.465172e-08 9.465172e-08
#> [20,24)  5.467673e-08 5.467673e-08
#> [25,29)  4.777273e-08 4.777273e-08
#> [30,34)  3.488339e-08 3.488339e-08
#> [35,39)  4.611959e-08 4.611959e-08
#> [40,44)  1.284464e-07 1.284464e-07
#> [45,49)  1.068351e-07 1.068351e-07
#> [50,54)  9.246504e-08 9.246504e-08
#> [55,59)  7.969682e-08 7.969682e-08
#> [60,64)  1.016733e-07 1.016733e-07
#> [65,69)  1.329549e-07 1.329549e-07
#> [70,74)  1.469234e-07 1.469234e-07
#> [75,79)  1.469234e-07 1.469234e-07
#> [80,100) 1.469234e-07 1.469234e-07
#> 
#> $p_star
#>  [1] 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2
#> 
#> $p_C
#>  [1] 0.1809 0.1809 0.1809 0.1809 0.2241 0.2241 0.2241 0.2241 0.3054 0.3054
#> [11] 0.3054 0.3054 0.3546 0.3546 0.3546 0.3546 0.6456
#> 
#> $p_G_D
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05
#> [16] 0.05 0.05
#> 
#> $p_H_D
#>  [1] 0.1258933 0.1226134 0.1356729 0.1526679 0.1743031 0.1941879 0.2093617
#>  [8] 0.2244326 0.2370135 0.2578281 0.2908746 0.3207640 0.3625638 0.3909655
#> [15] 0.4211515 0.4475459 0.4820000
#> 
#> $p_ICU_D
#>  [1] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#>  [8] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> [15] 0.5234896 0.5234896 0.5234896
#> 
#> $p_W_D
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05
#> [16] 0.05 0.05
#> 
#> $p_ICU
#>  [1] 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387
#>  [7] 0.11845387 0.12551064 0.14091042 0.17595537 0.24437611 0.35243489
#> [13] 0.52779221 0.72404472 0.18197685 0.05439050 0.01673267
#> 
#> $p_sero_pos_1
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394
#> 
#> $p_sero_pos_2
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394
#> 
#> $p_H
#>  [1] 0.000632689 0.000539148 0.000995640 0.002035807 0.004545371 0.008748134
#>  [7] 0.013802147 0.019850725 0.024607803 0.032822485 0.049084367 0.061571007
#> [13] 0.086379194 0.099452536 0.124208801 0.150169192 0.153030303
#> 
#> $p_R
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_recov_ICU
#>  [1] 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104
#>  [8] 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104 0.4765104
#> [15] 0.4765104 0.4765104 0.4765104
#> 
#> $p_recov_hosp
#>  [1] 0.7705654 0.7734568 0.7619442 0.7469624 0.7278899 0.7103605 0.6969841
#>  [8] 0.6782255 0.6554737 0.6115828 0.5358321 0.4398496 0.3010024 0.1680663
#> [15] 0.4735115 0.5224059 0.5093325
#> 
#> $p_death_hosp
#>  [1] 0.1109807 0.1080894 0.1196019 0.1345838 0.1536562 0.1711856 0.1845620
#>  [8] 0.1962639 0.2036158 0.2124618 0.2197918 0.2077156 0.1712054 0.1078890
#> [15] 0.3445117 0.4232036 0.4739349
#> 
#> $p_recov_sympt
#>  [1] 0.9993673 0.9994609 0.9990044 0.9979642 0.9954546 0.9912519 0.9861979
#>  [8] 0.9801493 0.9753922 0.9671775 0.9509156 0.9384290 0.9136208 0.9005475
#> [15] 0.8757912 0.8498308 0.8469697
#> 
#> $k_E
#> [1] 2
#> 
#> $k_A
#> [1] 1
#> 
#> $k_C
#> [1] 1
#> 
#> $k_hosp
#> [1] 2
#> 
#> $k_ICU
#> [1] 2
#> 
#> $k_rec
#> [1] 2
#> 
#> $gamma_E
#> [1] 0.4357298
#> 
#> $gamma_A
#> [1] 0.4784689
#> 
#> $gamma_C
#> [1] 0.25
#> 
#> $gamma_hosp
#> [1] 2
#> 
#> $gamma_ICU
#> [1] 0.4
#> 
#> $gamma_rec
#> [1] 0.4
#> 
#> $phi_ICU
#> [1] 0.95
#> 
#> $kappa_ICU
#> [1] 2
#> 
#> $phi_death
#> [1] 0.9087341
#> 
#> $kappa_death
#> [1] 2
#> 
#> $exp_noise
#> [1] 1e+06
#>