Index of carehomes model

Index of "interesting" elements for the carehomes model. This function conforms to the mcstate interface.

carehomes_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_index(mod$info())

#> $run
#>                                icu                            general 
#>                                 10                                 11 
#>               deaths_carehomes_inc                    deaths_comm_inc 
#>                                 17                                 15 
#>                    deaths_hosp_inc                       admitted_inc 
#>                                 18                                  2 
#>                      diagnoses_inc                           sero_pos 
#>                                  3                                 20 
#>                    sympt_cases_inc             sympt_cases_over25_inc 
#>                                 24                                 25 
#> sympt_cases_non_variant_over25_inc                          react_pos 
#>                                 26                                 27 
#> 
#> $state
#>                            icu                        general 
#>                             10                             11 
#>               deaths_carehomes                    deaths_comm 
#>                             16                             14 
#>                    deaths_hosp                       admitted 
#>                             13                              5 
#>                      diagnoses                       sero_pos 
#>                              6                             20 
#>                    sympt_cases             sympt_cases_over25 
#>                             21                             22 
#> sympt_cases_non_variant_over25                      react_pos 
#>                             23                             27 
#>                           hosp                         deaths 
#>                             12                             19 
#>                     infections                            S_0 
#>                              4                            219 
#>                            S_5                           S_10 
#>                            220                            221 
#>                           S_15                           S_20 
#>                            222                            223 
#>                           S_25                           S_30 
#>                            224                            225 
#>                           S_35                           S_40 
#>                            226                            227 
#>                           S_45                           S_50 
#>                            228                            229 
#>                           S_55                           S_60 
#>                            230                            231 
#>                           S_65                           S_70 
#>                            232                            233 
#>                           S_75                           S_80 
#>                            234                            235 
#>                          S_CHW                          S_CHR 
#>                            236                            237 
#>                    cum_admit_0                    cum_admit_5 
#>                             67                             68 
#>                   cum_admit_10                   cum_admit_15 
#>                             69                             70 
#>                   cum_admit_20                   cum_admit_25 
#>                             71                             72 
#>                   cum_admit_30                   cum_admit_35 
#>                             73                             74 
#>                   cum_admit_40                   cum_admit_45 
#>                             75                             76 
#>                   cum_admit_50                   cum_admit_55 
#>                             77                             78 
#>                   cum_admit_60                   cum_admit_65 
#>                             79                             80 
#>                   cum_admit_70                   cum_admit_75 
#>                             81                             82 
#>                   cum_admit_80                  cum_admit_CHW 
#>                             83                             84 
#>                  cum_admit_CHR                       D_hosp_0 
#>                             85                             29 
#>                       D_hosp_5                      D_hosp_10 
#>                             30                             31 
#>                      D_hosp_15                      D_hosp_20 
#>                             32                             33 
#>                      D_hosp_25                      D_hosp_30 
#>                             34                             35 
#>                      D_hosp_35                      D_hosp_40 
#>                             36                             37 
#>                      D_hosp_45                      D_hosp_50 
#>                             38                             39 
#>                      D_hosp_55                      D_hosp_60 
#>                             40                             41 
#>                      D_hosp_65                      D_hosp_70 
#>                             42                             43 
#>                      D_hosp_75                      D_hosp_80 
#>                             44                             45 
#>                     D_hosp_CHW                     D_hosp_CHR 
#>                             46                             47 
#>                   I_weighted_0                   I_weighted_5 
#>                            257                            258 
#>                  I_weighted_10                  I_weighted_15 
#>                            259                            260 
#>                  I_weighted_20                  I_weighted_25 
#>                            261                            262 
#>                  I_weighted_30                  I_weighted_35 
#>                            263                            264 
#>                  I_weighted_40                  I_weighted_45 
#>                            265                            266 
#>                  I_weighted_50                  I_weighted_55 
#>                            267                            268 
#>                  I_weighted_60                  I_weighted_65 
#>                            269                            270 
#>                  I_weighted_70                  I_weighted_75 
#>                            271                            272 
#>                  I_weighted_80                 I_weighted_CHW 
#>                            273                            274 
#>                 I_weighted_CHR                  prob_strain_0 
#>                            275                            238 
#>                  prob_strain_5                 prob_strain_10 
#>                            239                            240 
#>                 prob_strain_15                 prob_strain_20 
#>                            241                            242 
#>                 prob_strain_25                 prob_strain_30 
#>                            243                            244 
#>                 prob_strain_35                 prob_strain_40 
#>                            245                            246 
#>                 prob_strain_45                 prob_strain_50 
#>                            247                            248 
#>                 prob_strain_55                 prob_strain_60 
#>                            249                            250 
#>                 prob_strain_65                 prob_strain_70 
#>                            251                            252 
#>                 prob_strain_75                 prob_strain_80 
#>                            253                            254 
#>                prob_strain_CHW                prob_strain_CHR 
#>                            255                            256 
#>