Index of carehomes model

Index of "interesting" elements for the carehomes model. This function conforms to the mcstate interface.

carehomes_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_index(mod$info())

#> $run
#>                                icu                            general 
#>                                 10                                 11 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 15                                 16 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                           sero_pos                    sympt_cases_inc 
#>                                 18                                 22 
#>             sympt_cases_over25_inc sympt_cases_non_variant_over25_inc 
#>                                 23                                 24 
#>                          react_pos 
#>                                 25 
#> 
#> $state
#>                            icu                        general 
#>                             10                             11 
#>                    deaths_comm                    deaths_hosp 
#>                             14                             13 
#>                       admitted                      diagnoses 
#>                              5                              6 
#>                       sero_pos                    sympt_cases 
#>                             18                             19 
#>             sympt_cases_over25 sympt_cases_non_variant_over25 
#>                             20                             21 
#>                      react_pos                           hosp 
#>                             25                             12 
#>                         deaths                     infections 
#>                             17                              4 
#>                            S_0                            S_5 
#>                            198                            199 
#>                           S_10                           S_15 
#>                            200                            201 
#>                           S_20                           S_25 
#>                            202                            203 
#>                           S_30                           S_35 
#>                            204                            205 
#>                           S_40                           S_45 
#>                            206                            207 
#>                           S_50                           S_55 
#>                            208                            209 
#>                           S_60                           S_65 
#>                            210                            211 
#>                           S_70                           S_75 
#>                            212                            213 
#>                           S_80                          S_CHW 
#>                            214                            215 
#>                          S_CHR                    cum_admit_0 
#>                            216                             65 
#>                    cum_admit_5                   cum_admit_10 
#>                             66                             67 
#>                   cum_admit_15                   cum_admit_20 
#>                             68                             69 
#>                   cum_admit_25                   cum_admit_30 
#>                             70                             71 
#>                   cum_admit_35                   cum_admit_40 
#>                             72                             73 
#>                   cum_admit_45                   cum_admit_50 
#>                             74                             75 
#>                   cum_admit_55                   cum_admit_60 
#>                             76                             77 
#>                   cum_admit_65                   cum_admit_70 
#>                             78                             79 
#>                   cum_admit_75                   cum_admit_80 
#>                             80                             81 
#>                  cum_admit_CHW                  cum_admit_CHR 
#>                             82                             83 
#>                   I_weighted_0                   I_weighted_5 
#>                            236                            237 
#>                  I_weighted_10                  I_weighted_15 
#>                            238                            239 
#>                  I_weighted_20                  I_weighted_25 
#>                            240                            241 
#>                  I_weighted_30                  I_weighted_35 
#>                            242                            243 
#>                  I_weighted_40                  I_weighted_45 
#>                            244                            245 
#>                  I_weighted_50                  I_weighted_55 
#>                            246                            247 
#>                  I_weighted_60                  I_weighted_65 
#>                            248                            249 
#>                  I_weighted_70                  I_weighted_75 
#>                            250                            251 
#>                  I_weighted_80                 I_weighted_CHW 
#>                            252                            253 
#>                 I_weighted_CHR                  prob_strain_0 
#>                            254                            217 
#>                  prob_strain_5                 prob_strain_10 
#>                            218                            219 
#>                 prob_strain_15                 prob_strain_20 
#>                            220                            221 
#>                 prob_strain_25                 prob_strain_30 
#>                            222                            223 
#>                 prob_strain_35                 prob_strain_40 
#>                            224                            225 
#>                 prob_strain_45                 prob_strain_50 
#>                            226                            227 
#>                 prob_strain_55                 prob_strain_60 
#>                            228                            229 
#>                 prob_strain_65                 prob_strain_70 
#>                            230                            231 
#>                 prob_strain_75                 prob_strain_80 
#>                            232                            233 
#>                prob_strain_CHW                prob_strain_CHR 
#>                            234                            235 
#>