Index of carehomes model

Index of "interesting" elements for the carehomes model. This function conforms to the mcstate interface.

carehomes_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_index(mod$info())

#> $run
#>                                icu                            general 
#>                                 11                                 12 
#>               deaths_carehomes_inc                    deaths_comm_inc 
#>                                 18                                 16 
#>                    deaths_hosp_inc                       admitted_inc 
#>                                 19                                  2 
#>                      diagnoses_inc                         sero_pos_1 
#>                                  3                                 21 
#>                         sero_pos_2                    sympt_cases_inc 
#>                                 22                                 26 
#>             sympt_cases_over25_inc sympt_cases_non_variant_over25_inc 
#>                                 27                                 28 
#>                          react_pos 
#>                                 29 
#> 
#> $state
#>                            icu                        general 
#>                             11                             12 
#>               deaths_carehomes                    deaths_comm 
#>                             17                             15 
#>                    deaths_hosp                       admitted 
#>                             14                              5 
#>                      diagnoses                     sero_pos_1 
#>                              6                             21 
#>                     sero_pos_2                    sympt_cases 
#>                             22                             23 
#>             sympt_cases_over25 sympt_cases_non_variant_over25 
#>                             24                             25 
#>                      react_pos                           hosp 
#>                             29                             13 
#>                         deaths                     infections 
#>                             20                              4 
#>                            S_0                            S_5 
#>                            260                            261 
#>                           S_10                           S_15 
#>                            262                            263 
#>                           S_20                           S_25 
#>                            264                            265 
#>                           S_30                           S_35 
#>                            266                            267 
#>                           S_40                           S_45 
#>                            268                            269 
#>                           S_50                           S_55 
#>                            270                            271 
#>                           S_60                           S_65 
#>                            272                            273 
#>                           S_70                           S_75 
#>                            274                            275 
#>                           S_80                          S_CHW 
#>                            276                            277 
#>                          S_CHR                            R_0 
#>                            278                            431 
#>                            R_5                           R_10 
#>                            432                            433 
#>                           R_15                           R_20 
#>                            434                            435 
#>                           R_25                           R_30 
#>                            436                            437 
#>                           R_35                           R_40 
#>                            438                            439 
#>                           R_45                           R_50 
#>                            440                            441 
#>                           R_55                           R_60 
#>                            442                            443 
#>                           R_65                           R_70 
#>                            444                            445 
#>                           R_75                           R_80 
#>                            446                            447 
#>                          R_CHW                          R_CHR 
#>                            448                            449 
#>                    cum_admit_0                    cum_admit_5 
#>                             69                             70 
#>                   cum_admit_10                   cum_admit_15 
#>                             71                             72 
#>                   cum_admit_20                   cum_admit_25 
#>                             73                             74 
#>                   cum_admit_30                   cum_admit_35 
#>                             75                             76 
#>                   cum_admit_40                   cum_admit_45 
#>                             77                             78 
#>                   cum_admit_50                   cum_admit_55 
#>                             79                             80 
#>                   cum_admit_60                   cum_admit_65 
#>                             81                             82 
#>                   cum_admit_70                   cum_admit_75 
#>                             83                             84 
#>                   cum_admit_80                  cum_admit_CHW 
#>                             85                             86 
#>                  cum_admit_CHR                       D_hosp_0 
#>                             87                             31 
#>                       D_hosp_5                      D_hosp_10 
#>                             32                             33 
#>                      D_hosp_15                      D_hosp_20 
#>                             34                             35 
#>                      D_hosp_25                      D_hosp_30 
#>                             36                             37 
#>                      D_hosp_35                      D_hosp_40 
#>                             38                             39 
#>                      D_hosp_45                      D_hosp_50 
#>                             40                             41 
#>                      D_hosp_55                      D_hosp_60 
#>                             42                             43 
#>                      D_hosp_65                      D_hosp_70 
#>                             44                             45 
#>                      D_hosp_75                      D_hosp_80 
#>                             46                             47 
#>                     D_hosp_CHW                     D_hosp_CHR 
#>                             48                             49 
#>                   I_weighted_0                   I_weighted_5 
#>                            279                            280 
#>                  I_weighted_10                  I_weighted_15 
#>                            281                            282 
#>                  I_weighted_20                  I_weighted_25 
#>                            283                            284 
#>                  I_weighted_30                  I_weighted_35 
#>                            285                            286 
#>                  I_weighted_40                  I_weighted_45 
#>                            287                            288 
#>                  I_weighted_50                  I_weighted_55 
#>                            289                            290 
#>                  I_weighted_60                  I_weighted_65 
#>                            291                            292 
#>                  I_weighted_70                  I_weighted_75 
#>                            293                            294 
#>                  I_weighted_80                 I_weighted_CHW 
#>                            295                            296 
#>                 I_weighted_CHR                    prob_strain 
#>                            297                            107 
#>             cum_n_vaccinated_0             cum_n_vaccinated_5 
#>                            203                            204 
#>            cum_n_vaccinated_10            cum_n_vaccinated_15 
#>                            205                            206 
#>            cum_n_vaccinated_20            cum_n_vaccinated_25 
#>                            207                            208 
#>            cum_n_vaccinated_30            cum_n_vaccinated_35 
#>                            209                            210 
#>            cum_n_vaccinated_40            cum_n_vaccinated_45 
#>                            211                            212 
#>            cum_n_vaccinated_50            cum_n_vaccinated_55 
#>                            213                            214 
#>            cum_n_vaccinated_60            cum_n_vaccinated_65 
#>                            215                            216 
#>            cum_n_vaccinated_70            cum_n_vaccinated_75 
#>                            217                            218 
#>            cum_n_vaccinated_80           cum_n_vaccinated_CHW 
#>                            219                            220 
#>           cum_n_vaccinated_CHR 
#>                            221 
#>