Index of carehomes model

Index of "interesting" elements for the carehomes model. This function conforms to the mcstate interface.

carehomes_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_index(mod$info())

#> $run
#>                               time                                icu 
#>                                  1                                 12 
#>                            general               deaths_carehomes_inc 
#>                                 13                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 22                                 23 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 28                                 29 
#>             sympt_cases_over25_inc sympt_cases_non_variant_over25_inc 
#>                                 30                                 31 
#>                          react_pos 
#>                                 32 
#> 
#> $state
#>                                icu                            general 
#>                                 12                                 13 
#>                        deaths_comm                   deaths_carehomes 
#>                                 16                                 18 
#>                        deaths_hosp                           admitted 
#>                                 15                                  6 
#>                          diagnoses               deaths_carehomes_inc 
#>                                  7                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 22                                 23 
#>                        sympt_cases            sympt_cases_non_variant 
#>                                 24                                 25 
#>                 sympt_cases_over25     sympt_cases_non_variant_over25 
#>                                 26                                 27 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 28                                 29 
#>             sympt_cases_over25_inc sympt_cases_non_variant_over25_inc 
#>                                 30                                 31 
#>                          react_pos                               hosp 
#>                                 32                                 14 
#>                             deaths                         infections 
#>                                 21                                  4 
#>                     infections_inc                                S_0 
#>                                  5                                282 
#>                                S_5                               S_10 
#>                                283                                284 
#>                               S_15                               S_20 
#>                                285                                286 
#>                               S_25                               S_30 
#>                                287                                288 
#>                               S_35                               S_40 
#>                                289                                290 
#>                               S_45                               S_50 
#>                                291                                292 
#>                               S_55                               S_60 
#>                                293                                294 
#>                               S_65                               S_70 
#>                                295                                296 
#>                               S_75                               S_80 
#>                                297                                298 
#>                              S_CHW                              S_CHR 
#>                                299                                300 
#>                                R_0                                R_5 
#>                                434                                435 
#>                               R_10                               R_15 
#>                                436                                437 
#>                               R_20                               R_25 
#>                                438                                439 
#>                               R_30                               R_35 
#>                                440                                441 
#>                               R_40                               R_45 
#>                                442                                443 
#>                               R_50                               R_55 
#>                                444                                445 
#>                               R_60                               R_65 
#>                                446                                447 
#>                               R_70                               R_75 
#>                                448                                449 
#>                               R_80                              R_CHW 
#>                                450                                451 
#>                              R_CHR                        cum_admit_0 
#>                                452                                 72 
#>                        cum_admit_5                       cum_admit_10 
#>                                 73                                 74 
#>                       cum_admit_15                       cum_admit_20 
#>                                 75                                 76 
#>                       cum_admit_25                       cum_admit_30 
#>                                 77                                 78 
#>                       cum_admit_35                       cum_admit_40 
#>                                 79                                 80 
#>                       cum_admit_45                       cum_admit_50 
#>                                 81                                 82 
#>                       cum_admit_55                       cum_admit_60 
#>                                 83                                 84 
#>                       cum_admit_65                       cum_admit_70 
#>                                 85                                 86 
#>                       cum_admit_75                       cum_admit_80 
#>                                 87                                 88 
#>                      cum_admit_CHW                      cum_admit_CHR 
#>                                 89                                 90 
#>                           D_hosp_0                           D_hosp_5 
#>                                 34                                 35 
#>                          D_hosp_10                          D_hosp_15 
#>                                 36                                 37 
#>                          D_hosp_20                          D_hosp_25 
#>                                 38                                 39 
#>                          D_hosp_30                          D_hosp_35 
#>                                 40                                 41 
#>                          D_hosp_40                          D_hosp_45 
#>                                 42                                 43 
#>                          D_hosp_50                          D_hosp_55 
#>                                 44                                 45 
#>                          D_hosp_60                          D_hosp_65 
#>                                 46                                 47 
#>                          D_hosp_70                          D_hosp_75 
#>                                 48                                 49 
#>                          D_hosp_80                         D_hosp_CHW 
#>                                 50                                 51 
#>                         D_hosp_CHR                            D_all_0 
#>                                 52                                263 
#>                            D_all_5                           D_all_10 
#>                                264                                265 
#>                           D_all_15                           D_all_20 
#>                                266                                267 
#>                           D_all_25                           D_all_30 
#>                                268                                269 
#>                           D_all_35                           D_all_40 
#>                                270                                271 
#>                           D_all_45                           D_all_50 
#>                                272                                273 
#>                           D_all_55                           D_all_60 
#>                                274                                275 
#>                           D_all_65                           D_all_70 
#>                                276                                277 
#>                           D_all_75                           D_all_80 
#>                                278                                279 
#>                          D_all_CHW                          D_all_CHR 
#>                                280                                281 
#>               diagnoses_admitted_0               diagnoses_admitted_5 
#>                                225                                226 
#>              diagnoses_admitted_10              diagnoses_admitted_15 
#>                                227                                228 
#>              diagnoses_admitted_20              diagnoses_admitted_25 
#>                                229                                230 
#>              diagnoses_admitted_30              diagnoses_admitted_35 
#>                                231                                232 
#>              diagnoses_admitted_40              diagnoses_admitted_45 
#>                                233                                234 
#>              diagnoses_admitted_50              diagnoses_admitted_55 
#>                                235                                236 
#>              diagnoses_admitted_60              diagnoses_admitted_65 
#>                                237                                238 
#>              diagnoses_admitted_70              diagnoses_admitted_75 
#>                                239                                240 
#>              diagnoses_admitted_80             diagnoses_admitted_CHW 
#>                                241                                242 
#>             diagnoses_admitted_CHR             cum_infections_disag_0 
#>                                243                                244 
#>             cum_infections_disag_5            cum_infections_disag_10 
#>                                245                                246 
#>            cum_infections_disag_15            cum_infections_disag_20 
#>                                247                                248 
#>            cum_infections_disag_25            cum_infections_disag_30 
#>                                249                                250 
#>            cum_infections_disag_35            cum_infections_disag_40 
#>                                251                                252 
#>            cum_infections_disag_45            cum_infections_disag_50 
#>                                253                                254 
#>            cum_infections_disag_55            cum_infections_disag_60 
#>                                255                                256 
#>            cum_infections_disag_65            cum_infections_disag_70 
#>                                257                                258 
#>            cum_infections_disag_75            cum_infections_disag_80 
#>                                259                                260 
#>           cum_infections_disag_CHW           cum_infections_disag_CHR 
#>                                261                                262 
#>                       I_weighted_0                       I_weighted_5 
#>                                472                                473 
#>                      I_weighted_10                      I_weighted_15 
#>                                474                                475 
#>                      I_weighted_20                      I_weighted_25 
#>                                476                                477 
#>                      I_weighted_30                      I_weighted_35 
#>                                478                                479 
#>                      I_weighted_40                      I_weighted_45 
#>                                480                                481 
#>                      I_weighted_50                      I_weighted_55 
#>                                482                                483 
#>                      I_weighted_60                      I_weighted_65 
#>                                484                                485 
#>                      I_weighted_70                      I_weighted_75 
#>                                486                                487 
#>                      I_weighted_80                     I_weighted_CHW 
#>                                488                                489 
#>                     I_weighted_CHR                        prob_strain 
#>                                490                                110 
#>                 cum_n_vaccinated_0                 cum_n_vaccinated_5 
#>                                206                                207 
#>                cum_n_vaccinated_10                cum_n_vaccinated_15 
#>                                208                                209 
#>                cum_n_vaccinated_20                cum_n_vaccinated_25 
#>                                210                                211 
#>                cum_n_vaccinated_30                cum_n_vaccinated_35 
#>                                212                                213 
#>                cum_n_vaccinated_40                cum_n_vaccinated_45 
#>                                214                                215 
#>                cum_n_vaccinated_50                cum_n_vaccinated_55 
#>                                216                                217 
#>                cum_n_vaccinated_60                cum_n_vaccinated_65 
#>                                218                                219 
#>                cum_n_vaccinated_70                cum_n_vaccinated_75 
#>                                220                                221 
#>                cum_n_vaccinated_80               cum_n_vaccinated_CHW 
#>                                222                                223 
#>               cum_n_vaccinated_CHR 
#>                                224 
#>