Initial conditions for the carehomes model

Create initial conditions for the carehomes model. This matches the interface required for mcstate

carehomes_initial(info, n_particles, pars)

Arguments

info
The result of running the $info() method on an initialised basic model

n_particles
The number of particles required. Currently only uniform initial seeding is implemented so this has no effect

pars
A parameter list created by basic_parameters(); from this list we will use the population and initial_step elements.

Value

A numeric vector of initial conditions

Examples

p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_initial(mod$info(), 10, p)

#> $state
#>    [1]        0.00        0.00        0.00        0.00        0.00        0.00
#>    [7]        0.00        0.00 56286962.00 56286962.00 56286962.00        0.00
#>   [13]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [19]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [25]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [31]        0.00       10.00        0.00        0.00        0.00        0.00
#>   [37]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [43]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [49]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [55]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [61]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [67]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [73]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [79]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [85]        0.00        0.00        0.00        0.00        0.00        0.00
#>   [91]  3299637.00  3538206.00  3354246.00  3090232.00  3487863.00  3758950.00
#>   [97]  3765495.00  3691183.00  3371838.00  3673353.00  3865002.00  3628192.00
#>  [103]  3069376.00  2779756.00  2762342.00  1889735.00  2582208.00   339674.00
#>  [109]   339674.00        1.00        0.00        0.00        0.00        0.00
#>  [115]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [121]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [127]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [133]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [139]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [145]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [151]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [157]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [163]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [169]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [175]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [181]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [187]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [193]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [199]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [205]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [211]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [217]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [223]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [229]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [235]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [241]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [247]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [253]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [259]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [265]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [271]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [277]        0.00        0.00        0.00        0.00        0.00  3299637.00
#>  [283]  3538206.00  3354246.00  3090222.00  3487863.00  3758950.00  3765495.00
#>  [289]  3691183.00  3371838.00  3673353.00  3865002.00  3628192.00  3069376.00
#>  [295]  2779756.00  2762342.00  1889735.00  2582208.00   339674.00   339674.00
#>  [301]        0.00        0.00        0.00        3.63        0.00        0.00
#>  [307]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [313]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [319]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [325]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [331]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [337]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [343]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [349]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [355]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [361]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [367]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [373]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [379]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [385]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [391]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [397]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [403]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [409]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [415]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [421]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [427]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [433]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [439]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [445]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [451]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [457]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [463]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [469]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [475]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [481]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [487]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [493]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [499]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [505]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [511]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [517]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [523]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [529]        0.00        0.00        0.00       10.00        0.00        0.00
#>  [535]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [541]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [547]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [553]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [559]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [565]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [571]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [577]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [583]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [589]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [595]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [601]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [607]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [613]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [619]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [625]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [631]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [637]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [643]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [649]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [655]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [661]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [667]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [673]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [679]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [685]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [691]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [697]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [703]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [709]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [715]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [721]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [727]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [733]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [739]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [745]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [751]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [757]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [763]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [769]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [775]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [781]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [787]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [793]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [799]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [805]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [811]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [817]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [823]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [829]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [835]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [841]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [847]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [853]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [859]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [865]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [871]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [877]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [883]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [889]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [895]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [901]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [907]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [913]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [919]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [925]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [931]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [937]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [943]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [949]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [955]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [961]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [967]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [973]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [979]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [985]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [991]        0.00        0.00        0.00        0.00        0.00        0.00
#>  [997]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1003]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1009]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1015]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1021]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1027]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1033]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1039]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1045]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1051]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1057]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1063]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1069]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1075]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1081]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1087]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1093]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1099]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1105]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1111]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1117]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1123]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1129]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1135]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1141]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1147]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1153]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1159]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1165]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1171]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1177]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1183]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1189]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1195]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1201]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1207]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1213]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1219]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1225]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1231]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1237]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1243]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1249]        0.00        0.00        0.00        0.00        0.00       10.00
#> [1255]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1261]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1267]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1273]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1279]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1285]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1291]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1297]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1303]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1309]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1315]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1321]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1327]        0.00        0.00        0.00       10.00        0.00        0.00
#> [1333]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1339]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1345]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1351]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1357]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1363]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1369]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1375]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1381]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1387]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1393]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1399]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1405]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1411]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1417]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1423]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1429]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1435]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1441]        0.00        0.00        0.00       10.00        0.00        0.00
#> [1447]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1453]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1459]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1465]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1471]        0.00        0.00        0.00        0.00        0.00        0.00
#> [1477]        0.00        0.00
#> 
#> $step
#> [1] 152
#>