Parameters for the carehomes model

Parameters for the "carehomes" model.

carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  p_death_carehome = 0.7,
  sero_specificity = 0.9,
  sero_sensitivity = 0.99,
  progression = NULL,
  eps = 0.1,
  m_CHW = 4e-06,
  m_CHR = 5e-05,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99,
  p_NC = 0.01,
  strain_transmission = 1,
  strain_seed_date = NULL,
  strain_seed_value = NULL,
  rel_susceptibility = 1,
  rel_p_sympt = 1,
  rel_p_hosp_if_sympt = 1,
  rel_infectivity = 1,
  vaccine_progression_rate = NULL,
  vaccine_uptake = NULL,
  vaccine_daily_doses = 0,
  waning_rate = 0,
  model_pcr_and_serology_user = 1,
  exp_noise = 1e+06
)

Arguments

start_date
Thee start date, as a sircovid_date() (i.e., the number of days into 2020)

region
The region to run the model for. This will bee used to get population data, which is currently fixed within the package and is limited to "uk", the four constituent nations ("england", "wales", "scotland", "northern_ireland") and the 7 NHS regions (e.g., "midlands"). These names are case insensitive.

beta_date
A vector of dates for changes in beta (the contact rate parameter), or NULL if a single value is used for all times (see sircovid_parameters_beta(), where this is passed as date).

beta_value
A vector of values for beta (the contact rate parameter). If not given, and if beta_date is NULL then a value of 0.08 will be used through the whole simulation, otherwise if beta_date is NULL this must be a scalar. If beta_date is given then beta_date and beta_value must have the same length (see sircovid_parameters_beta(), where this is passed as value).

severity
Severity data, via Bob Verity's markovid package. This needs to be NULL (use the default bundled data version in the package), a data.frame object (for raw severity data) or a list (for data that has already been processed by sircovid for use). New severity data comes from Bob Verity via the markovid package, and needs to be carefully calibrated with the progression parameters.

p_death_carehome
Probability of death within carehomes (conditional on having an "inflenza-like-illness")

sero_specificity
Specificity of the serology test

sero_sensitivity
Sensitivity of the serology test

progression
Progression data

eps
Change in contact rate for carehome residents

m_CHW
Contact rate between carehome workers and either residents or workers

m_CHR
Contact rate between carehome residents

pillar2_specificity
Specificity of the Pillar 2 test

pillar2_sensitivity
Sensitivity of the Pillar 2 test

react_specificity
Specificity of the REACT test

react_sensitivity
Sensitivity of the REACT test

p_NC
Proportion of population who do not have covid but have covid-like symptoms

strain_transmission
Vector of relative transmissibility of each strain modelled. First element should be 1. Length will define the number of strains used in the model

strain_seed_date
Either NULL (no seeding) or a vector of exactly two sircovid_date values corresponding to the start and stop dates of seeding (inclusive). For example, to seed for a single day, these two values should have the same date

strain_seed_value
Either NULL (no seeding) or a single integer value represending the daily rate of seeding. Because each day is dividied into 1 / dt steps, this value will be spread out fairly evenly across the steps that occur within each date. Seeding is not stochastic; this many individuals will become infected with the new strain, unless the pool of susceptibles has been exhausted.

rel_susceptibility
A vector or matrix of values representing the relative susceptibility of individuals in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case relative susceptibility will be the same across all age groups within one vaccination category. Specifying a matrix instead of a vector allows different relative susceptibilities by age (rows of the matrix) and vaccination group (columns of the matrix); in that case, in each row of the matrix, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1

rel_p_sympt
A vector or matrix of values of same dimension as rel_susceptibility representing the relative probability of symptomatic infection in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case the relative reduction in probability of symptomatic infection will be the same across all age groups within one vaccination category. Specifying a matrix instead of a vector allows different relative reductions in probability of symptomatic infection by age (rows of the matrix) and vaccination group (columns of the matrix); in that case, in each row of the matrix, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1

rel_p_hosp_if_sympt
A vector or matrix of values of same dimension as rel_susceptibility representing the relative probability of hospitalisation for symptomatic cases in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case the relative reduction in probability of hospitalisation for symptomatic cases will be the same across all age groups within one vaccination category. Specifying a matrix instead of a vector allows different relative reductions in probability of hospitalisation for symptomatic cases by age (rows of the matrix) and vaccination group (columns of the matrix); in that case, in each row of the matrix, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1

rel_infectivity
A vector or matrix of values representing the relative infectivity of individuals in different vaccination groups, if they are infected. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case relative infectivity will be the same across all age groups within one vaccination category. Specifying a matrix instead of a vector allows different relative infectivities by age (rows of the matrix) and vaccination group (columns of the matrix); in that case, in each row of the matrix, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1

vaccine_progression_rate
A vector or matrix of values of same dimension as rel_susceptibility representing the rate of movement between different vaccination classes. If a vector, it should have as many values as vaccination classes, and the same rates of progression will be used for all age groups (the first rate is the vaccination rate and the last of the rates is the rate of returning to the initial vacination class); if a matrix, the element on row i and column j is the rate of progression from the jth vaccination class to the (j+1)th for age group i.

vaccine_uptake
A vector of length 19 with the proportion of the population who are able to be vaccinated.

vaccine_daily_doses
A single value indicating the number of (first) vaccine doses per day to distribute. The actual number distributed will be stochastically distributed around this, provided there are sufficient eligible candidates.

waning_rate
A single value or a vector of values representing the rates of waning of immunity after infection; if a single value the same rate is used for all age groups; if a vector of values if used it should have one value per age group.

model_pcr_and_serology_user
A value of 1 or 0 so switch on or off the flows out of T_PCR_neg and T_sero_neg and the corresponding cap on the number of individuals leaving the R compartments

exp_noise
Rate of exponential noise used in the compare function - typically set to a large value so that noise is small but non-zero. If set to Inf then there is no noise in the observation process (not realistic but useful for testing).

Value

A list of inputs to the model, many of which are fixed and represent data. These correspond largely to user() calls within the odin code, though some are also used in processing just before the model is run.

Examples

carehomes_parameters(sircovid_date("2020-02-01"), "uk")

#> $hosp_transmission
#> [1] 0.1
#> 
#> $ICU_transmission
#> [1] 0.05
#> 
#> $G_D_transmission
#> [1] 0.05
#> 
#> $dt
#> [1] 0.25
#> 
#> $initial_step
#> [1] 128
#> 
#> $n_age_groups
#> [1] 17
#> 
#> $beta_step
#> [1] 0.08
#> 
#> $population
#>  [1] 3857263 4149852 3953866 3656968 4153080 4514249 4497132 4395667 4019539
#> [10] 4402122 4661015 4405908 3755185 3368199 3318867 2325296 3362599
#> 
#> $carehome_beds
#> [1] 537521
#> 
#> $carehome_residents
#> [1] 398841
#> 
#> $carehome_workers
#> [1] 398841
#> 
#> $m
#>                 [0,4)        [5,9)      [10,14)      [15,19)      [20,24)
#> [0,4)    4.966707e-07 1.980202e-07 1.146078e-07 7.547886e-08 1.138137e-07
#> [5,9)    1.980202e-07 1.599396e-06 2.972491e-07 1.352721e-07 1.101069e-07
#> [10,14)  1.146078e-07 2.972491e-07 1.733225e-06 3.334453e-07 6.026086e-08
#> [15,19)  7.547886e-08 1.352721e-07 3.334453e-07 1.836025e-06 3.278825e-07
#> [20,24)  1.138137e-07 1.101069e-07 6.026086e-08 3.278825e-07 6.244090e-07
#> [25,29)  1.819153e-07 1.626085e-07 8.744544e-08 1.819264e-07 3.222430e-07
#> [30,34)  1.940880e-07 2.332740e-07 1.297928e-07 1.120304e-07 1.882853e-07
#> [35,39)  2.235471e-07 2.784775e-07 2.396054e-07 1.996738e-07 1.624326e-07
#> [40,44)  9.944949e-08 2.332186e-07 2.679243e-07 2.341623e-07 1.838660e-07
#> [45,49)  6.175863e-08 8.096090e-08 1.519569e-07 2.467186e-07 2.148012e-07
#> [50,54)  8.214138e-08 7.752704e-08 8.335625e-08 1.392241e-07 1.530723e-07
#> [55,59)  6.104769e-08 5.013907e-08 8.513107e-08 7.833919e-08 1.178128e-07
#> [60,64)  6.057159e-08 9.418311e-08 5.929099e-08 5.863294e-08 8.866898e-08
#> [65,69)  4.767006e-08 7.402091e-08 6.898452e-08 9.552015e-08 7.226646e-08
#> [70,74)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [75,79)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [80,100) 2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> CHW      1.205649e-07 1.512986e-07 1.380629e-07 1.563385e-07 1.788978e-07
#> CHR      2.100618e-09 2.893729e-09 5.503570e-09 9.539437e-09 5.468746e-09
#>               [25,29)      [30,34)      [35,39)      [40,44)      [45,49)
#> [0,4)    1.819153e-07 1.940880e-07 2.235471e-07 9.944949e-08 6.175863e-08
#> [5,9)    1.626085e-07 2.332740e-07 2.784775e-07 2.332186e-07 8.096090e-08
#> [10,14)  8.744544e-08 1.297928e-07 2.396054e-07 2.679243e-07 1.519569e-07
#> [15,19)  1.819264e-07 1.120304e-07 1.996738e-07 2.341623e-07 2.467186e-07
#> [20,24)  3.222430e-07 1.882853e-07 1.624326e-07 1.838660e-07 2.148012e-07
#> [25,29)  4.054957e-07 2.389792e-07 1.870791e-07 1.918110e-07 2.005309e-07
#> [30,34)  2.389792e-07 3.706066e-07 2.728896e-07 2.145623e-07 1.654374e-07
#> [35,39)  1.870791e-07 2.728896e-07 3.412451e-07 3.032663e-07 2.149544e-07
#> [40,44)  1.918110e-07 2.145623e-07 3.032663e-07 3.370632e-07 3.085384e-07
#> [45,49)  2.005309e-07 1.654374e-07 2.149544e-07 3.085384e-07 4.254147e-07
#> [50,54)  1.882440e-07 1.671228e-07 1.606864e-07 1.873283e-07 1.702589e-07
#> [55,59)  1.560297e-07 1.454919e-07 1.379231e-07 1.451614e-07 1.482183e-07
#> [60,64)  1.189756e-07 1.374885e-07 1.668294e-07 1.791329e-07 1.414642e-07
#> [65,69)  1.147535e-07 8.979643e-08 1.540697e-07 1.348771e-07 8.620090e-08
#> [70,74)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [75,79)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [80,100) 4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> CHW      2.108932e-07 2.140723e-07 2.231092e-07 2.333580e-07 2.218521e-07
#> CHR      4.760237e-09 3.516410e-09 4.653274e-09 1.307023e-08 1.059467e-08
#>               [50,54)      [55,59)      [60,64)      [65,69)      [70,74)
#> [0,4)    8.214138e-08 6.104769e-08 6.057159e-08 4.767006e-08 2.100618e-08
#> [5,9)    7.752704e-08 5.013907e-08 9.418311e-08 7.402091e-08 2.893729e-08
#> [10,14)  8.335625e-08 8.513107e-08 5.929099e-08 6.898452e-08 5.503570e-08
#> [15,19)  1.392241e-07 7.833919e-08 5.863294e-08 9.552015e-08 9.539437e-08
#> [20,24)  1.530723e-07 1.178128e-07 8.866898e-08 7.226646e-08 5.468746e-08
#> [25,29)  1.882440e-07 1.560297e-07 1.189756e-07 1.147535e-07 4.760237e-08
#> [30,34)  1.671228e-07 1.454919e-07 1.374885e-07 8.979643e-08 3.516410e-08
#> [35,39)  1.606864e-07 1.379231e-07 1.668294e-07 1.540697e-07 4.653274e-08
#> [40,44)  1.873283e-07 1.451614e-07 1.791329e-07 1.348771e-07 1.307023e-07
#> [45,49)  1.702589e-07 1.482183e-07 1.414642e-07 8.620090e-08 1.059467e-07
#> [50,54)  1.592838e-07 2.060829e-07 1.178586e-07 9.108480e-08 9.232320e-08
#> [55,59)  2.060829e-07 2.647960e-07 1.998500e-07 1.379069e-07 8.128181e-08
#> [60,64)  1.178586e-07 1.998500e-07 1.734975e-07 1.674043e-07 1.041277e-07
#> [65,69)  9.108480e-08 1.379069e-07 1.674043e-07 2.089258e-07 1.332611e-07
#> [70,74)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [75,79)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [80,100) 9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> CHW      1.696082e-07 1.754442e-07 1.543871e-07 1.220117e-07 8.046011e-08
#> CHR      9.232320e-09 8.128181e-09 1.041277e-08 1.332611e-08 1.493131e-08
#>               [75,79)     [80,100)          CHW          CHR
#> [0,4)    2.100618e-08 2.100618e-08 1.205649e-07 2.100618e-09
#> [5,9)    2.893729e-08 2.893729e-08 1.512986e-07 2.893729e-09
#> [10,14)  5.503570e-08 5.503570e-08 1.380629e-07 5.503570e-09
#> [15,19)  9.539437e-08 9.539437e-08 1.563385e-07 9.539437e-09
#> [20,24)  5.468746e-08 5.468746e-08 1.788978e-07 5.468746e-09
#> [25,29)  4.760237e-08 4.760237e-08 2.108932e-07 4.760237e-09
#> [30,34)  3.516410e-08 3.516410e-08 2.140723e-07 3.516410e-09
#> [35,39)  4.653274e-08 4.653274e-08 2.231092e-07 4.653274e-09
#> [40,44)  1.307023e-07 1.307023e-07 2.333580e-07 1.307023e-08
#> [45,49)  1.059467e-07 1.059467e-07 2.218521e-07 1.059467e-08
#> [50,54)  9.232320e-08 9.232320e-08 1.696082e-07 9.232320e-09
#> [55,59)  8.128181e-08 8.128181e-08 1.754442e-07 8.128181e-09
#> [60,64)  1.041277e-07 1.041277e-07 1.543871e-07 1.041277e-08
#> [65,69)  1.332611e-07 1.332611e-07 1.220117e-07 1.332611e-08
#> [70,74)  1.493131e-07 1.493131e-07 8.046011e-08 1.493131e-08
#> [75,79)  1.493131e-07 1.493131e-07 8.046011e-08 1.493131e-08
#> [80,100) 1.493131e-07 1.493131e-07 8.046011e-08 1.493131e-08
#> CHW      8.046011e-08 8.046011e-08 4.000000e-06 4.000000e-06
#> CHR      1.493131e-08 1.493131e-08 4.000000e-06 5.000000e-05
#> 
#> $N_tot
#>  [1] 3857263 4149852 3953866 3656968 4153080 4464394 4447277 4345812 3969684
#> [10] 4352267 4611160 4356053 3705330 3348257 3298925 2265470 3063468  398841
#> [19]  398841
#> 
#> $N_tot_15_64
#> [1] 42062025
#> 
#> $sero_specificity
#> [1] 0.9
#> 
#> $sero_sensitivity
#> [1] 0.99
#> 
#> $pillar2_specificity
#> [1] 0.99
#> 
#> $pillar2_sensitivity
#> [1] 0.99
#> 
#> $react_specificity
#> [1] 0.99
#> 
#> $react_sensitivity
#> [1] 0.99
#> 
#> $p_NC
#> [1] 0.01
#> 
#> $observation
#> $observation$phi_ICU
#> [1] 1
#> 
#> $observation$kappa_ICU
#> [1] 2
#> 
#> $observation$phi_general
#> [1] 1
#> 
#> $observation$kappa_general
#> [1] 2
#> 
#> $observation$phi_hosp
#> [1] 1
#> 
#> $observation$kappa_hosp
#> [1] 2
#> 
#> $observation$phi_death_hosp
#> [1] 1
#> 
#> $observation$kappa_death_hosp
#> [1] 2
#> 
#> $observation$phi_death_comm
#> [1] 1
#> 
#> $observation$kappa_death_comm
#> [1] 2
#> 
#> $observation$kappa_death
#> [1] 2
#> 
#> $observation$phi_admitted
#> [1] 1
#> 
#> $observation$kappa_admitted
#> [1] 2
#> 
#> $observation$phi_new
#> [1] 1
#> 
#> $observation$kappa_new
#> [1] 2
#> 
#> $observation$phi_new_admitted
#> [1] 1
#> 
#> $observation$kappa_new_admitted
#> [1] 2
#> 
#> $observation$phi_pillar2_cases
#> [1] 1
#> 
#> $observation$kappa_pillar2_cases
#> [1] 2
#> 
#> $observation$rho_pillar2_tests
#> [1] 0.1
#> 
#> $observation$exp_noise
#> [1] 1e+06
#> 
#> 
#> $n_groups
#> [1] 19
#> 
#> $p_star
#>  [1] 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2
#> 
#> $p_C
#>  [1] 0.1809000 0.1809000 0.1809000 0.1809000 0.2241000 0.2241000 0.2241000
#>  [8] 0.2241000 0.3054000 0.3054000 0.3054000 0.3054000 0.3546000 0.3546000
#> [15] 0.3546000 0.3546000 0.6456000 0.2810625 0.6456000
#> 
#> $p_G_D
#>  [1] 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.7
#> 
#> $p_H_D
#>  [1] 0.1258933 0.1226134 0.1356729 0.1526679 0.1743031 0.1941879 0.2093617
#>  [8] 0.2244326 0.2370135 0.2578281 0.2908746 0.3207640 0.3625638 0.3909655
#> [15] 0.4211515 0.4475459 0.4820000 0.2621283 0.4820000
#> 
#> $p_ICU_D
#>  [1] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#>  [8] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> [15] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> 
#> $p_W_D
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05
#> [16] 0.05 0.05 0.05 0.05
#> 
#> $p_ICU
#>  [1] 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387
#>  [7] 0.11845387 0.12551064 0.14091042 0.17595537 0.24437611 0.35243489
#> [13] 0.52779221 0.72404472 0.18197685 0.05439050 0.01673267 0.22548592
#> [19] 0.01673267
#> 
#> $p_sero_pos
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> 
#> $p_H
#>  [1] 0.000632689 0.000539148 0.000995640 0.002035807 0.004545371 0.008748134
#>  [7] 0.013802147 0.019850725 0.024607803 0.032822485 0.049084367 0.061571007
#> [13] 0.086379194 0.099452536 0.124208801 0.150169192 0.153030303 0.037108233
#> [19] 0.153030303
#> 
#> $psi_H
#>  [1] 0.004134403 0.003523145 0.006506162 0.013303293 0.029702424 0.057166024
#>  [7] 0.090192248 0.129717609 0.160803465 0.214483565 0.320749329 0.402345194
#> [13] 0.564458100 0.649887859 0.811661472 0.981303631 1.000000000 0.242489442
#> [19] 1.000000000
#> 
#> $p_H_step
#> [1] 0.1530303
#> 
#> $psi_ICU
#>  [1] 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021
#>  [7] 0.16360021 0.17334653 0.19461564 0.24301727 0.33751521 0.48675845
#> [13] 0.72894974 1.00000000 0.25133371 0.07512036 0.02311000 0.31142541
#> [19] 0.02311000
#> 
#> $p_ICU_step
#> [1] 0.7240447
#> 
#> $psi_ICU_D
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_ICU_D_step
#> [1] 0.5234896
#> 
#> $psi_H_D
#>  [1] 0.2611893 0.2543846 0.2814790 0.3167383 0.3616246 0.4028795 0.4343604
#>  [8] 0.4656277 0.4917293 0.5349130 0.6034743 0.6654854 0.7522070 0.8111317
#> [15] 0.8737583 0.9285184 1.0000000 0.5438346 1.0000000
#> 
#> $p_H_D_step
#> [1] 0.482
#> 
#> $psi_W_D
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_W_D_step
#> [1] 0.05
#> 
#> $psi_G_D
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> $p_G_D_step
#> [1] 0.7
#> 
#> $psi_star
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_star_step
#> [1] 0.2
#> 
#> $k_E
#> [1] 2
#> 
#> $k_A
#> [1] 1
#> 
#> $k_C
#> [1] 1
#> 
#> $k_G_D
#> [1] 2
#> 
#> $k_H_D
#> [1] 2
#> 
#> $k_H_R
#> [1] 2
#> 
#> $k_ICU_D
#> [1] 2
#> 
#> $k_ICU_W_R
#> [1] 2
#> 
#> $k_ICU_W_D
#> [1] 2
#> 
#> $k_ICU_pre
#> [1] 2
#> 
#> $k_W_R
#> [1] 2
#> 
#> $k_W_D
#> [1] 2
#> 
#> $k_sero_pos
#> [1] 2
#> 
#> $k_PCR_pre
#> [1] 2
#> 
#> $k_PCR_pos
#> [1] 2
#> 
#> $gamma_E
#> [1] 0.4357298
#> 
#> $gamma_A
#> [1] 0.4784689
#> 
#> $gamma_C
#> [1] 0.25
#> 
#> $gamma_G_D
#> [1] 0.4
#> 
#> $gamma_H_D
#> [1] 0.4
#> 
#> $gamma_H_R
#> [1] 0.2
#> 
#> $gamma_ICU_D
#> [1] 0.4
#> 
#> $gamma_ICU_W_R
#> [1] 0.2
#> 
#> $gamma_ICU_W_D
#> [1] 0.2
#> 
#> $gamma_ICU_pre
#> [1] 2
#> 
#> $gamma_W_R
#> [1] 0.4
#> 
#> $gamma_W_D
#> [1] 0.4
#> 
#> $gamma_sero_pre_1
#> [1] 0.2
#> 
#> $gamma_sero_pre_2
#> [1] 0.1
#> 
#> $gamma_sero_pos
#> [1] 0.04
#> 
#> $gamma_U
#> [1] 0.3
#> 
#> $gamma_PCR_pre
#> [1] 0.6666667
#> 
#> $gamma_PCR_pos
#> [1] 0.2
#> 
#> $n_strains
#> [1] 1
#> 
#> $strain_transmission
#> [1] 1
#> 
#> $strain_seed_step
#> [1] 0
#> 
#> $rel_susceptibility
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $rel_p_sympt
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $rel_p_hosp_if_sympt
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $rel_infectivity
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $vaccine_progression_rate_base
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    0
#> [11,]    0
#> [12,]    0
#> [13,]    0
#> [14,]    0
#> [15,]    0
#> [16,]    0
#> [17,]    0
#> [18,]    0
#> [19,]    0
#> 
#> $vaccine_population_reluctant
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $vaccine_daily_doses
#> [1] 0
#> 
#> $waning_rate
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $model_pcr_and_serology_user
#> [1] 1
#> 

# example set up of vaccination parameters independent of age
# 3 groups: 1) unvaccinated, 2) vaccinated with partial immunity
# 3) fully vaccinated (but with an imperfect vaccine). People return
# to group 1 after a period of time in group 3 to mirror waning immunity

# Assumption: immediately after vaccination susceptibility is reduced by
# 20%, and then by 50% when you reach full effect of the vaccination,
# then susceptibility returns to 100% upon waning of vaccine=induced
# immunity
# effect of vaccination similar across all age groups
rel_susceptibility <- c(1, 0.8, 0.5)

# The vaccine also reduces the risk of symptoms
rel_p_sympt <- c(1, 0.6, 0.3)
# and the risk of hospitalisation for those with symptoms
rel_p_hosp_if_sympt <- c(1, 0.95, 0.95)

# The vaccine also reduces infectivity of infected individuals by half
rel_infectivity <- c(1, 0.5, 0.5)

# Vaccination occurs at a constant rate of 0.03 per day,
# (i.e. average time to vaccination is 33 days)
# similar across all age groups.
# the period of build-up of immunity following vaccination is exponentially
# distributed and lasts on average two weeks;
# vaccine-induced immunity wanes after a period which is exponentially
# distributed and lasts on average 26 weeks (half a year);
# they are similar across all age groups
vaccine_progression_rate <- c(0, 1/(2*7), 1/(26*7))

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), "uk",
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_daily_doses = 10000)

# vaccination parameters are automatically copied across all age groups
p$rel_susceptibility

#>       [,1] [,2] [,3]
#>  [1,]    1  0.8  0.5
#>  [2,]    1  0.8  0.5
#>  [3,]    1  0.8  0.5
#>  [4,]    1  0.8  0.5
#>  [5,]    1  0.8  0.5
#>  [6,]    1  0.8  0.5
#>  [7,]    1  0.8  0.5
#>  [8,]    1  0.8  0.5
#>  [9,]    1  0.8  0.5
#> [10,]    1  0.8  0.5
#> [11,]    1  0.8  0.5
#> [12,]    1  0.8  0.5
#> [13,]    1  0.8  0.5
#> [14,]    1  0.8  0.5
#> [15,]    1  0.8  0.5
#> [16,]    1  0.8  0.5
#> [17,]    1  0.8  0.5
#> [18,]    1  0.8  0.5
#> [19,]    1  0.8  0.5

p$rel_p_sympt

#>       [,1] [,2] [,3]
#>  [1,]    1  0.6  0.3
#>  [2,]    1  0.6  0.3
#>  [3,]    1  0.6  0.3
#>  [4,]    1  0.6  0.3
#>  [5,]    1  0.6  0.3
#>  [6,]    1  0.6  0.3
#>  [7,]    1  0.6  0.3
#>  [8,]    1  0.6  0.3
#>  [9,]    1  0.6  0.3
#> [10,]    1  0.6  0.3
#> [11,]    1  0.6  0.3
#> [12,]    1  0.6  0.3
#> [13,]    1  0.6  0.3
#> [14,]    1  0.6  0.3
#> [15,]    1  0.6  0.3
#> [16,]    1  0.6  0.3
#> [17,]    1  0.6  0.3
#> [18,]    1  0.6  0.3
#> [19,]    1  0.6  0.3

p$rel_p_hosp_if_sympt

#>       [,1] [,2] [,3]
#>  [1,]    1 0.95 0.95
#>  [2,]    1 0.95 0.95
#>  [3,]    1 0.95 0.95
#>  [4,]    1 0.95 0.95
#>  [5,]    1 0.95 0.95
#>  [6,]    1 0.95 0.95
#>  [7,]    1 0.95 0.95
#>  [8,]    1 0.95 0.95
#>  [9,]    1 0.95 0.95
#> [10,]    1 0.95 0.95
#> [11,]    1 0.95 0.95
#> [12,]    1 0.95 0.95
#> [13,]    1 0.95 0.95
#> [14,]    1 0.95 0.95
#> [15,]    1 0.95 0.95
#> [16,]    1 0.95 0.95
#> [17,]    1 0.95 0.95
#> [18,]    1 0.95 0.95
#> [19,]    1 0.95 0.95

p$rel_infectivity

#>       [,1] [,2] [,3]
#>  [1,]    1  0.5  0.5
#>  [2,]    1  0.5  0.5
#>  [3,]    1  0.5  0.5
#>  [4,]    1  0.5  0.5
#>  [5,]    1  0.5  0.5
#>  [6,]    1  0.5  0.5
#>  [7,]    1  0.5  0.5
#>  [8,]    1  0.5  0.5
#>  [9,]    1  0.5  0.5
#> [10,]    1  0.5  0.5
#> [11,]    1  0.5  0.5
#> [12,]    1  0.5  0.5
#> [13,]    1  0.5  0.5
#> [14,]    1  0.5  0.5
#> [15,]    1  0.5  0.5
#> [16,]    1  0.5  0.5
#> [17,]    1  0.5  0.5
#> [18,]    1  0.5  0.5
#> [19,]    1  0.5  0.5

# Note that this is only the "base" rate as we fill in the first
# column dynamically based on vaccine_daily_doses
p$vaccine_progression_rate

#>       [,1]       [,2]        [,3]
#>  [1,]    0 0.07142857 0.005494505
#>  [2,]    0 0.07142857 0.005494505
#>  [3,]    0 0.07142857 0.005494505
#>  [4,]    0 0.07142857 0.005494505
#>  [5,]    0 0.07142857 0.005494505
#>  [6,]    0 0.07142857 0.005494505
#>  [7,]    0 0.07142857 0.005494505
#>  [8,]    0 0.07142857 0.005494505
#>  [9,]    0 0.07142857 0.005494505
#> [10,]    0 0.07142857 0.005494505
#> [11,]    0 0.07142857 0.005494505
#> [12,]    0 0.07142857 0.005494505
#> [13,]    0 0.07142857 0.005494505
#> [14,]    0 0.07142857 0.005494505
#> [15,]    0 0.07142857 0.005494505
#> [16,]    0 0.07142857 0.005494505
#> [17,]    0 0.07142857 0.005494505
#> [18,]    0 0.07142857 0.005494505
#> [19,]    0 0.07142857 0.005494505

### same example as above BUT assume a different effect of vaccine in the
### first age group
n_groups <- 19

# Assumption: vaccine is twice more effective at reducing susceptibility
# in the first age group
rel_susceptibility_agegp1 <- c(1, 0.4, 0.25)
rel_susceptibility_other_agegp <- c(1, 0.8, 0.5)
rel_susceptibility <- matrix(NA, nrow = n_groups, ncol = 3)
rel_susceptibility[1, ] <- rel_susceptibility_agegp1
for (i in seq(2, n_groups)) {
  rel_susceptibility[i, ] <- rel_susceptibility_other_agegp
}
rel_susceptibility

#>       [,1] [,2] [,3]
#>  [1,]    1  0.4 0.25
#>  [2,]    1  0.8 0.50
#>  [3,]    1  0.8 0.50
#>  [4,]    1  0.8 0.50
#>  [5,]    1  0.8 0.50
#>  [6,]    1  0.8 0.50
#>  [7,]    1  0.8 0.50
#>  [8,]    1  0.8 0.50
#>  [9,]    1  0.8 0.50
#> [10,]    1  0.8 0.50
#> [11,]    1  0.8 0.50
#> [12,]    1  0.8 0.50
#> [13,]    1  0.8 0.50
#> [14,]    1  0.8 0.50
#> [15,]    1  0.8 0.50
#> [16,]    1  0.8 0.50
#> [17,]    1  0.8 0.50
#> [18,]    1  0.8 0.50
#> [19,]    1  0.8 0.50

# But vaccine has the same impact on probability of symptoms and
# hospitalisation for the symptomatic across all age groups
rel_p_sympt <- matrix(rep(rel_p_sympt, n_groups), nrow = n_groups,
  byrow = TRUE)
rel_p_hosp_if_sympt <-
  matrix(rep(rel_p_hosp_if_sympt, n_groups), nrow = n_groups, byrow = TRUE)

# And vaccine has the same impact on onwards infectivity across age groups
rel_infectivity <- matrix(rep(rel_infectivity, n_groups), nrow = n_groups,
  byrow = TRUE)

# the period of build-up of immunity is the same for all age groups,
# lasting on average 2 weeks,
# but the first age group loses immunity more quickly
# (on average after 3 months) than the other age groups
# (on average after 6 months)
vaccine_progression_rate <- cbind(0,
                                  rep(1 / (2 * 7), n_groups),
                                  c(1 / (13 * 7),
                                  rep( 1 / (26 * 7), n_groups - 1)))

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), "uk",
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_daily_doses = 10000)