Parameters for the carehomes model

Parameters for the "carehomes" model.

carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  p_death_carehome = 0.7,
  sero_specificity = 0.9,
  sero_sensitivity = 0.99,
  progression = NULL,
  observation = NULL,
  initial_I = 10,
  eps = 0.1,
  m_CHW = 4e-06,
  m_CHR = 5e-05,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99,
  p_NC = 0.01,
  strain_transmission = 1,
  strain_seed_date = NULL,
  strain_seed_rate = NULL,
  strain_rel_gamma_A = 1,
  strain_rel_gamma_P = 1,
  strain_rel_gamma_C_1 = 1,
  strain_rel_gamma_C_2 = 1,
  strain_rel_severity = 1,
  rel_susceptibility = 1,
  rel_p_sympt = 1,
  rel_p_hosp_if_sympt = 1,
  rel_infectivity = 1,
  vaccine_progression_rate = NULL,
  vaccine_schedule = NULL,
  vaccine_index_dose2 = NULL,
  vaccine_catchup_fraction = 1,
  waning_rate = 0,
  exp_noise = 1e+06,
  cross_immunity = 1
)

Arguments

start_date
Thee start date, as a sircovid_date() (i.e., the number of days into 2020)

region
The region to run the model for. This will bee used to get population data, which is currently fixed within the package and is limited to "uk", the four constituent nations ("england", "wales", "scotland", "northern_ireland") and the 7 NHS regions (e.g., "midlands"). These names are case insensitive.

beta_date
A vector of dates for changes in beta (the contact rate parameter), or NULL if a single value is used for all times (see sircovid_parameters_beta(), where this is passed as date).

beta_value
A vector of values for beta (the contact rate parameter). If not given, and if beta_date is NULL then a value of 0.08 will be used through the whole simulation, otherwise if beta_date is NULL this must be a scalar. If beta_date is given then beta_date and beta_value must have the same length (see sircovid_parameters_beta(), where this is passed as value).

severity
Severity data, via Bob Verity's markovid package. This needs to be NULL (use the default bundled data version in the package), a data.frame object (for raw severity data) or a list (for data that has already been processed by sircovid for use). New severity data comes from Bob Verity via the markovid package, and needs to be carefully calibrated with the progression parameters.

p_death_carehome
Probability of death within carehomes (conditional on having an "inflenza-like-illness")

sero_specificity
Specificity of the serology test

sero_sensitivity
Sensitivity of the serology test

progression
Progression data

observation
Either NULL or a list of observation parameters. If NULL, then a list of observation parameters will be generated using carehomes_parameters_observation(exp_noise)

initial_I
Initial number of infected indidviduals; these will enter the model as asymptomatic 15-19 year olds at start_date. The default is 10 individuals.

eps
Change in contact rate for carehome residents

m_CHW
Contact rate between carehome workers and either residents or workers

m_CHR
Contact rate between carehome residents

pillar2_specificity
Specificity of the Pillar 2 test

pillar2_sensitivity
Sensitivity of the Pillar 2 test

react_specificity
Specificity of the REACT test

react_sensitivity
Sensitivity of the REACT test

p_NC
Proportion of population who do not have covid but have covid-like symptoms

strain_transmission
Vector of length two for relative transmissibility of each strain modelled. First element should be 1. Length will define the number of strains used in the model, either 1 or 2.

strain_seed_date
Either NULL (no seeding) or a vector of sircovid_date values corresponding to the dates that strain_seed_rate should change. For example, to seed a constant rate from a given date, provide one value for each of strain_seed_date and strain_seed_rate; or to seed with a constant rate for a set period, provide two dates and two rates with the second rate equal to 0.

strain_seed_rate
Either NULL (no seeding) or a vector of values representing the daily rate of seeding, starting on the dates set in strain_seed_date. Seeding is drawn from Poisson(strain_seed_rate * dt) at each day and so the rate is spread evenly across the steps that occur within each date. For example, to seed a constant rate from a given date, provide one value for each of strain_seed_date and strain_seed_rate; or to seed with a constant rate for a set period, provide two dates and two rates with the second rate equal to 0.

strain_rel_gamma_A
Vector of relative rates of progression out of I_A (gamma_A) for each strain modelled. If 1 all strains have same rates. Otherwise vector of same length as strain_transmission, with entries that determines the relative scaling of the defaults for each strain.

strain_rel_gamma_P
Vector of relative rates of progression out of I_P (gamma_P) for each strain modelled. If 1 all strains have same rates. Otherwise vector of same length as strain_transmission, with entries that determines the relative scaling of the defaults for each strain.

strain_rel_gamma_C_1
Vector of relative rates of progression out of I_C_1 (gamma_C_1) for each strain modelled. If 1 all strains have same rates. Otherwise vector of same length as strain_transmission, with entries that determines the relative scaling of the defaults for each strain.

strain_rel_gamma_C_2
Vector of relative rates of progression out of I_C_2 (gamma_C_2) for each strain modelled. If 1 all strains have same rates. Otherwise vector of same length as strain_transmission, with entries that determines the relative scaling of the defaults for each strain.

strain_rel_severity
Vector of relative probabilities of death for each strain modelled. If 1 all strains have same probabilities of death. Otherwise vector of same length as strain_transmission, where the first value should be 1 (for the first strain) and subsequent values between 0 and 1. To ensure valid probabilities, severity is upper-truncated at 1 after scaling.

rel_susceptibility
A vector or array of values representing the relative susceptibility of individuals in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case relative susceptibility will be the same across all age groups within one vaccination category, and will be the same for all pathogen strains. Specifying an array instead of a vector allows different relative susceptibilities by age (first dimension of the array), pathogen strain (second dimension) and vaccination group (third dimension); in that case, the first layer (3rd dimension) of rel_susceptibility should be 1 (for the non-vaccinated group) and other values be between 0 and 1

rel_p_sympt
A vector or matrix of values of same dimension as rel_susceptibility representing the relative probability of symptomatic infection in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case the relative reduction in probability of symptomatic infection will be the same across all age groups and all strains within one vaccination category. Specifying an array instead of a vector allows different relative reductions in probability of symptomatic infection by age (first dimension of the array), pathogen strain (second dimension) and vaccination group (third dimension); in that case, the first layer of rel_p_sympt should be 1 (for the non-vaccinated group) and other values be between 0 and 1

rel_p_hosp_if_sympt
A vector or array of values of same dimension as rel_susceptibility representing the relative probability of hospitalisation for symptomatic cases in different vaccination groups. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case the relative reduction in probability of hospitalisation for symptomatic cases will be the same across all age groups and all strains within one vaccination category. Specifying an array instead of a vector allows different relative reductions in probability of hospitalisation for symptomatic cases by age (first dimension of the array), pathogen strain (second dimension) and vaccination group (third dimension); in that case, the first layer of rel_p_hosp_if_sympt should be 1 (for the non-vaccinated group) and other values be between 0 and 1

rel_infectivity
A vector or array of values representing the relative infectivity of individuals in different vaccination groups, if they are infected. If a vector, the first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1. In that case relative infectivity will be the same across all age groups and all strains within one vaccination category. Specifying an array instead of a vector allows different relative infectivities by age (first dimension of the array), pathogen strain (second dimension) and vaccination group (third dimension); in that case, the first layer of rel_infectivity should be 1 (for the non-vaccinated group) and other values be between 0 and 1

vaccine_progression_rate
A vector or matrix of values of same dimension as rel_susceptibility representing the rate of movement between different vaccination classes. If a vector, it should have as many values as vaccination classes, and the same rates of progression will be used for all age groups (the first rate is the vaccination rate and the last of the rates is the rate of returning to the initial vacination class); if a matrix, the element on row i and column j is the rate of progression from the jth vaccination class to the (j+1)th for age group i.

vaccine_schedule
A vaccine_schedule object indicating the people to be vaccinated by group over time

vaccine_index_dose2
The index to use for the second dose

vaccine_catchup_fraction
A value between 0 and 1 indicating the proportion of doses not distributed according to schedule (e.g. because too many people were in the I or H compartments and could not be vaccinated at the scheduled time) that we postpone to a later date. A value of 0 means we do not catch up at all on any missed doses; a value of 1 means we try to catch up for all missed doses. This is set to 1 by default

waning_rate
A single value or a vector of values representing the rates of waning of immunity after infection; if a single value the same rate is used for all age groups; if a vector of values if used it should have one value per age group.

exp_noise
Rate of exponential noise used in the compare function - typically set to a large value so that noise is small but non-zero. If set to Inf then there is no noise in the observation process (not realistic but useful for testing).

cross_immunity
A value or vector of same length as strain_transmission that controls the amount of immunity conferred by previous infection with one strain. If a scalar is given then same level of cross immunity is assumed between both strains. Otherwise a vector of length two should be provided where the first value is the relative protection against infection with strain 2 following infection with strain 1 (i.e. while in the R1 compartment), and vice versa for the second value. Values between 0 and 1 are allowed with values of 1 (default) indicating complete cross-immunity, and values of 0 mean no cross-immunity. Modelling 'superinfections' (being exposed to one strain after recovering from another) can be turned off by setting cross_immunity = 1.

Value

A list of inputs to the model, many of which are fixed and represent data. These correspond largely to user() calls within the odin code, though some are also used in processing just before the model is run.

Examples

region <- "london"
carehomes_parameters(sircovid_date("2020-02-01"), region)

#> $hosp_transmission
#> [1] 0
#> 
#> $ICU_transmission
#> [1] 0
#> 
#> $G_D_transmission
#> [1] 0
#> 
#> $dt
#> [1] 0.25
#> 
#> $steps_per_day
#> [1] 4
#> 
#> $initial_step
#> [1] 128
#> 
#> $n_age_groups
#> [1] 17
#> 
#> $beta_step
#> [1] 0.08
#> 
#> $population
#>  [1] 606004 602992 537605 467602 565427 770010 838109 781520 661032 596614
#> [11] 567017 495468 391074 312926 273546 193310 301733
#> 
#> $carehome_beds
#> [1] 35564
#> 
#> $carehome_residents
#> [1] 26388
#> 
#> $carehome_workers
#> [1] 26388
#> 
#> $m
#>                 [0,4)        [5,9)      [10,14)      [15,19)      [20,24)
#> [0,4)    3.161348e-06 1.319430e-06 7.900780e-07 5.453716e-07 7.778341e-07
#> [5,9)    1.319430e-06 1.100720e-05 2.124195e-06 9.749905e-07 7.747056e-07
#> [10,14)  7.900780e-07 2.124195e-06 1.274717e-05 2.517955e-06 4.429459e-07
#> [15,19)  5.453716e-07 9.749905e-07 2.517955e-06 1.435898e-05 2.510819e-06
#> [20,24)  7.778341e-07 7.747056e-07 4.429459e-07 2.510819e-06 4.586304e-06
#> [25,29)  1.113966e-06 1.044914e-06 5.702188e-07 1.261678e-06 2.155378e-06
#> [30,34)  1.127793e-06 1.425261e-06 8.180012e-07 7.409853e-07 1.189016e-06
#> [35,39)  1.336883e-06 1.776266e-06 1.513314e-06 1.290362e-06 1.043805e-06
#> [40,44)  6.136411e-07 1.505346e-06 1.790263e-06 1.627237e-06 1.235879e-06
#> [45,49)  4.270248e-07 5.875836e-07 1.119126e-06 1.890765e-06 1.581396e-06
#> [50,54)  5.877758e-07 5.884340e-07 6.503774e-07 1.106381e-06 1.165694e-06
#> [55,59)  4.761266e-07 3.920319e-07 6.705895e-07 6.532076e-07 9.337941e-07
#> [60,64)  4.989705e-07 7.404991e-07 4.918079e-07 5.150288e-07 7.073104e-07
#> [65,69)  3.652815e-07 5.715578e-07 5.668887e-07 8.344055e-07 6.254129e-07
#> [70,74)  1.743744e-07 2.386333e-07 4.546425e-07 7.687184e-07 4.592928e-07
#> [75,79)  1.743744e-07 2.386333e-07 4.546425e-07 7.687184e-07 4.592928e-07
#> [80,100) 1.743744e-07 2.386333e-07 4.546425e-07 7.687184e-07 4.592928e-07
#> CHW      7.727726e-07 1.007542e-06 9.529621e-07 1.135706e-06 1.251534e-06
#> CHR      1.743744e-08 2.386333e-08 4.546425e-08 7.687184e-08 4.592928e-08
#>               [25,29)      [30,34)      [35,39)      [40,44)      [45,49)
#> [0,4)    1.113966e-06 1.127793e-06 1.336883e-06 6.136411e-07 4.270248e-07
#> [5,9)    1.044914e-06 1.425261e-06 1.776266e-06 1.505346e-06 5.875836e-07
#> [10,14)  5.702188e-07 8.180012e-07 1.513314e-06 1.790263e-06 1.119126e-06
#> [15,19)  1.261678e-06 7.409853e-07 1.290362e-06 1.627237e-06 1.890765e-06
#> [20,24)  2.155378e-06 1.189016e-06 1.043805e-06 1.235879e-06 1.581396e-06
#> [25,29)  2.377253e-06 1.335506e-06 1.071064e-06 1.148422e-06 1.336137e-06
#> [30,34)  1.335506e-06 1.988604e-06 1.509395e-06 1.223924e-06 1.017249e-06
#> [35,39)  1.071064e-06 1.509395e-06 1.919337e-06 1.775442e-06 1.368391e-06
#> [40,44)  1.148422e-06 1.223924e-06 1.775442e-06 2.049581e-06 2.071088e-06
#> [45,49)  1.336137e-06 1.017249e-06 1.368391e-06 2.071088e-06 3.138926e-06
#> [50,54)  1.317929e-06 1.080470e-06 1.074831e-06 1.301969e-06 1.341626e-06
#> [55,59)  1.095398e-06 1.054161e-06 9.717732e-07 1.057665e-06 1.181094e-06
#> [60,64)  8.662949e-07 9.633676e-07 1.142620e-06 1.308573e-06 1.205297e-06
#> [65,69)  8.330567e-07 5.886681e-07 1.095430e-06 9.771187e-07 6.999890e-07
#> [70,74)  3.341726e-07 2.415837e-07 4.018786e-07 9.512518e-07 9.525721e-07
#> [75,79)  3.341726e-07 2.415837e-07 4.018786e-07 9.512518e-07 9.525721e-07
#> [80,100) 3.341726e-07 2.415837e-07 4.018786e-07 9.512518e-07 9.525721e-07
#> CHW      1.318500e-06 1.271584e-06 1.354107e-06 1.492083e-06 1.582476e-06
#> CHR      3.341726e-08 2.415837e-08 4.018786e-08 9.512518e-08 9.525721e-08
#>               [50,54)      [55,59)      [60,64)      [65,69)      [70,74)
#> [0,4)    5.877758e-07 4.761266e-07 4.989705e-07 3.652815e-07 1.743744e-07
#> [5,9)    5.884340e-07 3.920319e-07 7.404991e-07 5.715578e-07 2.386333e-07
#> [10,14)  6.503774e-07 6.705895e-07 4.918079e-07 5.668887e-07 4.546425e-07
#> [15,19)  1.106381e-06 6.532076e-07 5.150288e-07 8.344055e-07 7.687184e-07
#> [20,24)  1.165694e-06 9.337941e-07 7.073104e-07 6.254129e-07 4.592928e-07
#> [25,29)  1.317929e-06 1.095398e-06 8.662949e-07 8.330567e-07 3.341726e-07
#> [30,34)  1.080470e-06 1.054161e-06 9.633676e-07 5.886681e-07 2.415837e-07
#> [35,39)  1.074831e-06 9.717732e-07 1.142620e-06 1.095430e-06 4.018786e-07
#> [40,44)  1.301969e-06 1.057665e-06 1.308573e-06 9.771187e-07 9.512518e-07
#> [45,49)  1.341626e-06 1.181094e-06 1.205297e-06 6.999890e-07 9.525721e-07
#> [50,54)  1.309351e-06 1.756463e-06 1.044103e-06 8.288251e-07 8.619066e-07
#> [55,59)  1.756463e-06 2.354676e-06 1.850658e-06 1.303483e-06 8.419123e-07
#> [60,64)  1.044103e-06 1.850658e-06 1.665964e-06 1.670145e-06 1.055041e-06
#> [65,69)  8.288251e-07 1.303483e-06 1.670145e-06 2.248786e-06 1.473635e-06
#> [70,74)  8.619066e-07 8.419123e-07 1.055041e-06 1.473635e-06 1.749736e-06
#> [75,79)  8.619066e-07 8.419123e-07 1.055041e-06 1.473635e-06 1.749736e-06
#> [80,100) 8.619066e-07 8.419123e-07 1.055041e-06 1.473635e-06 1.749736e-06
#> CHW      1.278343e-06 1.415236e-06 1.255860e-06 9.995894e-07 7.050399e-07
#> CHR      8.619066e-08 8.419123e-08 1.055041e-07 1.473635e-07 1.749736e-07
#>               [75,79)     [80,100)          CHW          CHR
#> [0,4)    1.743744e-07 1.743744e-07 7.727726e-07 1.743744e-08
#> [5,9)    2.386333e-07 2.386333e-07 1.007542e-06 2.386333e-08
#> [10,14)  4.546425e-07 4.546425e-07 9.529621e-07 4.546425e-08
#> [15,19)  7.687184e-07 7.687184e-07 1.135706e-06 7.687184e-08
#> [20,24)  4.592928e-07 4.592928e-07 1.251534e-06 4.592928e-08
#> [25,29)  3.341726e-07 3.341726e-07 1.318500e-06 3.341726e-08
#> [30,34)  2.415837e-07 2.415837e-07 1.271584e-06 2.415837e-08
#> [35,39)  4.018786e-07 4.018786e-07 1.354107e-06 4.018786e-08
#> [40,44)  9.512518e-07 9.512518e-07 1.492083e-06 9.512518e-08
#> [45,49)  9.525721e-07 9.525721e-07 1.582476e-06 9.525721e-08
#> [50,54)  8.619066e-07 8.619066e-07 1.278343e-06 8.619066e-08
#> [55,59)  8.419123e-07 8.419123e-07 1.415236e-06 8.419123e-08
#> [60,64)  1.055041e-06 1.055041e-06 1.255860e-06 1.055041e-07
#> [65,69)  1.473635e-06 1.473635e-06 9.995894e-07 1.473635e-07
#> [70,74)  1.749736e-06 1.749736e-06 7.050399e-07 1.749736e-07
#> [75,79)  1.749736e-06 1.749736e-06 7.050399e-07 1.749736e-07
#> [80,100) 1.749736e-06 1.749736e-06 7.050399e-07 1.749736e-07
#> CHW      7.050399e-07 7.050399e-07 4.000000e-06 4.000000e-06
#> CHR      1.749736e-07 1.749736e-07 4.000000e-06 5.000000e-05
#> 
#> $N_tot
#>  [1] 606004 602992 537605 467602 565427 766712 834810 778222 657734 593316
#> [11] 563718 492170 387776 311607 272227 189352 281942  26388  26388
#> 
#> $initial_I
#> [1] 10
#> 
#> $cross_immunity
#> [1] 1
#> 
#> $N_tot_15_64
#> [1] 6107487
#> 
#> $N_tot_all
#> [1] 8961992
#> 
#> $N_tot_over25
#> [1] 6182362
#> 
#> $N_tot_react
#> [1] 8329600
#> 
#> $sero_specificity
#> [1] 0.9
#> 
#> $sero_sensitivity
#> [1] 0.99
#> 
#> $pillar2_specificity
#> [1] 0.99
#> 
#> $pillar2_sensitivity
#> [1] 0.99
#> 
#> $react_specificity
#> [1] 0.99
#> 
#> $react_sensitivity
#> [1] 0.99
#> 
#> $p_NC
#> [1] 0.01
#> 
#> $I_A_transmission
#> [1] 0.363
#> 
#> $I_P_transmission
#> [1] 1
#> 
#> $I_C_1_transmission
#> [1] 1
#> 
#> $I_C_2_transmission
#> [1] 0
#> 
#> $n_groups
#> [1] 19
#> 
#> $p_star
#>  [1] 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2
#> 
#> $p_C
#>  [1] 0.1809000 0.1809000 0.1809000 0.1809000 0.2241000 0.2241000 0.2241000
#>  [8] 0.2241000 0.3054000 0.3054000 0.3054000 0.3054000 0.3546000 0.3546000
#> [15] 0.3546000 0.3546000 0.6456000 0.2810625 0.6456000
#> 
#> $p_G_D
#>       [,1]
#>  [1,]  0.0
#>  [2,]  0.0
#>  [3,]  0.0
#>  [4,]  0.0
#>  [5,]  0.0
#>  [6,]  0.0
#>  [7,]  0.0
#>  [8,]  0.0
#>  [9,]  0.0
#> [10,]  0.0
#> [11,]  0.0
#> [12,]  0.0
#> [13,]  0.0
#> [14,]  0.0
#> [15,]  0.0
#> [16,]  0.0
#> [17,]  0.0
#> [18,]  0.0
#> [19,]  0.7
#> 
#> $p_H_D
#>            [,1]
#>  [1,] 0.1258933
#>  [2,] 0.1226134
#>  [3,] 0.1356729
#>  [4,] 0.1526679
#>  [5,] 0.1743031
#>  [6,] 0.1941879
#>  [7,] 0.2093617
#>  [8,] 0.2244326
#>  [9,] 0.2370135
#> [10,] 0.2578281
#> [11,] 0.2908746
#> [12,] 0.3207640
#> [13,] 0.3625638
#> [14,] 0.3909655
#> [15,] 0.4211515
#> [16,] 0.4475459
#> [17,] 0.4820000
#> [18,] 0.2621283
#> [19,] 0.4820000
#> 
#> $p_ICU_D
#>            [,1]
#>  [1,] 0.5234896
#>  [2,] 0.5234896
#>  [3,] 0.5234896
#>  [4,] 0.5234896
#>  [5,] 0.5234896
#>  [6,] 0.5234896
#>  [7,] 0.5234896
#>  [8,] 0.5234896
#>  [9,] 0.5234896
#> [10,] 0.5234896
#> [11,] 0.5234896
#> [12,] 0.5234896
#> [13,] 0.5234896
#> [14,] 0.5234896
#> [15,] 0.5234896
#> [16,] 0.5234896
#> [17,] 0.5234896
#> [18,] 0.5234896
#> [19,] 0.5234896
#> 
#> $p_W_D
#>       [,1]
#>  [1,] 0.05
#>  [2,] 0.05
#>  [3,] 0.05
#>  [4,] 0.05
#>  [5,] 0.05
#>  [6,] 0.05
#>  [7,] 0.05
#>  [8,] 0.05
#>  [9,] 0.05
#> [10,] 0.05
#> [11,] 0.05
#> [12,] 0.05
#> [13,] 0.05
#> [14,] 0.05
#> [15,] 0.05
#> [16,] 0.05
#> [17,] 0.05
#> [18,] 0.05
#> [19,] 0.05
#> 
#> $p_ICU
#>  [1] 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387
#>  [7] 0.11845387 0.12551064 0.14091042 0.17595537 0.24437611 0.35243489
#> [13] 0.52779221 0.72404472 0.18197685 0.05439050 0.01673267 0.22548592
#> [19] 0.01673267
#> 
#> $p_sero_pos
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> 
#> $p_H
#>  [1] 0.000632689 0.000539148 0.000995640 0.002035807 0.004545371 0.008748134
#>  [7] 0.013802147 0.019850725 0.024607803 0.032822485 0.049084367 0.061571007
#> [13] 0.086379194 0.099452536 0.124208801 0.150169192 0.153030303 0.037108233
#> [19] 0.153030303
#> 
#> $psi_H
#>  [1] 0.004134403 0.003523145 0.006506162 0.013303293 0.029702424 0.057166024
#>  [7] 0.090192248 0.129717609 0.160803465 0.214483565 0.320749329 0.402345194
#> [13] 0.564458100 0.649887859 0.811661472 0.981303631 1.000000000 0.242489442
#> [19] 1.000000000
#> 
#> $p_H_step
#> [1] 0.1530303
#> 
#> $psi_ICU
#>  [1] 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021
#>  [7] 0.16360021 0.17334653 0.19461564 0.24301727 0.33751521 0.48675845
#> [13] 0.72894974 1.00000000 0.25133371 0.07512036 0.02311000 0.31142541
#> [19] 0.02311000
#> 
#> $p_ICU_step
#> [1] 0.7240447
#> 
#> $psi_ICU_D
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $p_ICU_D_step
#>           [,1]
#> [1,] 0.5234896
#> 
#> $psi_H_D
#>            [,1]
#>  [1,] 0.2611893
#>  [2,] 0.2543846
#>  [3,] 0.2814790
#>  [4,] 0.3167383
#>  [5,] 0.3616246
#>  [6,] 0.4028795
#>  [7,] 0.4343604
#>  [8,] 0.4656277
#>  [9,] 0.4917293
#> [10,] 0.5349130
#> [11,] 0.6034743
#> [12,] 0.6654854
#> [13,] 0.7522070
#> [14,] 0.8111317
#> [15,] 0.8737583
#> [16,] 0.9285184
#> [17,] 1.0000000
#> [18,] 0.5438346
#> [19,] 1.0000000
#> 
#> $p_H_D_step
#>       [,1]
#> [1,] 0.482
#> 
#> $psi_W_D
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> $p_W_D_step
#>      [,1]
#> [1,] 0.05
#> 
#> $psi_G_D
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    0
#> [11,]    0
#> [12,]    0
#> [13,]    0
#> [14,]    0
#> [15,]    0
#> [16,]    0
#> [17,]    0
#> [18,]    0
#> [19,]    1
#> 
#> $p_G_D_step
#>      [,1]
#> [1,]  0.7
#> 
#> $psi_star
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_star_step
#> [1] 0.2
#> 
#> $k_E
#> [1] 2
#> 
#> $k_A
#> [1] 1
#> 
#> $k_P
#> [1] 1
#> 
#> $k_C_1
#> [1] 1
#> 
#> $k_C_2
#> [1] 1
#> 
#> $k_G_D
#> [1] 2
#> 
#> $k_H_D
#> [1] 2
#> 
#> $k_H_R
#> [1] 2
#> 
#> $k_ICU_D
#> [1] 2
#> 
#> $k_ICU_W_R
#> [1] 2
#> 
#> $k_ICU_W_D
#> [1] 2
#> 
#> $k_ICU_pre
#> [1] 2
#> 
#> $k_W_R
#> [1] 2
#> 
#> $k_W_D
#> [1] 2
#> 
#> $k_sero_pos
#> [1] 2
#> 
#> $k_PCR_pre
#> [1] 2
#> 
#> $k_PCR_pos
#> [1] 2
#> 
#> $gamma_E
#> [1] 0.5847953
#> 
#> $gamma_A
#> [1] 0.3472222
#> 
#> $gamma_P
#> [1] 0.5952381
#> 
#> $gamma_C_1
#> [1] 0.4672897
#> 
#> $gamma_C_2
#> [1] 0.5376344
#> 
#> $gamma_G_D
#> [1] 0.6666667
#> 
#> $gamma_H_D
#> [1] 0.4
#> 
#> $gamma_H_R
#> [1] 0.2
#> 
#> $gamma_ICU_D
#> [1] 0.4
#> 
#> $gamma_ICU_W_R
#> [1] 0.2
#> 
#> $gamma_ICU_W_D
#> [1] 0.2
#> 
#> $gamma_ICU_pre
#> [1] 2
#> 
#> $gamma_W_R
#> [1] 0.4
#> 
#> $gamma_W_D
#> [1] 0.4
#> 
#> $gamma_sero_pre_1
#> [1] 0.2
#> 
#> $gamma_sero_pre_2
#> [1] 0.1
#> 
#> $gamma_sero_pos
#> [1] 0.04
#> 
#> $gamma_U
#> [1] 0.3
#> 
#> $gamma_PCR_pre
#> [1] 0.6666667
#> 
#> $gamma_PCR_pos
#> [1] 0.2
#> 
#> $gamma_H_R_step
#> [1] 0.2
#> 
#> $gamma_W_R_step
#> [1] 0.4
#> 
#> $gamma_ICU_W_R_step
#> [1] 0.2
#> 
#> $gamma_H_D_step
#> [1] 0.4
#> 
#> $gamma_W_D_step
#> [1] 0.4
#> 
#> $gamma_ICU_W_D_step
#> [1] 0.2
#> 
#> $gamma_ICU_D_step
#> [1] 0.4
#> 
#> $gamma_ICU_pre_step
#> [1] 2
#> 
#> $n_strains
#> [1] 1
#> 
#> $strain_transmission
#> [1] 1
#> 
#> $strain_seed_step
#> [1] 0
#> 
#> $rel_susceptibility
#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> 
#> $rel_p_sympt
#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> 
#> $rel_p_hosp_if_sympt
#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> 
#> $rel_infectivity
#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> 
#> $vaccine_dose_step
#> , , 1
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> 
#> 
#> $index_dose
#> [1] 1 1
#> 
#> $n_vacc_classes
#> [1] 1
#> 
#> $vaccine_progression_rate_base
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    0
#> [11,]    0
#> [12,]    0
#> [13,]    0
#> [14,]    0
#> [15,]    0
#> [16,]    0
#> [17,]    0
#> [18,]    0
#> [19,]    0
#> 
#> $vaccine_catchup_fraction
#> [1] 1
#> 
#> $waning_rate
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $phi_ICU
#> [1] 1
#> 
#> $kappa_ICU
#> [1] 2
#> 
#> $phi_general
#> [1] 1
#> 
#> $kappa_general
#> [1] 2
#> 
#> $phi_hosp
#> [1] 1
#> 
#> $kappa_hosp
#> [1] 2
#> 
#> $phi_death_hosp
#> [1] 1
#> 
#> $kappa_death_hosp
#> [1] 2
#> 
#> $phi_death_carehomes
#> [1] 1
#> 
#> $kappa_death_carehomes
#> [1] 2
#> 
#> $phi_death_comm
#> [1] 1
#> 
#> $kappa_death_comm
#> [1] 2
#> 
#> $kappa_death_non_hosp
#> [1] 2
#> 
#> $kappa_death
#> [1] 2
#> 
#> $phi_admitted
#> [1] 1
#> 
#> $kappa_admitted
#> [1] 2
#> 
#> $phi_diagnoses
#> [1] 1
#> 
#> $kappa_diagnoses
#> [1] 2
#> 
#> $phi_all_admission
#> [1] 1
#> 
#> $kappa_all_admission
#> [1] 2
#> 
#> $phi_pillar2_cases
#> [1] 1
#> 
#> $kappa_pillar2_cases
#> [1] 2
#> 
#> $rho_pillar2_tests
#> [1] 0.1
#> 
#> $exp_noise
#> [1] 1e+06
#> 

# example set up of vaccination parameters independent of age
# 3 groups: 1) unvaccinated, 2) vaccinated with partial immunity
# 3) fully vaccinated (but with an imperfect vaccine). People return
# to group 1 after a period of time in group 3 to mirror waning immunity

# Assumption: immediately after vaccination susceptibility is reduced by
# 20%, and then by 50% when you reach full effect of the vaccination,
# then susceptibility returns to 100% upon waning of vaccine=induced
# immunity
# effect of vaccination similar across all age groups
rel_susceptibility <- c(1, 0.8, 0.5)

# The vaccine also reduces the risk of symptoms
rel_p_sympt <- c(1, 0.6, 0.3)
# and the risk of hospitalisation for those with symptoms
rel_p_hosp_if_sympt <- c(1, 0.95, 0.95)

# The vaccine also reduces infectivity of infected individuals by half
rel_infectivity <- c(1, 0.5, 0.5)

# On average 10000 first doses of vaccine are given every day
# Second doses are given on average 12 weeks after the first dose
# vaccine-induced immunity wanes after a period which is exponentially
# distributed and lasts on average 26 weeks (half a year);
# they are similar across all age groups
vaccine_progression_rate <- c(0, 0, 1/(26*7))

daily_doses <- rep(10000, 365)
mean_days_between_doses <- 12 * 7
n <- vaccine_priority_population(region, uptake = 1)
schedule <- vaccine_schedule_future(
  0, daily_doses, mean_days_between_doses, n)

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2)

# vaccination parameters are automatically copied across all age groups#
# (and across strains but here we only have 1 strain which is the 2nd
# dimension here)
p$rel_susceptibility

#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> , , 2
#> 
#>       [,1]
#>  [1,]  0.8
#>  [2,]  0.8
#>  [3,]  0.8
#>  [4,]  0.8
#>  [5,]  0.8
#>  [6,]  0.8
#>  [7,]  0.8
#>  [8,]  0.8
#>  [9,]  0.8
#> [10,]  0.8
#> [11,]  0.8
#> [12,]  0.8
#> [13,]  0.8
#> [14,]  0.8
#> [15,]  0.8
#> [16,]  0.8
#> [17,]  0.8
#> [18,]  0.8
#> [19,]  0.8
#> 
#> , , 3
#> 
#>       [,1]
#>  [1,]  0.5
#>  [2,]  0.5
#>  [3,]  0.5
#>  [4,]  0.5
#>  [5,]  0.5
#>  [6,]  0.5
#>  [7,]  0.5
#>  [8,]  0.5
#>  [9,]  0.5
#> [10,]  0.5
#> [11,]  0.5
#> [12,]  0.5
#> [13,]  0.5
#> [14,]  0.5
#> [15,]  0.5
#> [16,]  0.5
#> [17,]  0.5
#> [18,]  0.5
#> [19,]  0.5
#> 

p$rel_p_sympt

#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> , , 2
#> 
#>       [,1]
#>  [1,]  0.6
#>  [2,]  0.6
#>  [3,]  0.6
#>  [4,]  0.6
#>  [5,]  0.6
#>  [6,]  0.6
#>  [7,]  0.6
#>  [8,]  0.6
#>  [9,]  0.6
#> [10,]  0.6
#> [11,]  0.6
#> [12,]  0.6
#> [13,]  0.6
#> [14,]  0.6
#> [15,]  0.6
#> [16,]  0.6
#> [17,]  0.6
#> [18,]  0.6
#> [19,]  0.6
#> 
#> , , 3
#> 
#>       [,1]
#>  [1,]  0.3
#>  [2,]  0.3
#>  [3,]  0.3
#>  [4,]  0.3
#>  [5,]  0.3
#>  [6,]  0.3
#>  [7,]  0.3
#>  [8,]  0.3
#>  [9,]  0.3
#> [10,]  0.3
#> [11,]  0.3
#> [12,]  0.3
#> [13,]  0.3
#> [14,]  0.3
#> [15,]  0.3
#> [16,]  0.3
#> [17,]  0.3
#> [18,]  0.3
#> [19,]  0.3
#> 

p$rel_p_hosp_if_sympt

#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> , , 2
#> 
#>       [,1]
#>  [1,] 0.95
#>  [2,] 0.95
#>  [3,] 0.95
#>  [4,] 0.95
#>  [5,] 0.95
#>  [6,] 0.95
#>  [7,] 0.95
#>  [8,] 0.95
#>  [9,] 0.95
#> [10,] 0.95
#> [11,] 0.95
#> [12,] 0.95
#> [13,] 0.95
#> [14,] 0.95
#> [15,] 0.95
#> [16,] 0.95
#> [17,] 0.95
#> [18,] 0.95
#> [19,] 0.95
#> 
#> , , 3
#> 
#>       [,1]
#>  [1,] 0.95
#>  [2,] 0.95
#>  [3,] 0.95
#>  [4,] 0.95
#>  [5,] 0.95
#>  [6,] 0.95
#>  [7,] 0.95
#>  [8,] 0.95
#>  [9,] 0.95
#> [10,] 0.95
#> [11,] 0.95
#> [12,] 0.95
#> [13,] 0.95
#> [14,] 0.95
#> [15,] 0.95
#> [16,] 0.95
#> [17,] 0.95
#> [18,] 0.95
#> [19,] 0.95
#> 

p$rel_infectivity

#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> , , 2
#> 
#>       [,1]
#>  [1,]  0.5
#>  [2,]  0.5
#>  [3,]  0.5
#>  [4,]  0.5
#>  [5,]  0.5
#>  [6,]  0.5
#>  [7,]  0.5
#>  [8,]  0.5
#>  [9,]  0.5
#> [10,]  0.5
#> [11,]  0.5
#> [12,]  0.5
#> [13,]  0.5
#> [14,]  0.5
#> [15,]  0.5
#> [16,]  0.5
#> [17,]  0.5
#> [18,]  0.5
#> [19,]  0.5
#> 
#> , , 3
#> 
#>       [,1]
#>  [1,]  0.5
#>  [2,]  0.5
#>  [3,]  0.5
#>  [4,]  0.5
#>  [5,]  0.5
#>  [6,]  0.5
#>  [7,]  0.5
#>  [8,]  0.5
#>  [9,]  0.5
#> [10,]  0.5
#> [11,]  0.5
#> [12,]  0.5
#> [13,]  0.5
#> [14,]  0.5
#> [15,]  0.5
#> [16,]  0.5
#> [17,]  0.5
#> [18,]  0.5
#> [19,]  0.5
#> 

# Note that this is only the "base" rate as we fill in the first
# column dynamically based on vaccine_daily_doses
p$vaccine_progression_rate

#>       [,1] [,2]        [,3]
#>  [1,]    0    0 0.005494505
#>  [2,]    0    0 0.005494505
#>  [3,]    0    0 0.005494505
#>  [4,]    0    0 0.005494505
#>  [5,]    0    0 0.005494505
#>  [6,]    0    0 0.005494505
#>  [7,]    0    0 0.005494505
#>  [8,]    0    0 0.005494505
#>  [9,]    0    0 0.005494505
#> [10,]    0    0 0.005494505
#> [11,]    0    0 0.005494505
#> [12,]    0    0 0.005494505
#> [13,]    0    0 0.005494505
#> [14,]    0    0 0.005494505
#> [15,]    0    0 0.005494505
#> [16,]    0    0 0.005494505
#> [17,]    0    0 0.005494505
#> [18,]    0    0 0.005494505
#> [19,]    0    0 0.005494505

### same example as above BUT assume a different effect of vaccine in the
### first age group
n_groups <- 19
n_strains <- 1

# Assumption: vaccine is twice more effective at reducing susceptibility
# in the first age group
rel_susceptibility_agegp1 <- c(1, 0.4, 0.25)
rel_susceptibility_other_agegp <- c(1, 0.8, 0.5)
rel_susceptibility <- array(NA, dim = c(n_groups, n_strains, 3))
rel_susceptibility[1, , ] <- rel_susceptibility_agegp1
for (i in seq(2, n_groups)) {
  rel_susceptibility[i, , ] <- rel_susceptibility_other_agegp
}
rel_susceptibility

#> , , 1
#> 
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> 
#> , , 2
#> 
#>       [,1]
#>  [1,]  0.4
#>  [2,]  0.8
#>  [3,]  0.8
#>  [4,]  0.8
#>  [5,]  0.8
#>  [6,]  0.8
#>  [7,]  0.8
#>  [8,]  0.8
#>  [9,]  0.8
#> [10,]  0.8
#> [11,]  0.8
#> [12,]  0.8
#> [13,]  0.8
#> [14,]  0.8
#> [15,]  0.8
#> [16,]  0.8
#> [17,]  0.8
#> [18,]  0.8
#> [19,]  0.8
#> 
#> , , 3
#> 
#>       [,1]
#>  [1,] 0.25
#>  [2,] 0.50
#>  [3,] 0.50
#>  [4,] 0.50
#>  [5,] 0.50
#>  [6,] 0.50
#>  [7,] 0.50
#>  [8,] 0.50
#>  [9,] 0.50
#> [10,] 0.50
#> [11,] 0.50
#> [12,] 0.50
#> [13,] 0.50
#> [14,] 0.50
#> [15,] 0.50
#> [16,] 0.50
#> [17,] 0.50
#> [18,] 0.50
#> [19,] 0.50
#> 

# But vaccine has the same impact on probability of symptoms and
# hospitalisation for the symptomatic across all age groups
rel_p_sympt <- array(rep(rel_p_sympt, each = n_groups),
  dim = c(n_groups, n_strains, 3))
rel_p_hosp_if_sympt <-
  array(rep(rel_p_hosp_if_sympt, each = n_groups),
  dim = c(n_groups, n_strains, 3))

# And vaccine has the same impact on onwards infectivity across age groups
rel_infectivity <- array(rep(rel_infectivity, each = n_groups),
  dim = c(n_groups, n_strains, 3))

# the period of build-up of immunity is the same for all age groups,
# lasting on average 2 weeks,
# but the first age group loses immunity more quickly
# (on average after 3 months) than the other age groups
# (on average after 6 months)
vaccine_progression_rate <- cbind(0, 0,
                                  c(1 / (13 * 7),
                                  rep( 1 / (26 * 7), n_groups - 1)))

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2)

# TODO: add an example of manually set up vaccine schedule