Parameters for the carehomes model

Parameters for the "carehomes" model.

carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  p_death_carehome = 0.7,
  sero_specificity = 0.9,
  sero_sensitivity = 0.99,
  progression = NULL,
  eps = 0.1,
  C_1 = 4e-06,
  C_2 = 5e-05,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99,
  prop_noncovid_sympt = 0.01,
  rel_susceptibility = 1,
  waning_rate = 0,
  exp_noise = 1e+06
)

Arguments

start_date
Thee start date, as a sircovid_date() (i.e., the number of days into 2020)

region
The region to run the model for. This will bee used to get population data, which is currently fixed within the package and is limited to "uk", the four constituent nations ("england", "wales", "scotland", "northern_ireland") and thee 7 NHS regions (e.g., "midlands"). These names are case insensitive.

beta_date
A vector of dates for changes in beta (the contact rate parameter), or NULL if a single value is used for all times (see sircovid_parameters_beta(), where this is passed as date).

beta_value
A vector of values for beta (the contact rate parameter). If not given, and if beta_date is null then a value of 0.08 will be used through the whole simulation, otherwise if beta_date is NULL this must be a scalar. If beta_date is given then beta_date and beta_value must have the same length (see sircovid_parameters_beta(), where this is passed as value).

severity
Severity data, via Bob Verity's markovid package. This needs to be NULL (use the default bundled data version in the package), a data.frame object (for raw severity data) or a list (for data that has already been processed by sircovid for use). New severity data comes from Bob Verity via the markovid package, and needs to be carefully calibrated with the progression parameters.

p_death_carehome
Probability of death within carehomes (conditional on having an "inflenza-like-illness")

sero_specificity
Specificity of the serology test

sero_sensitivity
Sensitivity of the serology test

progression
Progression data

eps
Change in contact rate for carehome residents

C_1
Contact rate between carehome workers and either residents or workers

C_2
Contact rate between carehome residents

pillar2_specificity
Specificity of the Pillar 2 test

pillar2_sensitivity
Sensitivity of the Pillar 2 test

react_specificity
Specificity of the REACT test

react_sensitivity
Sensitivity of the REACT test

prop_noncovid_sympt
Proportion of population who do not have covid but have covid-like symptoms

rel_susceptibility
A vector of values representing the relative susceptibility of individuals in different vaccination groups. The first value should be 1 (for the non-vaccinated group) and subsequent values be between 0 and 1.

waning_rate
A single value or a vector of values representing the rates of waning of immunity after infection; if a single value the same rate is used for all age groups; if a vector of values if used it should have one value per age group.

exp_noise
Rate of exponential noise used in the compare function - typically set to a large value so that noise is small but non-zero. If set to Inf then there is no noise in the observation process (not realistic but useful for testing).

Value

A list of inputs to the model, many of which are fixed and represent data. These correspond largely to user() calls within the odin code, though some are also used in processing just before the model is run.

Examples

carehomes_parameters(sircovid_date("2020-02-01"), "uk")

#> $hosp_transmission
#> [1] 0.1
#> 
#> $ICU_transmission
#> [1] 0.05
#> 
#> $comm_D_transmission
#> [1] 0.05
#> 
#> $dt
#> [1] 0.25
#> 
#> $initial_step
#> [1] 128
#> 
#> $n_age_groups
#> [1] 17
#> 
#> $beta_step
#> [1] 0.08
#> 
#> $population
#>  [1] 3857263 4149852 3953866 3656968 4153080 4514249 4497132 4395667 4019539
#> [10] 4402122 4661015 4405908 3755185 3368199 3318867 2325296 3362599
#> 
#> $carehome_beds
#> [1] 537521
#> 
#> $carehome_residents
#> [1] 398841
#> 
#> $carehome_workers
#> [1] 398841
#> 
#> $m
#>                 [0,4)        [5,9)      [10,14)      [15,19)      [20,24)
#> [0,4)    4.966707e-07 1.980202e-07 1.146078e-07 7.547886e-08 1.138137e-07
#> [5,9)    1.980202e-07 1.599396e-06 2.972491e-07 1.352721e-07 1.101069e-07
#> [10,14)  1.146078e-07 2.972491e-07 1.733225e-06 3.334453e-07 6.026086e-08
#> [15,19)  7.547886e-08 1.352721e-07 3.334453e-07 1.836025e-06 3.278825e-07
#> [20,24)  1.138137e-07 1.101069e-07 6.026086e-08 3.278825e-07 6.244090e-07
#> [25,29)  1.819153e-07 1.626085e-07 8.744544e-08 1.819264e-07 3.222430e-07
#> [30,34)  1.940880e-07 2.332740e-07 1.297928e-07 1.120304e-07 1.882853e-07
#> [35,39)  2.235471e-07 2.784775e-07 2.396054e-07 1.996738e-07 1.624326e-07
#> [40,44)  9.944949e-08 2.332186e-07 2.679243e-07 2.341623e-07 1.838660e-07
#> [45,49)  6.175863e-08 8.096090e-08 1.519569e-07 2.467186e-07 2.148012e-07
#> [50,54)  8.214138e-08 7.752704e-08 8.335625e-08 1.392241e-07 1.530723e-07
#> [55,59)  6.104769e-08 5.013907e-08 8.513107e-08 7.833919e-08 1.178128e-07
#> [60,64)  6.057159e-08 9.418311e-08 5.929099e-08 5.863294e-08 8.866898e-08
#> [65,69)  4.767006e-08 7.402091e-08 6.898452e-08 9.552015e-08 7.226646e-08
#> [70,74)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [75,79)  2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> [80,100) 2.100618e-08 2.893729e-08 5.503570e-08 9.539437e-08 5.468746e-08
#> CHW      1.220055e-07 1.511359e-07 1.374799e-07 1.571202e-07 1.808205e-07
#> CHR      2.100618e-09 2.893729e-09 5.503570e-09 9.539437e-09 5.468746e-09
#>               [25,29)      [30,34)      [35,39)      [40,44)      [45,49)
#> [0,4)    1.819153e-07 1.940880e-07 2.235471e-07 9.944949e-08 6.175863e-08
#> [5,9)    1.626085e-07 2.332740e-07 2.784775e-07 2.332186e-07 8.096090e-08
#> [10,14)  8.744544e-08 1.297928e-07 2.396054e-07 2.679243e-07 1.519569e-07
#> [15,19)  1.819264e-07 1.120304e-07 1.996738e-07 2.341623e-07 2.467186e-07
#> [20,24)  3.222430e-07 1.882853e-07 1.624326e-07 1.838660e-07 2.148012e-07
#> [25,29)  4.054957e-07 2.389792e-07 1.870791e-07 1.918110e-07 2.005309e-07
#> [30,34)  2.389792e-07 3.706066e-07 2.728896e-07 2.145623e-07 1.654374e-07
#> [35,39)  1.870791e-07 2.728896e-07 3.412451e-07 3.032663e-07 2.149544e-07
#> [40,44)  1.918110e-07 2.145623e-07 3.032663e-07 3.370632e-07 3.085384e-07
#> [45,49)  2.005309e-07 1.654374e-07 2.149544e-07 3.085384e-07 4.254147e-07
#> [50,54)  1.882440e-07 1.671228e-07 1.606864e-07 1.873283e-07 1.702589e-07
#> [55,59)  1.560297e-07 1.454919e-07 1.379231e-07 1.451614e-07 1.482183e-07
#> [60,64)  1.189756e-07 1.374885e-07 1.668294e-07 1.791329e-07 1.414642e-07
#> [65,69)  1.147535e-07 8.979643e-08 1.540697e-07 1.348771e-07 8.620090e-08
#> [70,74)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [75,79)  4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> [80,100) 4.760237e-08 3.516410e-08 4.653274e-08 1.307023e-07 1.059467e-07
#> CHW      2.133559e-07 2.156373e-07 2.227971e-07 2.326728e-07 2.217800e-07
#> CHR      4.760237e-09 3.516410e-09 4.653274e-09 1.307023e-08 1.059467e-08
#>               [50,54)      [55,59)      [60,64)      [65,69)      [70,74)
#> [0,4)    8.214138e-08 6.104769e-08 6.057159e-08 4.767006e-08 2.100618e-08
#> [5,9)    7.752704e-08 5.013907e-08 9.418311e-08 7.402091e-08 2.893729e-08
#> [10,14)  8.335625e-08 8.513107e-08 5.929099e-08 6.898452e-08 5.503570e-08
#> [15,19)  1.392241e-07 7.833919e-08 5.863294e-08 9.552015e-08 9.539437e-08
#> [20,24)  1.530723e-07 1.178128e-07 8.866898e-08 7.226646e-08 5.468746e-08
#> [25,29)  1.882440e-07 1.560297e-07 1.189756e-07 1.147535e-07 4.760237e-08
#> [30,34)  1.671228e-07 1.454919e-07 1.374885e-07 8.979643e-08 3.516410e-08
#> [35,39)  1.606864e-07 1.379231e-07 1.668294e-07 1.540697e-07 4.653274e-08
#> [40,44)  1.873283e-07 1.451614e-07 1.791329e-07 1.348771e-07 1.307023e-07
#> [45,49)  1.702589e-07 1.482183e-07 1.414642e-07 8.620090e-08 1.059467e-07
#> [50,54)  1.592838e-07 2.060829e-07 1.178586e-07 9.108480e-08 9.232320e-08
#> [55,59)  2.060829e-07 2.647960e-07 1.998500e-07 1.379069e-07 8.128181e-08
#> [60,64)  1.178586e-07 1.998500e-07 1.734975e-07 1.674043e-07 1.041277e-07
#> [65,69)  9.108480e-08 1.379069e-07 1.674043e-07 2.089258e-07 1.332611e-07
#> [70,74)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [75,79)  9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> [80,100) 9.232320e-08 8.128181e-08 1.041277e-07 1.332611e-07 1.493131e-07
#> CHW      1.703607e-07 1.754236e-07 1.533259e-07 1.206750e-07 7.932803e-08
#> CHR      9.232320e-09 8.128181e-09 1.041277e-08 1.332611e-08 1.493131e-08
#>               [75,79)     [80,100)          CHW          CHR
#> [0,4)    2.100618e-08 2.100618e-08 1.220055e-07 2.100618e-09
#> [5,9)    2.893729e-08 2.893729e-08 1.511359e-07 2.893729e-09
#> [10,14)  5.503570e-08 5.503570e-08 1.374799e-07 5.503570e-09
#> [15,19)  9.539437e-08 9.539437e-08 1.571202e-07 9.539437e-09
#> [20,24)  5.468746e-08 5.468746e-08 1.808205e-07 5.468746e-09
#> [25,29)  4.760237e-08 4.760237e-08 2.133559e-07 4.760237e-09
#> [30,34)  3.516410e-08 3.516410e-08 2.156373e-07 3.516410e-09
#> [35,39)  4.653274e-08 4.653274e-08 2.227971e-07 4.653274e-09
#> [40,44)  1.307023e-07 1.307023e-07 2.326728e-07 1.307023e-08
#> [45,49)  1.059467e-07 1.059467e-07 2.217800e-07 1.059467e-08
#> [50,54)  9.232320e-08 9.232320e-08 1.703607e-07 9.232320e-09
#> [55,59)  8.128181e-08 8.128181e-08 1.754236e-07 8.128181e-09
#> [60,64)  1.041277e-07 1.041277e-07 1.533259e-07 1.041277e-08
#> [65,69)  1.332611e-07 1.332611e-07 1.206750e-07 1.332611e-08
#> [70,74)  1.493131e-07 1.493131e-07 7.932803e-08 1.493131e-08
#> [75,79)  1.493131e-07 1.493131e-07 7.932803e-08 1.493131e-08
#> [80,100) 1.493131e-07 1.493131e-07 7.932803e-08 1.493131e-08
#> CHW      7.932803e-08 7.932803e-08 4.000000e-06 4.000000e-06
#> CHR      1.493131e-08 1.493131e-08 4.000000e-06 5.000000e-05
#> 
#> $N_tot
#>  [1] 3857263 4149852 3953866 3656968 4153080 4462289 4445369 4345072 3973273
#> [10] 4351452 4607365 4355195 3711962 3348257 3298925 2265470 3063468  398841
#> [19]  398841
#> 
#> $N_tot_15_64
#> [1] 42062025
#> 
#> $sero_specificity
#> [1] 0.9
#> 
#> $sero_sensitivity
#> [1] 0.99
#> 
#> $pillar2_specificity
#> [1] 0.99
#> 
#> $pillar2_sensitivity
#> [1] 0.99
#> 
#> $react_specificity
#> [1] 0.99
#> 
#> $react_sensitivity
#> [1] 0.99
#> 
#> $prop_noncovid_sympt
#> [1] 0.01
#> 
#> $observation
#> $observation$phi_ICU
#> [1] 1
#> 
#> $observation$k_ICU
#> [1] 2
#> 
#> $observation$phi_general
#> [1] 1
#> 
#> $observation$k_general
#> [1] 2
#> 
#> $observation$phi_hosp
#> [1] 1
#> 
#> $observation$k_hosp
#> [1] 2
#> 
#> $observation$phi_death_hosp
#> [1] 1
#> 
#> $observation$k_death_hosp
#> [1] 2
#> 
#> $observation$phi_death_comm
#> [1] 1
#> 
#> $observation$k_death_comm
#> [1] 2
#> 
#> $observation$k_death
#> [1] 2
#> 
#> $observation$phi_admitted
#> [1] 1
#> 
#> $observation$k_admitted
#> [1] 2
#> 
#> $observation$phi_new
#> [1] 1
#> 
#> $observation$k_new
#> [1] 2
#> 
#> $observation$phi_new_admitted
#> [1] 1
#> 
#> $observation$k_new_admitted
#> [1] 2
#> 
#> $observation$phi_pillar2_cases
#> [1] 1
#> 
#> $observation$k_pillar2_cases
#> [1] 2
#> 
#> $observation$rho_pillar2_tests
#> [1] 0.1
#> 
#> $observation$exp_noise
#> [1] 1e+06
#> 
#> 
#> $n_groups
#> [1] 19
#> 
#> $p_admit_conf
#>  [1] 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2
#> 
#> $p_asympt
#>  [1] 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34 0.34
#> [16] 0.34 0.34 0.34 0.34
#> 
#> $p_death_comm
#>  [1] 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.7
#> 
#> $p_death_hosp_D
#>  [1] 0.1258933 0.1226134 0.1356729 0.1526679 0.1743031 0.1941879 0.2093617
#>  [8] 0.2244326 0.2370135 0.2578281 0.2908746 0.3207640 0.3625638 0.3909655
#> [15] 0.4211515 0.4475459 0.4820000 0.2603941 0.4820000
#> 
#> $p_death_ICU
#>  [1] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#>  [8] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> [15] 0.5234896 0.5234896 0.5234896 0.5234896 0.5234896
#> 
#> $p_hosp_ILI
#>  [1] 0.001772051 0.001510059 0.002681536 0.005482996 0.010801921 0.020789645
#>  [7] 0.030033884 0.043195770 0.050353126 0.067162222 0.084808192 0.106382665
#> [13] 0.131365666 0.151247633 0.169503156 0.204930342 0.200036782 0.065630479
#> [19] 0.200036782
#> 
#> $p_ICU_hosp
#>  [1] 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387 0.11845387
#>  [7] 0.11845387 0.12551064 0.14091042 0.17595537 0.24437611 0.35243489
#> [13] 0.52779221 0.72404472 0.18197685 0.05439050 0.01673267 0.22030943
#> [19] 0.01673267
#> 
#> $p_seroconversion
#>  [1] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#>  [8] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> [15] 0.8212394 0.8212394 0.8212394 0.8212394 0.8212394
#> 
#> $p_sympt_ILI
#>  [1] 0.2356449 0.2356449 0.2450545 0.2450545 0.2777233 0.2777233 0.3033047
#>  [8] 0.3033047 0.3225450 0.3225450 0.3819877 0.3819877 0.4339815 0.4339815
#> [15] 0.4836359 0.4836359 0.5049071 0.3393986 0.5049071
#> 
#> $psi_hosp_ILI
#>  [1] 0.008647088 0.007368644 0.013085111 0.026755414 0.052710207 0.101447374
#>  [7] 0.146556549 0.210782695 0.245708497 0.327731957 0.413839119 0.519116224
#> [13] 0.641025944 0.738044114 0.827125720 1.000000000 0.976120865 0.320257500
#> [19] 0.976120865
#> 
#> $p_hosp_ILI_step
#> [1] 0.2049303
#> 
#> $psi_ICU_hosp
#>  [1] 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021 0.16360021
#>  [7] 0.16360021 0.17334653 0.19461564 0.24301727 0.33751521 0.48675845
#> [13] 0.72894974 1.00000000 0.25133371 0.07512036 0.02311000 0.30427599
#> [19] 0.02311000
#> 
#> $p_ICU_hosp_step
#> [1] 0.7240447
#> 
#> $psi_death_ICU
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_death_ICU_step
#> [1] 0.5234896
#> 
#> $psi_death_hosp_D
#>  [1] 0.2611893 0.2543846 0.2814790 0.3167383 0.3616246 0.4028795 0.4343604
#>  [8] 0.4656277 0.4917293 0.5349130 0.6034743 0.6654854 0.7522070 0.8111317
#> [15] 0.8737583 0.9285184 1.0000000 0.5402367 1.0000000
#> 
#> $p_death_hosp_D_step
#> [1] 0.482
#> 
#> $psi_death_comm
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> $p_death_comm_step
#> [1] 0.7
#> 
#> $psi_admit_conf
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $p_admit_conf_step
#> [1] 0.2
#> 
#> $s_E
#> [1] 2
#> 
#> $s_asympt
#> [1] 1
#> 
#> $s_mild
#> [1] 1
#> 
#> $s_ILI
#> [1] 1
#> 
#> $s_comm_D
#> [1] 2
#> 
#> $s_hosp_D
#> [1] 2
#> 
#> $s_hosp_R
#> [1] 2
#> 
#> $s_ICU_D
#> [1] 2
#> 
#> $s_ICU_R
#> [1] 2
#> 
#> $s_triage
#> [1] 2
#> 
#> $s_stepdown
#> [1] 2
#> 
#> $s_R_pos
#> [1] 2
#> 
#> $s_PCR_pre
#> [1] 2
#> 
#> $s_PCR_pos
#> [1] 2
#> 
#> $gamma_E
#> [1] 0.4357298
#> 
#> $gamma_asympt
#> [1] 0.4784689
#> 
#> $gamma_mild
#> [1] 0.4784689
#> 
#> $gamma_ILI
#> [1] 0.25
#> 
#> $gamma_comm_D
#> [1] 0.4
#> 
#> $gamma_hosp_D
#> [1] 0.4
#> 
#> $gamma_hosp_R
#> [1] 0.2
#> 
#> $gamma_ICU_D
#> [1] 0.4
#> 
#> $gamma_ICU_R
#> [1] 0.2
#> 
#> $gamma_triage
#> [1] 2
#> 
#> $gamma_stepdown
#> [1] 0.4
#> 
#> $gamma_R_pre_1
#> [1] 0.2
#> 
#> $gamma_R_pre_2
#> [1] 0.1
#> 
#> $gamma_R_pos
#> [1] 0.04
#> 
#> $gamma_test
#> [1] 0.3
#> 
#> $gamma_PCR_pre
#> [1] 0.6666667
#> 
#> $gamma_PCR_pos
#> [1] 0.2
#> 
#> $rel_susceptibility
#> [1] 1
#> 
#> $waning_rate
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>