Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())

#> $run
#>                               time                                icu 
#>                                  1                                 12 
#>                            general               deaths_carehomes_inc 
#>                                 13                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 22                                 23 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 34                                 35 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 36                                 38 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 39                                 40 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 41                                 42 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 43                                 37 
#>                          react_pos 
#>                                 44 
#> 
#> $state
#>                                icu                            general 
#>                                 12                                 13 
#>                        deaths_comm                   deaths_carehomes 
#>                                 16                                 18 
#>                        deaths_hosp                           admitted 
#>                                 15                                  6 
#>                          diagnoses               deaths_carehomes_inc 
#>                                  7                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 22                                 23 
#>                        sympt_cases            sympt_cases_non_variant 
#>                                 24                                 25 
#>                 sympt_cases_over25     sympt_cases_non_variant_over25 
#>                                 26                                 27 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 34                                 35 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 36                                 38 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 39                                 40 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 41                                 42 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 43                                 37 
#>                          react_pos                               hosp 
#>                                 44                                 14 
#>                             deaths                         infections 
#>                                 21                                  4 
#>                     infections_inc                                S_0 
#>                                  5                                294 
#>                                S_5                               S_10 
#>                                295                                296 
#>                               S_15                               S_20 
#>                                297                                298 
#>                               S_25                               S_30 
#>                                299                                300 
#>                               S_35                               S_40 
#>                                301                                302 
#>                               S_45                               S_50 
#>                                303                                304 
#>                               S_55                               S_60 
#>                                305                                306 
#>                               S_65                               S_70 
#>                                307                                308 
#>                               S_75                               S_80 
#>                                309                                310 
#>                              S_CHW                              S_CHR 
#>                                311                                312 
#>                                R_0                                R_5 
#>                                446                                447 
#>                               R_10                               R_15 
#>                                448                                449 
#>                               R_20                               R_25 
#>                                450                                451 
#>                               R_30                               R_35 
#>                                452                                453 
#>                               R_40                               R_45 
#>                                454                                455 
#>                               R_50                               R_55 
#>                                456                                457 
#>                               R_60                               R_65 
#>                                458                                459 
#>                               R_70                               R_75 
#>                                460                                461 
#>                               R_80                              R_CHW 
#>                                462                                463 
#>                              R_CHR                        cum_admit_0 
#>                                464                                 84 
#>                        cum_admit_5                       cum_admit_10 
#>                                 85                                 86 
#>                       cum_admit_15                       cum_admit_20 
#>                                 87                                 88 
#>                       cum_admit_25                       cum_admit_30 
#>                                 89                                 90 
#>                       cum_admit_35                       cum_admit_40 
#>                                 91                                 92 
#>                       cum_admit_45                       cum_admit_50 
#>                                 93                                 94 
#>                       cum_admit_55                       cum_admit_60 
#>                                 95                                 96 
#>                       cum_admit_65                       cum_admit_70 
#>                                 97                                 98 
#>                       cum_admit_75                       cum_admit_80 
#>                                 99                                100 
#>                      cum_admit_CHW                      cum_admit_CHR 
#>                                101                                102 
#>                           D_hosp_0                           D_hosp_5 
#>                                 46                                 47 
#>                          D_hosp_10                          D_hosp_15 
#>                                 48                                 49 
#>                          D_hosp_20                          D_hosp_25 
#>                                 50                                 51 
#>                          D_hosp_30                          D_hosp_35 
#>                                 52                                 53 
#>                          D_hosp_40                          D_hosp_45 
#>                                 54                                 55 
#>                          D_hosp_50                          D_hosp_55 
#>                                 56                                 57 
#>                          D_hosp_60                          D_hosp_65 
#>                                 58                                 59 
#>                          D_hosp_70                          D_hosp_75 
#>                                 60                                 61 
#>                          D_hosp_80                         D_hosp_CHW 
#>                                 62                                 63 
#>                         D_hosp_CHR                            D_all_0 
#>                                 64                                275 
#>                            D_all_5                           D_all_10 
#>                                276                                277 
#>                           D_all_15                           D_all_20 
#>                                278                                279 
#>                           D_all_25                           D_all_30 
#>                                280                                281 
#>                           D_all_35                           D_all_40 
#>                                282                                283 
#>                           D_all_45                           D_all_50 
#>                                284                                285 
#>                           D_all_55                           D_all_60 
#>                                286                                287 
#>                           D_all_65                           D_all_70 
#>                                288                                289 
#>                           D_all_75                           D_all_80 
#>                                290                                291 
#>                          D_all_CHW                          D_all_CHR 
#>                                292                                293 
#>               diagnoses_admitted_0               diagnoses_admitted_5 
#>                                237                                238 
#>              diagnoses_admitted_10              diagnoses_admitted_15 
#>                                239                                240 
#>              diagnoses_admitted_20              diagnoses_admitted_25 
#>                                241                                242 
#>              diagnoses_admitted_30              diagnoses_admitted_35 
#>                                243                                244 
#>              diagnoses_admitted_40              diagnoses_admitted_45 
#>                                245                                246 
#>              diagnoses_admitted_50              diagnoses_admitted_55 
#>                                247                                248 
#>              diagnoses_admitted_60              diagnoses_admitted_65 
#>                                249                                250 
#>              diagnoses_admitted_70              diagnoses_admitted_75 
#>                                251                                252 
#>              diagnoses_admitted_80             diagnoses_admitted_CHW 
#>                                253                                254 
#>             diagnoses_admitted_CHR             cum_infections_disag_0 
#>                                255                                256 
#>             cum_infections_disag_5            cum_infections_disag_10 
#>                                257                                258 
#>            cum_infections_disag_15            cum_infections_disag_20 
#>                                259                                260 
#>            cum_infections_disag_25            cum_infections_disag_30 
#>                                261                                262 
#>            cum_infections_disag_35            cum_infections_disag_40 
#>                                263                                264 
#>            cum_infections_disag_45            cum_infections_disag_50 
#>                                265                                266 
#>            cum_infections_disag_55            cum_infections_disag_60 
#>                                267                                268 
#>            cum_infections_disag_65            cum_infections_disag_70 
#>                                269                                270 
#>            cum_infections_disag_75            cum_infections_disag_80 
#>                                271                                272 
#>           cum_infections_disag_CHW           cum_infections_disag_CHR 
#>                                273                                274 
#>                       I_weighted_0                       I_weighted_5 
#>                                484                                485 
#>                      I_weighted_10                      I_weighted_15 
#>                                486                                487 
#>                      I_weighted_20                      I_weighted_25 
#>                                488                                489 
#>                      I_weighted_30                      I_weighted_35 
#>                                490                                491 
#>                      I_weighted_40                      I_weighted_45 
#>                                492                                493 
#>                      I_weighted_50                      I_weighted_55 
#>                                494                                495 
#>                      I_weighted_60                      I_weighted_65 
#>                                496                                497 
#>                      I_weighted_70                      I_weighted_75 
#>                                498                                499 
#>                      I_weighted_80                     I_weighted_CHW 
#>                                500                                501 
#>                     I_weighted_CHR                        prob_strain 
#>                                502                                122 
#>                 cum_n_vaccinated_0                 cum_n_vaccinated_5 
#>                                218                                219 
#>                cum_n_vaccinated_10                cum_n_vaccinated_15 
#>                                220                                221 
#>                cum_n_vaccinated_20                cum_n_vaccinated_25 
#>                                222                                223 
#>                cum_n_vaccinated_30                cum_n_vaccinated_35 
#>                                224                                225 
#>                cum_n_vaccinated_40                cum_n_vaccinated_45 
#>                                226                                227 
#>                cum_n_vaccinated_50                cum_n_vaccinated_55 
#>                                228                                229 
#>                cum_n_vaccinated_60                cum_n_vaccinated_65 
#>                                230                                231 
#>                cum_n_vaccinated_70                cum_n_vaccinated_75 
#>                                232                                233 
#>                cum_n_vaccinated_80               cum_n_vaccinated_CHW 
#>                                234                                235 
#>               cum_n_vaccinated_CHR 
#>                                236 
#>