Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(info)

Arguments

info
The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
#> $run
#>                               time                                icu 
#>                                  1                                 12 
#>                            general               deaths_carehomes_inc 
#>                                 13                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 23                                 24 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 35                                 36 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 37                                 39 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 40                                 41 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 42                                 43 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 44                                 38 
#>                          react_pos 
#>                                 45 
#> 
#> $state
#>                                icu                            general 
#>                                 12                                 13 
#>                        deaths_comm                   deaths_carehomes 
#>                                 16                                 18 
#>                        deaths_hosp                           admitted 
#>                                 15                                  6 
#>                          diagnoses               deaths_carehomes_inc 
#>                                  7                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 23                                 24 
#>                        sympt_cases            sympt_cases_non_variant 
#>                                 25                                 26 
#>                 sympt_cases_over25     sympt_cases_non_variant_over25 
#>                                 27                                 28 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 35                                 36 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 37                                 39 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 40                                 41 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 42                                 43 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 44                                 38 
#>                          react_pos                               hosp 
#>                                 45                                 14 
#>                             deaths                         deaths_inc 
#>                                 21                                 22 
#>                         infections                     infections_inc 
#>                                  4                                  5 
#>                                S_0                                S_5 
#>                                295                                296 
#>                               S_10                               S_15 
#>                                297                                298 
#>                               S_20                               S_25 
#>                                299                                300 
#>                               S_30                               S_35 
#>                                301                                302 
#>                               S_40                               S_45 
#>                                303                                304 
#>                               S_50                               S_55 
#>                                305                                306 
#>                               S_60                               S_65 
#>                                307                                308 
#>                               S_70                               S_75 
#>                                309                                310 
#>                               S_80                              S_CHW 
#>                                311                                312 
#>                              S_CHR                                R_0 
#>                                313                                447 
#>                                R_5                               R_10 
#>                                448                                449 
#>                               R_15                               R_20 
#>                                450                                451 
#>                               R_25                               R_30 
#>                                452                                453 
#>                               R_35                               R_40 
#>                                454                                455 
#>                               R_45                               R_50 
#>                                456                                457 
#>                               R_55                               R_60 
#>                                458                                459 
#>                               R_65                               R_70 
#>                                460                                461 
#>                               R_75                               R_80 
#>                                462                                463 
#>                              R_CHW                              R_CHR 
#>                                464                                465 
#>                        cum_admit_0                        cum_admit_5 
#>                                 85                                 86 
#>                       cum_admit_10                       cum_admit_15 
#>                                 87                                 88 
#>                       cum_admit_20                       cum_admit_25 
#>                                 89                                 90 
#>                       cum_admit_30                       cum_admit_35 
#>                                 91                                 92 
#>                       cum_admit_40                       cum_admit_45 
#>                                 93                                 94 
#>                       cum_admit_50                       cum_admit_55 
#>                                 95                                 96 
#>                       cum_admit_60                       cum_admit_65 
#>                                 97                                 98 
#>                       cum_admit_70                       cum_admit_75 
#>                                 99                                100 
#>                       cum_admit_80                      cum_admit_CHW 
#>                                101                                102 
#>                      cum_admit_CHR                           D_hosp_0 
#>                                103                                 47 
#>                           D_hosp_5                          D_hosp_10 
#>                                 48                                 49 
#>                          D_hosp_15                          D_hosp_20 
#>                                 50                                 51 
#>                          D_hosp_25                          D_hosp_30 
#>                                 52                                 53 
#>                          D_hosp_35                          D_hosp_40 
#>                                 54                                 55 
#>                          D_hosp_45                          D_hosp_50 
#>                                 56                                 57 
#>                          D_hosp_55                          D_hosp_60 
#>                                 58                                 59 
#>                          D_hosp_65                          D_hosp_70 
#>                                 60                                 61 
#>                          D_hosp_75                          D_hosp_80 
#>                                 62                                 63 
#>                         D_hosp_CHW                         D_hosp_CHR 
#>                                 64                                 65 
#>                            D_all_0                            D_all_5 
#>                                276                                277 
#>                           D_all_10                           D_all_15 
#>                                278                                279 
#>                           D_all_20                           D_all_25 
#>                                280                                281 
#>                           D_all_30                           D_all_35 
#>                                282                                283 
#>                           D_all_40                           D_all_45 
#>                                284                                285 
#>                           D_all_50                           D_all_55 
#>                                286                                287 
#>                           D_all_60                           D_all_65 
#>                                288                                289 
#>                           D_all_70                           D_all_75 
#>                                290                                291 
#>                           D_all_80                          D_all_CHW 
#>                                292                                293 
#>                          D_all_CHR               diagnoses_admitted_0 
#>                                294                                238 
#>               diagnoses_admitted_5              diagnoses_admitted_10 
#>                                239                                240 
#>              diagnoses_admitted_15              diagnoses_admitted_20 
#>                                241                                242 
#>              diagnoses_admitted_25              diagnoses_admitted_30 
#>                                243                                244 
#>              diagnoses_admitted_35              diagnoses_admitted_40 
#>                                245                                246 
#>              diagnoses_admitted_45              diagnoses_admitted_50 
#>                                247                                248 
#>              diagnoses_admitted_55              diagnoses_admitted_60 
#>                                249                                250 
#>              diagnoses_admitted_65              diagnoses_admitted_70 
#>                                251                                252 
#>              diagnoses_admitted_75              diagnoses_admitted_80 
#>                                253                                254 
#>             diagnoses_admitted_CHW             diagnoses_admitted_CHR 
#>                                255                                256 
#>             cum_infections_disag_0             cum_infections_disag_5 
#>                                257                                258 
#>            cum_infections_disag_10            cum_infections_disag_15 
#>                                259                                260 
#>            cum_infections_disag_20            cum_infections_disag_25 
#>                                261                                262 
#>            cum_infections_disag_30            cum_infections_disag_35 
#>                                263                                264 
#>            cum_infections_disag_40            cum_infections_disag_45 
#>                                265                                266 
#>            cum_infections_disag_50            cum_infections_disag_55 
#>                                267                                268 
#>            cum_infections_disag_60            cum_infections_disag_65 
#>                                269                                270 
#>            cum_infections_disag_70            cum_infections_disag_75 
#>                                271                                272 
#>            cum_infections_disag_80           cum_infections_disag_CHW 
#>                                273                                274 
#>           cum_infections_disag_CHR                       I_weighted_0 
#>                                275                                485 
#>                       I_weighted_5                      I_weighted_10 
#>                                486                                487 
#>                      I_weighted_15                      I_weighted_20 
#>                                488                                489 
#>                      I_weighted_25                      I_weighted_30 
#>                                490                                491 
#>                      I_weighted_35                      I_weighted_40 
#>                                492                                493 
#>                      I_weighted_45                      I_weighted_50 
#>                                494                                495 
#>                      I_weighted_55                      I_weighted_60 
#>                                496                                497 
#>                      I_weighted_65                      I_weighted_70 
#>                                498                                499 
#>                      I_weighted_75                      I_weighted_80 
#>                                500                                501 
#>                     I_weighted_CHW                     I_weighted_CHR 
#>                                502                                503 
#>                        prob_strain                 cum_n_vaccinated_0 
#>                                123                                219 
#>                 cum_n_vaccinated_5                cum_n_vaccinated_10 
#>                                220                                221 
#>                cum_n_vaccinated_15                cum_n_vaccinated_20 
#>                                222                                223 
#>                cum_n_vaccinated_25                cum_n_vaccinated_30 
#>                                224                                225 
#>                cum_n_vaccinated_35                cum_n_vaccinated_40 
#>                                226                                227 
#>                cum_n_vaccinated_45                cum_n_vaccinated_50 
#>                                228                                229 
#>                cum_n_vaccinated_55                cum_n_vaccinated_60 
#>                                230                                231 
#>                cum_n_vaccinated_65                cum_n_vaccinated_70 
#>                                232                                233 
#>                cum_n_vaccinated_75                cum_n_vaccinated_80 
#>                                234                                235 
#>               cum_n_vaccinated_CHW               cum_n_vaccinated_CHR 
#>                                236                                237 
#>