Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(info)

Arguments

info

The result of running the $info() method on an initialised basic model

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
#> $run
#>                               time                                icu 
#>                                  1                                 12 
#>                            general               deaths_carehomes_inc 
#>                                 13                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 23                                 24 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 35                                 36 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 37                                 39 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 40                                 41 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 42                                 43 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 44                                 38 
#>                          react_pos 
#>                                 45 
#> 
#> $state
#>                                icu                            general 
#>                                 12                                 13 
#>                        deaths_comm                   deaths_carehomes 
#>                                 16                                 18 
#>                        deaths_hosp                           admitted 
#>                                 15                                  6 
#>                          diagnoses               deaths_carehomes_inc 
#>                                  7                                 19 
#>                    deaths_comm_inc                    deaths_hosp_inc 
#>                                 17                                 20 
#>                       admitted_inc                      diagnoses_inc 
#>                                  2                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 23                                 24 
#>                        sympt_cases            sympt_cases_non_variant 
#>                                 25                                 26 
#>                 sympt_cases_over25     sympt_cases_non_variant_over25 
#>                                 27                                 28 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 35                                 36 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 37                                 39 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 40                                 41 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 42                                 43 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 44                                 38 
#>                          react_pos                               hosp 
#>                                 45                                 14 
#>                             deaths                         deaths_inc 
#>                                 21                                 22 
#>                         infections                     infections_inc 
#>                                  4                                  5 
#>                                S_0                                S_5 
#>                                409                                410 
#>                               S_10                               S_15 
#>                                411                                412 
#>                               S_20                               S_25 
#>                                413                                414 
#>                               S_30                               S_35 
#>                                415                                416 
#>                               S_40                               S_45 
#>                                417                                418 
#>                               S_50                               S_55 
#>                                419                                420 
#>                               S_60                               S_65 
#>                                421                                422 
#>                               S_70                               S_75 
#>                                423                                424 
#>                               S_80                              S_CHW 
#>                                425                                426 
#>                              S_CHR                                R_0 
#>                                427                                561 
#>                                R_5                               R_10 
#>                                562                                563 
#>                               R_15                               R_20 
#>                                564                                565 
#>                               R_25                               R_30 
#>                                566                                567 
#>                               R_35                               R_40 
#>                                568                                569 
#>                               R_45                               R_50 
#>                                570                                571 
#>                               R_55                               R_60 
#>                                572                                573 
#>                               R_65                               R_70 
#>                                574                                575 
#>                               R_75                               R_80 
#>                                576                                577 
#>                              R_CHW                              R_CHR 
#>                                578                                579 
#>                        cum_admit_0                        cum_admit_5 
#>                                199                                200 
#>                       cum_admit_10                       cum_admit_15 
#>                                201                                202 
#>                       cum_admit_20                       cum_admit_25 
#>                                203                                204 
#>                       cum_admit_30                       cum_admit_35 
#>                                205                                206 
#>                       cum_admit_40                       cum_admit_45 
#>                                207                                208 
#>                       cum_admit_50                       cum_admit_55 
#>                                209                                210 
#>                       cum_admit_60                       cum_admit_65 
#>                                211                                212 
#>                       cum_admit_70                       cum_admit_75 
#>                                213                                214 
#>                       cum_admit_80                      cum_admit_CHW 
#>                                215                                216 
#>                      cum_admit_CHR                           D_hosp_0 
#>                                217                                161 
#>                           D_hosp_5                          D_hosp_10 
#>                                162                                163 
#>                          D_hosp_15                          D_hosp_20 
#>                                164                                165 
#>                          D_hosp_25                          D_hosp_30 
#>                                166                                167 
#>                          D_hosp_35                          D_hosp_40 
#>                                168                                169 
#>                          D_hosp_45                          D_hosp_50 
#>                                170                                171 
#>                          D_hosp_55                          D_hosp_60 
#>                                172                                173 
#>                          D_hosp_65                          D_hosp_70 
#>                                174                                175 
#>                          D_hosp_75                          D_hosp_80 
#>                                176                                177 
#>                         D_hosp_CHW                         D_hosp_CHR 
#>                                178                                179 
#>                            D_all_0                            D_all_5 
#>                                390                                391 
#>                           D_all_10                           D_all_15 
#>                                392                                393 
#>                           D_all_20                           D_all_25 
#>                                394                                395 
#>                           D_all_30                           D_all_35 
#>                                396                                397 
#>                           D_all_40                           D_all_45 
#>                                398                                399 
#>                           D_all_50                           D_all_55 
#>                                400                                401 
#>                           D_all_60                           D_all_65 
#>                                402                                403 
#>                           D_all_70                           D_all_75 
#>                                404                                405 
#>                           D_all_80                          D_all_CHW 
#>                                406                                407 
#>                          D_all_CHR               diagnoses_admitted_0 
#>                                408                                352 
#>               diagnoses_admitted_5              diagnoses_admitted_10 
#>                                353                                354 
#>              diagnoses_admitted_15              diagnoses_admitted_20 
#>                                355                                356 
#>              diagnoses_admitted_25              diagnoses_admitted_30 
#>                                357                                358 
#>              diagnoses_admitted_35              diagnoses_admitted_40 
#>                                359                                360 
#>              diagnoses_admitted_45              diagnoses_admitted_50 
#>                                361                                362 
#>              diagnoses_admitted_55              diagnoses_admitted_60 
#>                                363                                364 
#>              diagnoses_admitted_65              diagnoses_admitted_70 
#>                                365                                366 
#>              diagnoses_admitted_75              diagnoses_admitted_80 
#>                                367                                368 
#>             diagnoses_admitted_CHW             diagnoses_admitted_CHR 
#>                                369                                370 
#>             cum_infections_disag_0             cum_infections_disag_5 
#>                                371                                372 
#>            cum_infections_disag_10            cum_infections_disag_15 
#>                                373                                374 
#>            cum_infections_disag_20            cum_infections_disag_25 
#>                                375                                376 
#>            cum_infections_disag_30            cum_infections_disag_35 
#>                                377                                378 
#>            cum_infections_disag_40            cum_infections_disag_45 
#>                                379                                380 
#>            cum_infections_disag_50            cum_infections_disag_55 
#>                                381                                382 
#>            cum_infections_disag_60            cum_infections_disag_65 
#>                                383                                384 
#>            cum_infections_disag_70            cum_infections_disag_75 
#>                                385                                386 
#>            cum_infections_disag_80           cum_infections_disag_CHW 
#>                                387                                388 
#>           cum_infections_disag_CHR                       I_weighted_0 
#>                                389                                599 
#>                       I_weighted_5                      I_weighted_10 
#>                                600                                601 
#>                      I_weighted_15                      I_weighted_20 
#>                                602                                603 
#>                      I_weighted_25                      I_weighted_30 
#>                                604                                605 
#>                      I_weighted_35                      I_weighted_40 
#>                                606                                607 
#>                      I_weighted_45                      I_weighted_50 
#>                                608                                609 
#>                      I_weighted_55                      I_weighted_60 
#>                                610                                611 
#>                      I_weighted_65                      I_weighted_70 
#>                                612                                613 
#>                      I_weighted_75                      I_weighted_80 
#>                                614                                615 
#>                     I_weighted_CHW                     I_weighted_CHR 
#>                                616                                617 
#>                        prob_strain                 cum_n_vaccinated_0 
#>                                237                                333 
#>                 cum_n_vaccinated_5                cum_n_vaccinated_10 
#>                                334                                335 
#>                cum_n_vaccinated_15                cum_n_vaccinated_20 
#>                                336                                337 
#>                cum_n_vaccinated_25                cum_n_vaccinated_30 
#>                                338                                339 
#>                cum_n_vaccinated_35                cum_n_vaccinated_40 
#>                                340                                341 
#>                cum_n_vaccinated_45                cum_n_vaccinated_50 
#>                                342                                343 
#>                cum_n_vaccinated_55                cum_n_vaccinated_60 
#>                                344                                345 
#>                cum_n_vaccinated_65                cum_n_vaccinated_70 
#>                                346                                347 
#>                cum_n_vaccinated_75                cum_n_vaccinated_80 
#>                                348                                349 
#>               cum_n_vaccinated_CHW               cum_n_vaccinated_CHR 
#>                                350                                351 
#>