Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(
  info,
  rt = TRUE,
  cum_admit = TRUE,
  diagnoses_admitted = TRUE,
  cum_infections_disag = TRUE,
  cum_n_vaccinated = TRUE,
  D_all = TRUE,
  D_hosp = TRUE,
  infections_inc_per_strain = TRUE,
  severity = FALSE
)

Arguments

info

The result of running the $info() method on an initialised basic model

rt

Logical, whether to output trajectories required for calculating Rt (default = TRUE)

cum_admit

Logical, whether to output cumulative admissions by age (default = TRUE)

diagnoses_admitted

Logical, whether to output cumulative combined confirmed admissions and inpatient diagnoses by age and vaccine class (default = TRUE)

cum_infections_disag

Logical, whether to output cumulative infections by age and vaccine class (default = TRUE)

cum_n_vaccinated

Logical, whether to output cumulative number vaccinated by age and vaccine class (default = TRUE)

D_all

Logical, whether to output all deaths by age and vaccine class (default = TRUE)

D_hosp

Logical, whether to output hospital deaths by age (default = TRUE)

infections_inc_per_strain

Logical, whether to output infections incidence per strain (default = TRUE)

severity

Logical, whether to output severity calculations (default = FALSE)

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
#> $run
#>                               time                                icu 
#>                                  1                                 21 
#>                            general               deaths_carehomes_inc 
#>                                 22                                 44 
#>                    deaths_comm_inc               deaths_comm_0_49_inc 
#>                                 34                                 35 
#>              deaths_comm_50_54_inc              deaths_comm_55_59_inc 
#>                                 36                                 37 
#>              deaths_comm_60_64_inc              deaths_comm_65_69_inc 
#>                                 38                                 39 
#>              deaths_comm_70_74_inc              deaths_comm_75_79_inc 
#>                                 40                                 41 
#>            deaths_comm_80_plus_inc                    deaths_hosp_inc 
#>                                 42                                 47 
#>               deaths_hosp_0_49_inc              deaths_hosp_50_54_inc 
#>                                 48                                 49 
#>              deaths_hosp_55_59_inc              deaths_hosp_60_64_inc 
#>                                 50                                 51 
#>              deaths_hosp_65_69_inc              deaths_hosp_70_74_inc 
#>                                 52                                 53 
#>              deaths_hosp_75_79_inc            deaths_hosp_80_plus_inc 
#>                                 54                                 55 
#>                       admitted_inc              all_admission_0_9_inc 
#>                                  2                                  4 
#>            all_admission_10_19_inc            all_admission_20_29_inc 
#>                                  5                                  6 
#>            all_admission_30_39_inc            all_admission_40_49_inc 
#>                                  7                                  8 
#>            all_admission_50_59_inc            all_admission_60_69_inc 
#>                                  9                                 10 
#>            all_admission_70_79_inc          all_admission_80_plus_inc 
#>                                 11                                 12 
#>                      diagnoses_inc                         sero_pos_1 
#>                                  3                                 56 
#>                         sero_pos_2                    sympt_cases_inc 
#>                                 57                                 68 
#>        sympt_cases_non_variant_inc             sympt_cases_over25_inc 
#>                                 69                                 70 
#>            sympt_cases_under15_inc              sympt_cases_15_24_inc 
#>                                 72                                 73 
#>              sympt_cases_25_49_inc              sympt_cases_50_64_inc 
#>                                 74                                 75 
#>              sympt_cases_65_79_inc            sympt_cases_80_plus_inc 
#>                                 76                                 77 
#> sympt_cases_non_variant_over25_inc                          react_pos 
#>                                 71                                 78 
#>                     react_5_24_pos                    react_25_34_pos 
#>                                 79                                 80 
#>                    react_35_44_pos                    react_45_54_pos 
#>                                 81                                 82 
#>                    react_55_64_pos                  react_65_plus_pos 
#>                                 83                                 84 
#> 
#> $state
#>                                icu                            general 
#>                                 21                                 22 
#>               deaths_carehomes_inc                    deaths_comm_inc 
#>                                 44                                 34 
#>               deaths_comm_0_49_inc              deaths_comm_50_54_inc 
#>                                 35                                 36 
#>              deaths_comm_55_59_inc              deaths_comm_60_64_inc 
#>                                 37                                 38 
#>              deaths_comm_65_69_inc              deaths_comm_70_74_inc 
#>                                 39                                 40 
#>              deaths_comm_75_79_inc            deaths_comm_80_plus_inc 
#>                                 41                                 42 
#>                    deaths_hosp_inc               deaths_hosp_0_49_inc 
#>                                 47                                 48 
#>              deaths_hosp_50_54_inc              deaths_hosp_55_59_inc 
#>                                 49                                 50 
#>              deaths_hosp_60_64_inc              deaths_hosp_65_69_inc 
#>                                 51                                 52 
#>              deaths_hosp_70_74_inc              deaths_hosp_75_79_inc 
#>                                 53                                 54 
#>            deaths_hosp_80_plus_inc                       admitted_inc 
#>                                 55                                  2 
#>              all_admission_0_9_inc            all_admission_10_19_inc 
#>                                  4                                  5 
#>            all_admission_20_29_inc            all_admission_30_39_inc 
#>                                  6                                  7 
#>            all_admission_40_49_inc            all_admission_50_59_inc 
#>                                  8                                  9 
#>            all_admission_60_69_inc            all_admission_70_79_inc 
#>                                 10                                 11 
#>          all_admission_80_plus_inc                      diagnoses_inc 
#>                                 12                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 56                                 57 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 68                                 69 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 70                                 72 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 73                                 74 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 75                                 76 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 77                                 71 
#>                          react_pos                     react_5_24_pos 
#>                                 78                                 79 
#>                    react_25_34_pos                    react_35_44_pos 
#>                                 80                                 81 
#>                    react_45_54_pos                    react_55_64_pos 
#>                                 82                                 83 
#>                  react_65_plus_pos                               hosp 
#>                                 84                                 23 
#>                           admitted                          diagnoses 
#>                                 15                                 16 
#>                             deaths                         deaths_inc 
#>                                 45                                 46 
#>                         infections                     infections_inc 
#>                                 13                                 14 
#>            effective_susceptible_1                                S_0 
#>                                167                                453 
#>                                S_5                               S_10 
#>                                454                                455 
#>                               S_15                               S_20 
#>                                456                                457 
#>                               S_25                               S_30 
#>                                458                                459 
#>                               S_35                               S_40 
#>                                460                                461 
#>                               S_45                               S_50 
#>                                462                                463 
#>                               S_55                               S_60 
#>                                464                                465 
#>                               S_65                               S_70 
#>                                466                                467 
#>                               S_75                               S_80 
#>                                468                                469 
#>                              S_CHW                              S_CHR 
#>                                470                                471 
#>                                R_0                                R_5 
#>                                605                                606 
#>                               R_10                               R_15 
#>                                607                                608 
#>                               R_20                               R_25 
#>                                609                                610 
#>                               R_30                               R_35 
#>                                611                                612 
#>                               R_40                               R_45 
#>                                613                                614 
#>                               R_50                               R_55 
#>                                615                                616 
#>                               R_60                               R_65 
#>                                617                                618 
#>                               R_70                               R_75 
#>                                619                                620 
#>                               R_80                              R_CHW 
#>                                621                                622 
#>                              R_CHR                        prob_strain 
#>                                623                                166 
#>                        cum_admit_0                        cum_admit_5 
#>                                128                                129 
#>                       cum_admit_10                       cum_admit_15 
#>                                130                                131 
#>                       cum_admit_20                       cum_admit_25 
#>                                132                                133 
#>                       cum_admit_30                       cum_admit_35 
#>                                134                                135 
#>                       cum_admit_40                       cum_admit_45 
#>                                136                                137 
#>                       cum_admit_50                       cum_admit_55 
#>                                138                                139 
#>                       cum_admit_60                       cum_admit_65 
#>                                140                                141 
#>                       cum_admit_70                       cum_admit_75 
#>                                142                                143 
#>                       cum_admit_80                      cum_admit_CHW 
#>                                144                                145 
#>                      cum_admit_CHR               diagnoses_admitted_0 
#>                                146                                396 
#>               diagnoses_admitted_5              diagnoses_admitted_10 
#>                                397                                398 
#>              diagnoses_admitted_15              diagnoses_admitted_20 
#>                                399                                400 
#>              diagnoses_admitted_25              diagnoses_admitted_30 
#>                                401                                402 
#>              diagnoses_admitted_35              diagnoses_admitted_40 
#>                                403                                404 
#>              diagnoses_admitted_45              diagnoses_admitted_50 
#>                                405                                406 
#>              diagnoses_admitted_55              diagnoses_admitted_60 
#>                                407                                408 
#>              diagnoses_admitted_65              diagnoses_admitted_70 
#>                                409                                410 
#>              diagnoses_admitted_75              diagnoses_admitted_80 
#>                                411                                412 
#>             diagnoses_admitted_CHW             diagnoses_admitted_CHR 
#>                                413                                414 
#>             cum_infections_disag_0             cum_infections_disag_5 
#>                                415                                416 
#>            cum_infections_disag_10            cum_infections_disag_15 
#>                                417                                418 
#>            cum_infections_disag_20            cum_infections_disag_25 
#>                                419                                420 
#>            cum_infections_disag_30            cum_infections_disag_35 
#>                                421                                422 
#>            cum_infections_disag_40            cum_infections_disag_45 
#>                                423                                424 
#>            cum_infections_disag_50            cum_infections_disag_55 
#>                                425                                426 
#>            cum_infections_disag_60            cum_infections_disag_65 
#>                                427                                428 
#>            cum_infections_disag_70            cum_infections_disag_75 
#>                                429                                430 
#>            cum_infections_disag_80           cum_infections_disag_CHW 
#>                                431                                432 
#>           cum_infections_disag_CHR                 cum_n_vaccinated_0 
#>                                433                                263 
#>                 cum_n_vaccinated_5                cum_n_vaccinated_10 
#>                                264                                265 
#>                cum_n_vaccinated_15                cum_n_vaccinated_20 
#>                                266                                267 
#>                cum_n_vaccinated_25                cum_n_vaccinated_30 
#>                                268                                269 
#>                cum_n_vaccinated_35                cum_n_vaccinated_40 
#>                                270                                271 
#>                cum_n_vaccinated_45                cum_n_vaccinated_50 
#>                                272                                273 
#>                cum_n_vaccinated_55                cum_n_vaccinated_60 
#>                                274                                275 
#>                cum_n_vaccinated_65                cum_n_vaccinated_70 
#>                                276                                277 
#>                cum_n_vaccinated_75                cum_n_vaccinated_80 
#>                                278                                279 
#>               cum_n_vaccinated_CHW               cum_n_vaccinated_CHR 
#>                                280                                281 
#>                            D_all_0                            D_all_5 
#>                                434                                435 
#>                           D_all_10                           D_all_15 
#>                                436                                437 
#>                           D_all_20                           D_all_25 
#>                                438                                439 
#>                           D_all_30                           D_all_35 
#>                                440                                441 
#>                           D_all_40                           D_all_45 
#>                                442                                443 
#>                           D_all_50                           D_all_55 
#>                                444                                445 
#>                           D_all_60                           D_all_65 
#>                                446                                447 
#>                           D_all_70                           D_all_75 
#>                                448                                449 
#>                           D_all_80                          D_all_CHW 
#>                                450                                451 
#>                          D_all_CHR                           D_hosp_0 
#>                                452                                 90 
#>                           D_hosp_5                          D_hosp_10 
#>                                 91                                 92 
#>                          D_hosp_15                          D_hosp_20 
#>                                 93                                 94 
#>                          D_hosp_25                          D_hosp_30 
#>                                 95                                 96 
#>                          D_hosp_35                          D_hosp_40 
#>                                 97                                 98 
#>                          D_hosp_45                          D_hosp_50 
#>                                 99                                100 
#>                          D_hosp_55                          D_hosp_60 
#>                                101                                102 
#>                          D_hosp_65                          D_hosp_70 
#>                                103                                104 
#>                          D_hosp_75                          D_hosp_80 
#>                                105                                106 
#>                         D_hosp_CHW                         D_hosp_CHR 
#>                                107                                108 
#>            infections_inc_strain_1 
#>                                 89 
#>