Index of lancelot model

Index of "interesting" elements for the lancelot model. This function conforms to the mcstate interface.

lancelot_index(
  info,
  rt = TRUE,
  cum_admit = TRUE,
  diagnoses_admitted = TRUE,
  cum_infections_disag = TRUE,
  cum_n_vaccinated = TRUE,
  D_all = TRUE,
  D_hosp = TRUE,
  infections_inc_per_strain = TRUE,
  severity = FALSE,
  protected = FALSE
)

Arguments

info

The result of running the $info() method on an initialised basic model

rt

Logical, whether to output trajectories required for calculating Rt (default = TRUE)

cum_admit

Logical, whether to output cumulative admissions by age (default = TRUE)

diagnoses_admitted

Logical, whether to output cumulative combined confirmed admissions and inpatient diagnoses by age and vaccine class (default = TRUE)

cum_infections_disag

Logical, whether to output cumulative infections by age and vaccine class (default = TRUE)

cum_n_vaccinated

Logical, whether to output cumulative number vaccinated by age and vaccine class (default = TRUE)

D_all

Logical, whether to output all deaths by age and vaccine class (default = TRUE)

D_hosp

Logical, whether to output hospital deaths by age (default = TRUE)

infections_inc_per_strain

Logical, whether to output infections incidence per strain (default = TRUE)

severity

Logical, whether to output severity calculations (default = FALSE)

protected

Logical, whether to output protected levels

Value

A list with element run, indicating the locations of (in order) (1) ICU, (2) general, (3) deaths in community, (4) deaths in hospital, (5) total deaths, (6) cumulative confirmed admissions,(7) cumulative confirmed new admissions, (8) probability of a positive sero test, (9) probability of a positive pillar 2 test, and with element state containing the same values followed by 17 S compartments (one per age group, then one for carehome workers and carehome residents respectively) and 17 "cumulative admission" compartments.

Examples

p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
#> $run
#>                               time                                icu 
#>                                  1                                 22 
#>                            general               deaths_carehomes_inc 
#>                                 23                                 45 
#>                    deaths_comm_inc               deaths_comm_0_49_inc 
#>                                 35                                 36 
#>              deaths_comm_50_54_inc              deaths_comm_55_59_inc 
#>                                 37                                 38 
#>              deaths_comm_60_64_inc              deaths_comm_65_69_inc 
#>                                 39                                 40 
#>              deaths_comm_70_74_inc              deaths_comm_75_79_inc 
#>                                 41                                 42 
#>            deaths_comm_80_plus_inc                    deaths_hosp_inc 
#>                                 43                                 48 
#>               deaths_hosp_0_49_inc              deaths_hosp_50_54_inc 
#>                                 49                                 50 
#>              deaths_hosp_55_59_inc              deaths_hosp_60_64_inc 
#>                                 51                                 52 
#>              deaths_hosp_65_69_inc              deaths_hosp_70_74_inc 
#>                                 53                                 54 
#>              deaths_hosp_75_79_inc            deaths_hosp_80_plus_inc 
#>                                 55                                 56 
#>                       admitted_inc              all_admission_0_9_inc 
#>                                  2                                  4 
#>            all_admission_10_19_inc            all_admission_20_29_inc 
#>                                  5                                  6 
#>            all_admission_30_39_inc            all_admission_40_49_inc 
#>                                  7                                  8 
#>            all_admission_50_59_inc            all_admission_60_69_inc 
#>                                  9                                 10 
#>            all_admission_70_79_inc          all_admission_80_plus_inc 
#>                                 11                                 12 
#>                      diagnoses_inc                         sero_pos_1 
#>                                  3                                 57 
#>                         sero_pos_2                    sympt_cases_inc 
#>                                 58                                 69 
#>        sympt_cases_non_variant_inc             sympt_cases_over25_inc 
#>                                 70                                 71 
#>            sympt_cases_under15_inc              sympt_cases_15_24_inc 
#>                                 73                                 74 
#>              sympt_cases_25_49_inc              sympt_cases_50_64_inc 
#>                                 75                                 76 
#>              sympt_cases_65_79_inc            sympt_cases_80_plus_inc 
#>                                 77                                 78 
#> sympt_cases_non_variant_over25_inc                            ons_pos 
#>                                 72                                 79 
#>                          react_pos                     react_5_24_pos 
#>                                 80                                 81 
#>                    react_25_34_pos                    react_35_44_pos 
#>                                 82                                 83 
#>                    react_45_54_pos                    react_55_64_pos 
#>                                 84                                 85 
#>                  react_65_plus_pos 
#>                                 86 
#> 
#> $state
#>                                icu                            general 
#>                                 22                                 23 
#>               deaths_carehomes_inc                    deaths_comm_inc 
#>                                 45                                 35 
#>               deaths_comm_0_49_inc              deaths_comm_50_54_inc 
#>                                 36                                 37 
#>              deaths_comm_55_59_inc              deaths_comm_60_64_inc 
#>                                 38                                 39 
#>              deaths_comm_65_69_inc              deaths_comm_70_74_inc 
#>                                 40                                 41 
#>              deaths_comm_75_79_inc            deaths_comm_80_plus_inc 
#>                                 42                                 43 
#>                    deaths_hosp_inc               deaths_hosp_0_49_inc 
#>                                 48                                 49 
#>              deaths_hosp_50_54_inc              deaths_hosp_55_59_inc 
#>                                 50                                 51 
#>              deaths_hosp_60_64_inc              deaths_hosp_65_69_inc 
#>                                 52                                 53 
#>              deaths_hosp_70_74_inc              deaths_hosp_75_79_inc 
#>                                 54                                 55 
#>            deaths_hosp_80_plus_inc                       admitted_inc 
#>                                 56                                  2 
#>              all_admission_0_9_inc            all_admission_10_19_inc 
#>                                  4                                  5 
#>            all_admission_20_29_inc            all_admission_30_39_inc 
#>                                  6                                  7 
#>            all_admission_40_49_inc            all_admission_50_59_inc 
#>                                  8                                  9 
#>            all_admission_60_69_inc            all_admission_70_79_inc 
#>                                 10                                 11 
#>          all_admission_80_plus_inc                      diagnoses_inc 
#>                                 12                                  3 
#>                         sero_pos_1                         sero_pos_2 
#>                                 57                                 58 
#>                    sympt_cases_inc        sympt_cases_non_variant_inc 
#>                                 69                                 70 
#>             sympt_cases_over25_inc            sympt_cases_under15_inc 
#>                                 71                                 73 
#>              sympt_cases_15_24_inc              sympt_cases_25_49_inc 
#>                                 74                                 75 
#>              sympt_cases_50_64_inc              sympt_cases_65_79_inc 
#>                                 76                                 77 
#>            sympt_cases_80_plus_inc sympt_cases_non_variant_over25_inc 
#>                                 78                                 72 
#>                            ons_pos                          react_pos 
#>                                 79                                 80 
#>                     react_5_24_pos                    react_25_34_pos 
#>                                 81                                 82 
#>                    react_35_44_pos                    react_45_54_pos 
#>                                 83                                 84 
#>                    react_55_64_pos                  react_65_plus_pos 
#>                                 85                                 86 
#>                               hosp                           admitted 
#>                                 24                                 16 
#>                          diagnoses                             deaths 
#>                                 17                                 46 
#>                         deaths_inc                         infections 
#>                                 47                                 13 
#>                     infections_inc                        susceptible 
#>                                 14                                 87 
#>                         recovered1            effective_susceptible_1 
#>                                191                                190 
#>                                S_0                                S_5 
#>                                540                                541 
#>                               S_10                               S_15 
#>                                542                                543 
#>                               S_20                               S_25 
#>                                544                                545 
#>                               S_30                               S_35 
#>                                546                                547 
#>                               S_40                               S_45 
#>                                548                                549 
#>                               S_50                               S_55 
#>                                550                                551 
#>                               S_60                               S_65 
#>                                552                                553 
#>                               S_70                               S_75 
#>                                554                                555 
#>                               S_80                              S_CHW 
#>                                556                                557 
#>                              S_CHR                                R_0 
#>                                558                                711 
#>                                R_5                               R_10 
#>                                712                                713 
#>                               R_15                               R_20 
#>                                714                                715 
#>                               R_25                               R_30 
#>                                716                                717 
#>                               R_35                               R_40 
#>                                718                                719 
#>                               R_45                               R_50 
#>                                720                                721 
#>                               R_55                               R_60 
#>                                722                                723 
#>                               R_65                               R_70 
#>                                724                                725 
#>                               R_75                               R_80 
#>                                726                                727 
#>                              R_CHW                              R_CHR 
#>                                728                                729 
#>                        prob_strain                        cum_admit_0 
#>                                189                                151 
#>                        cum_admit_5                       cum_admit_10 
#>                                152                                153 
#>                       cum_admit_15                       cum_admit_20 
#>                                154                                155 
#>                       cum_admit_25                       cum_admit_30 
#>                                156                                157 
#>                       cum_admit_35                       cum_admit_40 
#>                                158                                159 
#>                       cum_admit_45                       cum_admit_50 
#>                                160                                161 
#>                       cum_admit_55                       cum_admit_60 
#>                                162                                163 
#>                       cum_admit_65                       cum_admit_70 
#>                                164                                165 
#>                       cum_admit_75                       cum_admit_80 
#>                                166                                167 
#>                      cum_admit_CHW                      cum_admit_CHR 
#>                                168                                169 
#>               diagnoses_admitted_0               diagnoses_admitted_5 
#>                                483                                484 
#>              diagnoses_admitted_10              diagnoses_admitted_15 
#>                                485                                486 
#>              diagnoses_admitted_20              diagnoses_admitted_25 
#>                                487                                488 
#>              diagnoses_admitted_30              diagnoses_admitted_35 
#>                                489                                490 
#>              diagnoses_admitted_40              diagnoses_admitted_45 
#>                                491                                492 
#>              diagnoses_admitted_50              diagnoses_admitted_55 
#>                                493                                494 
#>              diagnoses_admitted_60              diagnoses_admitted_65 
#>                                495                                496 
#>              diagnoses_admitted_70              diagnoses_admitted_75 
#>                                497                                498 
#>              diagnoses_admitted_80             diagnoses_admitted_CHW 
#>                                499                                500 
#>             diagnoses_admitted_CHR             cum_infections_disag_0 
#>                                501                                502 
#>             cum_infections_disag_5            cum_infections_disag_10 
#>                                503                                504 
#>            cum_infections_disag_15            cum_infections_disag_20 
#>                                505                                506 
#>            cum_infections_disag_25            cum_infections_disag_30 
#>                                507                                508 
#>            cum_infections_disag_35            cum_infections_disag_40 
#>                                509                                510 
#>            cum_infections_disag_45            cum_infections_disag_50 
#>                                511                                512 
#>            cum_infections_disag_55            cum_infections_disag_60 
#>                                513                                514 
#>            cum_infections_disag_65            cum_infections_disag_70 
#>                                515                                516 
#>            cum_infections_disag_75            cum_infections_disag_80 
#>                                517                                518 
#>           cum_infections_disag_CHW           cum_infections_disag_CHR 
#>                                519                                520 
#>                 cum_n_vaccinated_0                 cum_n_vaccinated_5 
#>                                350                                351 
#>                cum_n_vaccinated_10                cum_n_vaccinated_15 
#>                                352                                353 
#>                cum_n_vaccinated_20                cum_n_vaccinated_25 
#>                                354                                355 
#>                cum_n_vaccinated_30                cum_n_vaccinated_35 
#>                                356                                357 
#>                cum_n_vaccinated_40                cum_n_vaccinated_45 
#>                                358                                359 
#>                cum_n_vaccinated_50                cum_n_vaccinated_55 
#>                                360                                361 
#>                cum_n_vaccinated_60                cum_n_vaccinated_65 
#>                                362                                363 
#>                cum_n_vaccinated_70                cum_n_vaccinated_75 
#>                                364                                365 
#>                cum_n_vaccinated_80               cum_n_vaccinated_CHW 
#>                                366                                367 
#>               cum_n_vaccinated_CHR                            D_all_0 
#>                                368                                521 
#>                            D_all_5                           D_all_10 
#>                                522                                523 
#>                           D_all_15                           D_all_20 
#>                                524                                525 
#>                           D_all_25                           D_all_30 
#>                                526                                527 
#>                           D_all_35                           D_all_40 
#>                                528                                529 
#>                           D_all_45                           D_all_50 
#>                                530                                531 
#>                           D_all_55                           D_all_60 
#>                                532                                533 
#>                           D_all_65                           D_all_70 
#>                                534                                535 
#>                           D_all_75                           D_all_80 
#>                                536                                537 
#>                          D_all_CHW                          D_all_CHR 
#>                                538                                539 
#>                           D_hosp_0                           D_hosp_5 
#>                                559                                560 
#>                          D_hosp_10                          D_hosp_15 
#>                                561                                562 
#>                          D_hosp_20                          D_hosp_25 
#>                                563                                564 
#>                          D_hosp_30                          D_hosp_35 
#>                                565                                566 
#>                          D_hosp_40                          D_hosp_45 
#>                                567                                568 
#>                          D_hosp_50                          D_hosp_55 
#>                                569                                570 
#>                          D_hosp_60                          D_hosp_65 
#>                                571                                572 
#>                          D_hosp_70                          D_hosp_75 
#>                                573                                574 
#>                          D_hosp_80                         D_hosp_CHW 
#>                                575                                576 
#>                         D_hosp_CHR            infections_inc_strain_1 
#>                                577                                 92 
#>